Function zng::access::on_access_number

pub fn on_access_number(
    child: impl UiNode,
    handler: impl WidgetHandler<AccessNumberArgs>
) -> impl UiNode

P Access requested a number input.

Preview

You can preview this event using on_pre_access_number. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.