zng::app

Function on_app_start

pub fn on_app_start(handler: impl FnMut(&AppStartArgs) + Send + 'static)

Register a handler to run when an APP starts running in the process.

The handler is called in the new app context, just before the “run” future executes, all app service are already available in it.

In "multi_app" builds the handler can be called more them once.