Function zng::app::on_pre_save

pub fn on_pre_save(
    child: impl UiNode,
    handler: impl WidgetHandler<CommandArgs>
) -> impl UiNode

P Preview on_save command.

Preview

Preview event properties call the handler before the main event property and before the widget content, if you stop the propagation of a preview event the main event handler is not called.

Async

You can use async event handlers with this property, note that only the code before the fist .await is preview, subsequent code runs in widget updates.