zng::app

Function print_tracing

pub fn print_tracing(max: Level) -> bool

Enables tracing events printing if a subscriber is not already set.

All non-fatal errors in the Zng project are logged using tracing.

In debug builds this function is called automatically with level INFO on app start.

In "wasm32" builds logs to the browser console.

In "android" builds logs to logcat.

See also test_log to enable panicking on error log.

See also print_tracing_filter for the filter used by this.