zng::app

Module raw_device_events

Input device hardware ID and events.

Full API

See zng_app::view_process::raw_device_events for the full API.

Structs

AxisId

Identifier for a specific analog axis on some device.

AxisMotionArgs

Arguments for AXIS_MOTION_EVENT.

INPUT_DEVICES

Input devices info service.

InputDeviceCapability

Capabilities of an input device.

InputDeviceId

Unique identifier of an input device event source.

InputDeviceInfo

Info about an human input device.

InputDevicesChangedArgs

Arguments for INPUT_DEVICES_CHANGED_EVENT.

Statics

AXIS_MOTION_EVENT

Motion on some analog axis.

INPUT_DEVICES_CHANGED_EVENT

Raw input devices have been added or removed from the system.