zng::app

Struct AppExtended

pub struct AppExtended<E>
where
    E: AppExtension,
{ /* private fields */ }

Application builder.

You can use APP to start building the app.

Implementations

impl<E> AppExtended<E>
where E: AppExtension,

pub fn view_process_exe( self, view_process_exe: impl Into<PathBuf>, ) -> AppExtended<E>

Set the path to the executable for the View Process.

By the default the current executable is started again as a View Process, you can use two executables instead, by setting this value.

Note that the view_process_exe must start a view server and both executables must be build using the same exact VERSION.

pub fn view_process_env( self, name: impl Into<Txt>, value: impl Into<Txt>, ) -> AppExtended<E>

Set an env variable for the view-process.

pub fn run<F>(self, start: impl IntoFuture<IntoFuture = F>)
where F: Future<Output = ()> + Send + 'static,

Starts the app, then starts polling start to run.

This method only returns when the app has exited.

The start task runs in a UiTask in the app context, note that it only needs to start the app, usually by opening a window, the app will keep running after start is finished.

pub fn run_headless(self, with_renderer: bool) -> HeadlessApp

Initializes extensions in headless mode and returns an HeadlessApp.

If with_renderer is true spawns a renderer process for headless rendering. See HeadlessApp::renderer_enabled for more details.