zng::app

Struct Deadline

pub struct Deadline(pub DInstant);

Represents a timeout instant.

Deadlines and timeouts can be specified as a DInstant in the future or as a Duration from now, both of these types can be converted to this struct.

Examples

In the example below the timer function accepts Deadline, DInstant and Duration inputs.

fn timer(deadline: impl Into<Deadline>) {
    let deadline = deadline.into();
    // ..
}

timer(5.secs());

Tuple Fields
0: DInstant
Implementations

impl Deadline

pub const ELAPSED: Deadline

Deadline that is always elapsed.

pub const MAX: Deadline

Deadline that is practically never reached.

pub fn timeout(dur: Duration) -> Deadline

New deadline from now + dur.

pub fn has_elapsed(self) -> bool

Returns true if the deadline was reached.

pub fn time_left(self) -> Option<Duration>

Returns the time left until the deadline is reached.

pub fn min(self, other: Deadline) -> Deadline

Returns the deadline further into the past or closest to now.

pub fn max(self, other: Deadline) -> Deadline

Returns the deadline further into the future.