zng::clipboard

Function on_paste

pub fn on_paste(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<CommandArgs>,
) -> UiNode

P On paste command.

Receives PASTE_CMD command events scoped on the widget. You can use the CLIPBOARD service to receive data from the clipboard.

Preview

You can preview this command event using on_pre_paste. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.

Enabled

You can use the can_paste context property to disable the command handle. When disabled the handler is not called and the command handle signals as disabled.