zng::color::filter

Function child_filter

pub fn child_filter(
    child: impl IntoUiNode,
    filter: impl IntoVar<Filter>,
) -> UiNode

P Color filter, or combination of filters targeting the widget’s descendants and not the widget itself.

This property allows setting multiple filters at once, there is also a property for every filter for easier value updating.

Performance

The performance for setting specific filter properties versus this one is the same, except for child_opacity which can be animated using only frame updates instead of generating a new frame every change.