Struct zng::color::filter::Filter

pub struct Filter { /* private fields */ }

A color filter or combination of filters.

Examples

use zng_color::filter::Filter;
use zng_layout::unit::*;

let filter = Filter::new_opacity(50.pct()).blur(3);

The example above creates a filter that lowers the opacity to 50% and blurs by 3px.

Implementations

impl Filter

pub fn opacity<A>(self, alpha: A) -> Filter
where A: Into<Factor>,

Add an opacity adjustment to the filter, zero is fully transparent, one is the input transparency.

pub fn invert<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a color inversion filter, zero does not invert, one fully inverts.

pub fn blur<R>(self, radius: R) -> Filter
where R: Into<Length>,

Add a blue effect to the filter, the blue radius is defined by a Length.

Relative lengths are calculated by the width of the available space.

pub fn sepia<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a sepia color effect to the filter, zero is the input color, one is the full desaturated brown look.

pub fn grayscale<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a grayscale color effect to the filter, zero is the input color, one if the full grayscale.

pub fn drop_shadow<O, R, C>(self, offset: O, blur_radius: R, color: C) -> Filter
where O: Into<Point>, R: Into<Length>, C: Into<Rgba>,

Add a drop-shadow to the effect.

pub fn brightness<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a brightness adjustment to the filter, zero removes all brightness, one is the input brightness.

pub fn contrast<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a contrast adjustment to the filter, zero removes all contrast, one is the input contrast.

pub fn saturate<A>(self, amount: A) -> Filter
where A: Into<Factor>,

Add a saturation adjustment to the filter, zero fully desaturates, one is the input saturation.

pub fn hue_rotate<A>(self, angle: A) -> Filter
where A: Into<AngleDegree>,

Add a filter that adds the angle to each color hue value.

pub fn flood<C>(self, color: C) -> Filter
where C: Into<Rgba>,

Add a filter that fills the pixel space with color.

pub fn color_matrix<M>(self, matrix: M) -> Filter
where M: Into<ColorMatrix>,

Custom filter.

The color matrix is in the format of SVG color matrix, [0..5] is the first matrix row.

impl Filter

pub const fn new() -> Filter

New default empty.

pub fn new_opacity<A>(alpha: A) -> Filter
where A: Into<Factor>,

New Filter::opacity.

pub fn new_invert<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::invert.

pub fn new_blur<R>(radius: R) -> Filter
where R: Into<Length>,

New Filter::blur.

pub fn new_sepia<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::sepia.

pub fn new_grayscale<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::grayscale.

pub fn new_drop_shadow<O, R, C>(offset: O, blur_radius: R, color: C) -> Filter
where O: Into<Point>, R: Into<Length>, C: Into<Rgba>,

New Filter::drop_shadow.

pub fn new_brightness<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::brightness.

pub fn new_contrast<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::contrast.

pub fn new_saturate<A>(amount: A) -> Filter
where A: Into<Factor>,

New Filter::saturate.

pub fn new_hue_rotate<A>(angle: A) -> Filter
where A: Into<AngleDegree>,

New Filter::hue_rotate.

pub fn new_flood<C>(color: C) -> Filter
where C: Into<Rgba>,

New Filter::flood.

pub fn new_color_matrix<M>(matrix: M) -> Filter
where M: Into<ColorMatrix>,

New Filter::color_matrix.

impl Filter

pub fn layout(&self) -> Vec<FilterOp>

Compute a RenderFilter in the current LAYOUT context.

Most filters convert one-to-one, effects that have a Length value use the layout context to calculate relative values.

Relative blur radius lengths are calculated using the constraints().fill_size().width value.

pub fn try_render(&self) -> Option<Vec<FilterOp>>

Compute a RenderFilter if the filter is not affected by layout.

pub fn needs_layout(&self) -> bool

Returns true if this filter is affected by the layout context where it is evaluated.