Function zng::color::gradient::linear_gradient

pub fn linear_gradient<A, S>(
    axis: A,
    stops: S,
) -> LinearGradient<<S as IntoVar<GradientStops>>::Var, <A as IntoVar<LinearGradientAxis>>::Var, LocalVar<ExtendMode>>
where
    A: IntoVar<LinearGradientAxis>,
    S: IntoVar<GradientStops>,

Starts building a linear gradient with the axis and color stops.

Returns a node that is also a builder that can be used to refine the gradient definition.