zng::color::gradient

Struct ConicGradient

pub struct ConicGradient { /* private fields */ }

Conic gradient.

Can be used as a node that fills the available space with the gradient, or can continue building the conic or a tiled conic gradient.

Use gradient or conic_gradient to start building.

Implementations

impl ConicGradient

pub fn extend_mode(self, mode: impl IntoVar<ExtendMode>) -> ConicGradient

Sets the extend mode of the conic gradient.

By default is ExtendMode::Clamp.

pub fn repeat(self) -> ConicGradient

Sets the extend mode to ExtendMode::Repeat.

pub fn reflect(self) -> ConicGradient

Sets the extend mode to ExtendMode::Reflect.

pub fn tile( self, tile_size: impl IntoVar<Size>, tile_spacing: impl IntoVar<Size>, ) -> TiledConicGradient

Continue building a tiled radial gradient.

pub fn tile_size(self, size: impl IntoVar<Size>) -> TiledConicGradient

Continue building a tiled radial gradient.