Struct zng::color::gradient::TiledConicGradient

pub struct TiledConicGradient<S, C, A, E, O, T, TS> { /* private fields */ }

Tiled conic gradient.

Can be used as a node that fills the available space with the gradient tiles, or can continue building the gradient.

Use ConicGradient::tile to build.

Implementations

impl<S, C, A, E, O, T, TS> TiledConicGradient<S, C, A, E, O, T, TS>
where S: Var<GradientStops>, C: Var<Point>, A: Var<AngleRadian>, E: Var<ExtendMode>, O: Var<Point>, T: Var<Size>, TS: Var<Size>,

pub fn tile_spacing<TS2>( self, spacing: TS2, ) -> TiledConicGradient<S, C, A, E, O, T, <TS2 as IntoVar<Size>>::Var>
where TS2: IntoVar<Size>,

Set the space between tiles.

Relative values are resolved on the tile size, so setting this to 100.pct() will skip a tile.

Leftover values are resolved on the space taken by tiles that do not fully fit in the available space, so setting this to 1.lft() will cause the border tiles to always touch the full bounds and the middle filled with the maximum full tiles that fit or empty space.

pub fn tile_origin<O2>( self, origin: O2, ) -> TiledConicGradient<S, C, A, E, <O2 as IntoVar<Point>>::Var, T, TS>
where O2: IntoVar<Point>,

Sets the tile offset.

Relative values are resolved on the tile size, so setting this to 100.pct() will offset a full turn.