Trait zng::config::ConfigValue

pub trait ConfigValue: VarValue + Serialize + DeserializeOwned { }

Marker trait for types that can stored in a Config.

This trait is already implemented for types it applies.

Object Safety

This trait is not object safe.