Trait zng::config::FallbackConfigReset

pub trait FallbackConfigReset: AnyConfig + Sync {
    // Required methods
    fn reset(&self, key: &Txt);
    fn can_reset(&self, key: Txt) -> Box<dyn VarBoxed<bool>>;
    fn clone_boxed(&self) -> Box<dyn FallbackConfigReset>;
}

Reset controls of a FallbackConfig.

Required Methods

fn reset(&self, key: &Txt)

Removes the key from the config and updates all active config variables back to the fallback value. Note that if you assign the config variable the key will be re-inserted on the config.

fn can_reset(&self, key: Txt) -> Box<dyn VarBoxed<bool>>

Returns a read-only var that is true when the key has an entry in the read-write config.

fn clone_boxed(&self) -> Box<dyn FallbackConfigReset>

Clone a reference to the config.