Function zng::data_context::extend_data_note_colors

pub fn extend_data_note_colors(
    child: impl UiNode,
    colors: impl IntoVar<HashMap<DataNoteLevel, ColorPair>>
) -> impl UiNode

P Extend the data note level colors, the colors extend the parent colors, only entries of the same level are replaced.

The colors will be used directly as text color.

This property sets the DATA_NOTE_COLORS_VAR.