zng::data_context

Struct DATA

pub struct DATA;

Data context and validation.

This service enables data flow from a context to descendants, and from descendants up-to contexts, like an anonymous context var.

Arbitrary data can be set on a context using the data property and retrieved using DATA.get or DATA.req. Only one data entry and type can exist in a context, nested data properties override the parent data and type in their context.

Annotation on the data can be set back using DATA.annotate and can be retrieved using the get_data_notes property, annotations are classified by DataNoteLevel, including INFO, WARN and ERROR. For each level there are specialized methods and properties, as an example, the DATA.invalidate is used to set an error note, and the get_data_error_txt property gets the error formatted for display. Data notes are aggregated from descendants up-to the context, continuing up to outer nested contexts too, this means that you can get data errors for a form field by setting get_data_error_txt on the field widget, and get all form errors from that field and others by also setting get_data_error_txt in the form widget.

Implementations

impl DATA

pub fn req<T>(&self) -> Var<T>
where T: VarValue,

Require context data of type T.

See get for capabilities of the returned var.

Panics

Panics if the context data is not set to a variable of type T on the first usage of the returned variable.

pub fn get<T>(&self, fallback: impl Fn() -> T + Send + Sync + 'static) -> Var<T>
where T: VarValue,

Gets a var of T, if the contextual DATA type is not T the fallback closure is called to produce a value.

Capabilities

The returned var is contextual, it will get the data var in the final context it is used not on the context that calls this method.

pub fn get_any(&self) -> AnyVar

Get the current context variable.

Note that this is different from get, the return variable is “actualized” on the calling context, not on the first var use context.

pub fn is<T>(&self) -> bool
where T: VarValue,

Gets if the current context data type is T.

pub fn value_type(&self) -> TypeId

Gets the current context value type.

pub fn value_type_name(&self) -> &'static str

Gets the current context value type name.

pub fn annotate( &self, level: DataNoteLevel, note: impl DataNoteValue, ) -> DataNoteHandle

Insert a data note in the current context.

The note will stay in context until the context is unloaded or the handle is dropped.

pub fn inform(&self, note: impl DataNoteValue) -> DataNoteHandle

Insert an INFO note in the current context.

The note will stay in context until the context is unloaded or the handle is dropped.

pub fn warn(&self, note: impl DataNoteValue) -> DataNoteHandle

Insert a WARN note in the current context.

The note will stay in context until the context is unloaded or the handle is dropped.

pub fn invalidate(&self, note: impl DataNoteValue) -> DataNoteHandle

Insert an ERROR note in the current context.

The note will stay in context until the context is unloaded or the handle is dropped.

pub fn note_color(&self, level: impl IntoVar<DataNoteLevel>) -> Var<Rgba>

Read-only variable that is the best color for the note level in the context of the current color scheme.

If the level is not found, gets the nearest less than level, if no color is set in the context gets the black/white for dark/light.

The color can be used directly as text color, it probably needs mixing or desaturating to use as background.

pub fn info_color(&self) -> Var<Rgba>

Read-only variable that is the best color for INFO notes in the context of the current color scheme.

The color can be used directly as text color, it probably needs mixing or desaturating to use as background.

pub fn warn_color(&self) -> Var<Rgba>

Read-only variable that is the best color for WARN notes in the context of the current color scheme.

The color can be used directly as text color, it probably needs mixing or desaturating to use as background.

pub fn error_color(&self) -> Var<Rgba>

Read-only variable that is the best color for ERROR notes in the context of the current color scheme.

The color can be used directly as text color, it probably needs mixing or desaturating to use as background.