Module zng::dialog::backdrop

Modal dialog parent widget that fills the window.

Note that the actual DialogBackdrop! widget is not included in this module because it is instantiated by the DIALOG service. The backdrop can be customized by setting the backdrop::style_fn.

use zng::prelude::*;

Window! {
    dialog::backdrop::style_fn = Style! {
        replace = true;
        color::filter::backdrop_blur = 2;
    };
}

The example above configures the backdrop to blur the window content when any dialog is open.

Full API

See zng_wgt_dialog::backdrop for the fill view API.

Structs

DefaultStyle

W Dialog backdrop default style.

Functions

style_fn

P Extends or replaces the widget style.