zng::drag_drop

Function on_pre_drag_end

pub fn on_pre_drag_end(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<DragEndArgs>,
) -> UiNode

P Preview on_drag_end event.

Preview

Preview event properties call the handler before the main event property and before the widget content, if you stop the propagation of a preview event the main event handler is not called.

Async

You can use async event handlers with this property, note that only the code before the fist .await is preview, subsequent code runs in widget updates.