zng::drag_drop

Struct DragDropEffect

pub struct DragDropEffect(/* private fields */);

Drag&drop drop effect on the data source.

Implementations

impl DragDropEffect

pub const COPY: DragDropEffect

Indicates that the dragged data will be copied from its present location to the drop location.

pub const MOVE: DragDropEffect

Indicates that the dragged data will be moved from its present location to the drop location.

pub const LINK: DragDropEffect

Indicates that some form of relationship or connection will be created between the source and drop locations.

impl DragDropEffect

pub const fn empty() -> DragDropEffect

Get a flags value with all bits unset.

pub const fn all() -> DragDropEffect

Get a flags value with all known bits set.

pub const fn bits(&self) -> u8

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u8) -> Option<DragDropEffect>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u8) -> DragDropEffect

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u8) -> DragDropEffect

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<DragDropEffect>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesnโt correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: DragDropEffect) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: DragDropEffect) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: DragDropEffect)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: DragDropEffect)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove wonโt truncate other, but the ! operator will.

pub fn toggle(&mut self, other: DragDropEffect)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: DragDropEffect, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: DragDropEffect) -> DragDropEffect

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: DragDropEffect) -> DragDropEffect

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: DragDropEffect) -> DragDropEffect

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference wonโt truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: DragDropEffect) -> DragDropEffect

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> DragDropEffect

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl DragDropEffect

pub const fn iter(&self) -> Iter<DragDropEffect>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<DragDropEffect>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl DragDropEffect

pub fn len(&self) -> u8

Count effects flagged.