Module zng::env

Process external directories and files.

This module contains functions to get external files associated with the installed app package. Note that these values are associated with the executable, not just the APP context.

fn main() {
   // optional, but recommended package metadata init.
   zng::env::init!();

   // get a path in the app config dir, the config dir is created if needed.    
   let my_config = zng::env::config("my-config.txt");

   // read a config file, or create it
   if let Ok(c) = std::fs::read_to_string(&my_config) {
      println!("{c}");
   } else {
      std::fs::write(zng::env::config("my-config.txt"), b"Hello!").unwrap();
   }
}

The example above uses init! to initialize the metadata used to find a good place for each directory, it then uses config to write and read a file.

Macros

init

Init about from cargo manifest.

Structs

About

Metadata about the app and main crate.

Functions

about

Gets metadata about the application.

bin

Gets a path relative to the package binaries.

cache

Gets a path relative to the cache directory for the app.

clear_cache

Removes all cache files possible.

config

Gets a path relative to the user config directory for the app.

init

Initialize about manually.

init_built_res

Sets a custom path for the “built resources” override checked by res in debug builds.

init_cache

Sets a custom cache path.

init_config

Sets a custom original_config path.

init_res

Sets a custom res path.

migrate_cache

Save new_path as the new cache path and make a best effort to move existing cache files.

migrate_config

Copied all config to new_path and saves it as the config path.

res

Gets a path relative to the package resources.