Macro zng::env::init

macro_rules! init {
    () => { ... };
}

Inits process metadata, calls process start handlers and defines the process lifetime in main.

This must be called in main.

Init about an About for the process metadata. See on_process_start! for process start handlers. See on_process_exit for exit handlers called at the end of the main function.

Process Start

A single Zng executable can be built with multiple components that spawn different instances of the executable that must run as different processes. If the current instance is requested by component init! runs it and exits the process, never returning flow to the normal main function.

fn main() {
    println!("print in all processes");
    zng::env::init!();
    println!("print only in the app-process");

    // directories are available after `init!`.
    let _res = zng::env::res("");
     
    // APP.defaults().run(...);

    // on_exit handlers are called here
}