zng::event

Trait AnyEventArgs

pub trait AnyEventArgs:
    Debug
    + Send
    + Sync
    + Any {
    // Required methods
    fn clone_any(&self) -> Box<dyn AnyEventArgs>;
    fn as_any(&self) -> &(dyn Any + 'static);
    fn timestamp(&self) -> DInstant;
    fn delivery_list(&self, list: &mut UpdateDeliveryList);
    fn propagation(&self) -> &EventPropagationHandle;
}

Methods of EventArgs that are object safe.

Required Methods

fn clone_any(&self) -> Box<dyn AnyEventArgs>

Clone the event into a type erased box.

fn as_any(&self) -> &(dyn Any + 'static)

Access to dyn Any methods.

fn timestamp(&self) -> DInstant

Gets the instant this event happened.

fn delivery_list(&self, list: &mut UpdateDeliveryList)

Insert all targets of this event on the UpdateDeliveryList.

fn propagation(&self) -> &EventPropagationHandle

Propagation handle associated with this event instance.

Cloned arguments share the same handle, some arguments may also share the handle of another event if they share the same cause.