zng::event

Trait CommandInfoExt

pub trait CommandInfoExt {
    // Required methods
    fn info(self) -> Var<Txt>;
    fn init_info(self, info: impl Into<Txt>) -> Self;
}

Adds the info command metadata.

Required Methods

fn info(self) -> Var<Txt>

Gets a read-write variable that is a short informational string about the command.

fn init_info(self, info: impl Into<Txt>) -> Self

Sets the initial info if it is not set.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.