Function zng::focus::is_return_focus

pub fn is_return_focus(
    child: impl UiNode,
    state: impl IntoVar<bool>,
) -> impl UiNode

P If the widget will be focused when a parent scope is focused.

Focus scopes can remember the last focused widget inside, the focus returns to this widget when the scope receives focus. Alt scopes also remember the widget from which the alt focus happened and can also return focus back to that widget.

Usually input widgets that have a visual state for is_focused also have a visual for this, a common example is the text-input widget that shows an emphasized border and blinking cursor when focused and still shows the emphasized border without cursor when a menu is open and it is only the return focus.

Note that a widget can be is_focused and is_return_focus, this property is true if any focus scope considers the widget its return focus, you probably want to declare the widget visual states in such a order that is_focused overrides the state of this property.