Enum zng::font::font_features::FontStyleSet

#[repr(u8)]
pub enum FontStyleSet {

    Auto = 0,
    S01 = 1,
    S02 = 2,
    S03 = 3,
    S04 = 4,
    S05 = 5,
    S06 = 6,
    S07 = 7,
    S08 = 8,
    S09 = 9,
    S10 = 10,
    S11 = 11,
    S12 = 12,
    S13 = 13,
    S14 = 14,
    S15 = 15,
    S16 = 16,
    S17 = 17,
    S18 = 18,
    S19 = 19,
    S20 = 20,

}

All possible style_set features.

The styles depend on the font, it is recommended you create an enum with named sets that converts into this one for each font you wish to use.

Variants

Auto = 0

Donโt use alternative style set.

S01 = 1

S02 = 2

S03 = 3

S04 = 4

S05 = 5

S06 = 6

S07 = 7

S08 = 8

S09 = 9

S10 = 10

S11 = 11

S12 = 12

S13 = 13

S14 = 14

S15 = 15

S16 = 16

S17 = 17

S18 = 18

S19 = 19

S20 = 20