Struct zng::font::font_features::FontFeatures

pub struct FontFeatures(/* private fields */);

Font features configuration.

Implementations

impl FontFeatures

pub fn new() -> FontFeatures

New default.

pub fn builder() -> FontFeaturesBuilder

New builder.

pub fn set_all( &mut self, other: &FontFeatures ) -> Vec<(FontFeatureName, Option<u32>)>

Set or override the features of self from other.

Returns the previous state of all affected names.

pub fn restore(&mut self, prev: Vec<(FontFeatureName, Option<u32>)>)

Restore feature states that where overridden in set_all.

pub fn feature(&mut self, name: FontFeatureName) -> FontFeature<'_>

Access to the named feature.

pub fn feature_set( &mut self, names: &'static [FontFeatureName] ) -> FontFeatureSet<'_>

Access to a set of named features that are managed together.

Panics

If names has less than 2 names.

pub fn feature_exclusive_set<S>(&mut self) -> FontFeatureExclusiveSet<'_, S>
where S: FontFeatureExclusiveSetState,

Access to a set of named features where only one of the features can be enabled at a time.

Panics

If S::names() has less than 2 names.

pub fn feature_exclusive_sets<S>(&mut self) -> FontFeatureExclusiveSets<'_, S>
where S: FontFeatureExclusiveSetsState,

Access to a set of named features where only one or more features can be enabled but each combination represents a single distinct state.

Panics

If S::names() has less than 2 entries.

pub fn finalize(&self) -> Vec<Feature>

Generate the harfbuzz font features.

impl FontFeatures

pub fn caps(&mut self) -> FontFeatureExclusiveSets<'_, CapsVariant>

Font capital glyph variants.

See CapsVariant for more details.

pub fn kerning(&mut self) -> FontFeature<'_>

Allow glyphs boundaries to overlap for a more pleasant reading.

This corresponds to the kern feature.

Auto always activates these kerning.

pub fn common_lig(&mut self) -> FontFeatureSet<'_>

The most common ligatures, like for fi, ffi, th or similar.

This corresponds to OpenType liga and clig features.

Auto always activates these ligatures.

pub fn discretionary_lig(&mut self) -> FontFeature<'_>

Ligatures specific to the font, usually decorative.

This corresponds to OpenType dlig feature.

Auto usually disables these ligatures.

pub fn historical_lig(&mut self) -> FontFeature<'_>

Ligatures used historically, in old books, like the German tz digraph being displayed ß.

This corresponds to OpenType hlig feature.

Auto usually disables these ligatures.

pub fn contextual_alt(&mut self) -> FontFeature<'_>

Alternative letters that adapt to their surrounding letters.

This corresponds to OpenType calt feature.

Auto usually activates this feature.

pub fn ordinal(&mut self) -> FontFeature<'_>

Force usage of ordinal special glyphs, 1a becomes 1ª.

This corresponds to OpenType ordn feature.

Auto deactivates this feature.

pub fn slashed_zero(&mut self) -> FontFeature<'_>

Force use of a slashed zero for 0.

This corresponds to OpenType zero feature.

Auto deactivates this feature.

pub fn swash(&mut self) -> FontFeatureSet<'_>

Use swashes flourish style.

Fonts can have alternative swash styles, you can select then by enabling a number.

This corresponds to OpenType swsh and cswh feature.

Auto does not use swashes.

pub fn stylistic(&mut self) -> FontFeature<'_>

Use stylistic alternatives.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType salt feature.

Auto does not use alternative styles.

pub fn historical_forms(&mut self) -> FontFeature<'_>

Use glyphs that were common in the past but not today.

This corresponds to OpenType hist feature.

Auto does not use alternative styles.

pub fn ornaments(&mut self) -> FontFeature<'_>

Replace letter with fleurons, dingbats and border elements.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType ornm feature.

Auto does not enable this by default, but some fonts are purely dingbats glyphs.

pub fn annotation(&mut self) -> FontFeature<'_>

Font annotation alternatives, like circled digits or inverted characters.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType nalt feature.

Auto does not use alternative styles.

pub fn numeric(&mut self) -> FontFeatureExclusiveSet<'_, NumVariant>

Font numeric glyph variants.

See NumVariant for more details.

pub fn num_spacing(&mut self) -> FontFeatureExclusiveSet<'_, NumSpacing>

Font numeric spacing variants.

See NumSpacing for more details.

pub fn num_fraction(&mut self) -> FontFeatureExclusiveSet<'_, NumFraction>

Font numeric spacing variants.

See NumSpacing for more details.

pub fn style_set(&mut self) -> FontFeatureExclusiveSet<'_, FontStyleSet>

Font stylistic alternatives for sets of characters.

See FontStyleSet for more details.

pub fn char_variant(&mut self) -> FontFeatureExclusiveSet<'_, CharVariant>

Font stylistic alternatives for individual characters.

See CharVariant for more details.

pub fn position(&mut self) -> FontFeatureExclusiveSet<'_, FontPosition>

Font sub/super script alternatives.

See FontPosition for more details.

pub fn ruby(&mut self) -> FontFeature<'_>

Force the use of ruby (rubi) glyph variants.

This corresponds to OpenType ruby feature.

Auto does not force the use of ruby variants.

pub fn jp_variant(&mut self) -> FontFeatureExclusiveSet<'_, JpVariant>

Japanese logographic set selection.

See JpVariant for more details.

pub fn horizontal_kana(&mut self) -> FontFeature<'_>

Use kana glyphs optimized for horizontal writing.

This corresponds to OpenType hkna feature.

pub fn cn_variant(&mut self) -> FontFeatureExclusiveSet<'_, CnVariant>

Chinese logographic set selection.

See CnVariant for more details.

pub fn ea_width(&mut self) -> FontFeatureExclusiveSet<'_, EastAsianWidth>

East Asian figure width control

See EastAsianWidth for more details.