zng::font::font_features

Struct FontFeaturesBuilder

pub struct FontFeaturesBuilder(/* private fields */);

A builder for FontFeatures.

Examples

let features = FontFeatures::builder().kerning(false).build();

Implementations

impl FontFeaturesBuilder

pub fn build(self) -> FontFeatures

Finish building.

pub fn feature( self, name: FontFeatureName, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Set the named feature.

pub fn feature_set( self, names: &'static [FontFeatureName], state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Sets all the named features to the same value.

Panics

If names has less than 2 names.

pub fn feature_exclusive_set<S>( self, state: impl Into<S>, ) -> FontFeaturesBuilder
where S: FontFeatureExclusiveSetState,

Sets a single feature of a set of features.

Panics

If S::names() has less than 2 names.

pub fn feature_exclusive_sets<S>( self, state: impl Into<S>, ) -> FontFeaturesBuilder
where S: FontFeatureExclusiveSetsState,

Sets the features that represent the state.

Panics

If S::names() has less than 2 entries.

impl FontFeaturesBuilder

pub fn caps(self, state: impl Into<CapsVariant>) -> FontFeaturesBuilder

Font capital glyph variants.

See CapsVariant for more details.

pub fn kerning(self, state: impl Into<FontFeatureState>) -> FontFeaturesBuilder

Allow glyphs boundaries to overlap for a more pleasant reading.

This corresponds to the kern feature.

Auto always activates these kerning.

pub fn common_lig( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

The most common ligatures, like for fi, ffi, th or similar.

This corresponds to OpenType liga and clig features.

Auto always activates these ligatures.

pub fn discretionary_lig( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Ligatures specific to the font, usually decorative.

This corresponds to OpenType dlig feature.

Auto usually disables these ligatures.

pub fn historical_lig( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Ligatures used historically, in old books, like the German tz digraph being displayed ß.

This corresponds to OpenType hlig feature.

Auto usually disables these ligatures.

pub fn contextual_alt( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Alternative letters that adapt to their surrounding letters.

This corresponds to OpenType calt feature.

Auto usually activates this feature.

pub fn ordinal(self, state: impl Into<FontFeatureState>) -> FontFeaturesBuilder

Force usage of ordinal special glyphs, 1a becomes 1ª.

This corresponds to OpenType ordn feature.

Auto deactivates this feature.

pub fn slashed_zero( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Force use of a slashed zero for 0.

This corresponds to OpenType zero feature.

Auto deactivates this feature.

pub fn swash(self, state: impl Into<FontFeatureState>) -> FontFeaturesBuilder

Use swashes flourish style.

Fonts can have alternative swash styles, you can select then by enabling a number.

This corresponds to OpenType swsh and cswh feature.

Auto does not use swashes.

pub fn stylistic( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Use stylistic alternatives.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType salt feature.

Auto does not use alternative styles.

pub fn historical_forms( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Use glyphs that were common in the past but not today.

This corresponds to OpenType hist feature.

Auto does not use alternative styles.

pub fn ornaments( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Replace letter with fleurons, dingbats and border elements.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType ornm feature.

Auto does not enable this by default, but some fonts are purely dingbats glyphs.

pub fn annotation( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Font annotation alternatives, like circled digits or inverted characters.

Fonts can have multiple alternative styles, you can select then by enabling a number.

This corresponds to OpenType nalt feature.

Auto does not use alternative styles.

pub fn numeric(self, state: impl Into<NumVariant>) -> FontFeaturesBuilder

Font numeric glyph variants.

See NumVariant for more details.

pub fn num_spacing(self, state: impl Into<NumSpacing>) -> FontFeaturesBuilder

Font numeric spacing variants.

See NumSpacing for more details.

pub fn num_fraction(self, state: impl Into<NumFraction>) -> FontFeaturesBuilder

Font numeric spacing variants.

See NumSpacing for more details.

pub fn style_set(self, state: impl Into<FontStyleSet>) -> FontFeaturesBuilder

Font stylistic alternatives for sets of characters.

See FontStyleSet for more details.

pub fn char_variant(self, state: impl Into<CharVariant>) -> FontFeaturesBuilder

Font stylistic alternatives for individual characters.

See CharVariant for more details.

pub fn position(self, state: impl Into<FontPosition>) -> FontFeaturesBuilder

Font sub/super script alternatives.

See FontPosition for more details.

pub fn ruby(self, state: impl Into<FontFeatureState>) -> FontFeaturesBuilder

Force the use of ruby (rubi) glyph variants.

This corresponds to OpenType ruby feature.

Auto does not force the use of ruby variants.

pub fn jp_variant(self, state: impl Into<JpVariant>) -> FontFeaturesBuilder

Japanese logographic set selection.

See JpVariant for more details.

pub fn horizontal_kana( self, state: impl Into<FontFeatureState>, ) -> FontFeaturesBuilder

Use kana glyphs optimized for horizontal writing.

This corresponds to OpenType hkna feature.

pub fn cn_variant(self, state: impl Into<CnVariant>) -> FontFeaturesBuilder

Chinese logographic set selection.

See CnVariant for more details.

pub fn ea_width(self, state: impl Into<EastAsianWidth>) -> FontFeaturesBuilder

East Asian figure width control

See EastAsianWidth for more details.