Struct zng::font::ShapedSegment

pub struct ShapedSegment<'a> { /* private fields */ }

Represents a word or space selection of a ShapedText.

Implementations

impl<'a> ShapedSegment<'a>

pub fn kind(&self) -> TextSegmentKind

Segment kind.

pub fn level(&self) -> Level

Segment bidi level.

pub fn direction(&self) -> LayoutDirection

Layout direction of glyphs in the segment.

pub fn has_last_glyph(&self) -> bool

If the segment contains the last glyph of the line.

pub fn glyphs(&self) -> impl Iterator<Item = (&'a Font, &'a [GlyphInstance])>

Glyphs in the word or space.

The glyphs are in visual order (LTR) within segments, so the RTL text โูู ุงโ will yield โโรกlifโ, โmiimโ, โlรกamโ.

All glyph points are set as offsets to the top-left of the text full text.

Note that multiple glyphs can map to the same char and multiple chars can map to the same glyph, you can use the clusters map to find the char for each glyph. Some font ligatures also bridge multiple segments, in this case only the first shaped segment has glyphs the subsequent ones are empty.

pub fn glyph(&self, index: usize) -> Option<(&'a Font, GlyphInstance)>

Gets the specific glyph and font.

pub fn clusters(&self) -> &[u32]

Map glyph -> char.

Each glyphs glyph pairs with an entry in this slice that is the char byte index in text. If a font ligature bridges multiple segments only the first segment will have a non-empty map.

pub fn ligature_segs_count(&self) -> usize

Number of next segments that are empty because their text is included in a ligature glyph or glyphs started in this segment.

pub fn glyphs_with_x_advance( &self, ) -> impl Iterator<Item = (&'a Font, impl Iterator<Item = (GlyphInstance, f32)> + 'a)> + 'a

Glyphs in the segment, paired with the x-advance.

Yields (Font, [(glyph, advance)]).

pub fn cluster_glyphs_with_x_advance( &self, ) -> impl Iterator<Item = (&Font, impl Iterator<Item = (u32, &[GlyphInstance], f32)>)>

Glyphs per cluster in the segment, paired with the x-advance of the cluster.

Yields (Font, [(cluster, [glyph], advance)]).

pub fn x_width(&self) -> (Px, Px)

Gets the segment x offset and advance.

pub fn advance(&self) -> f32

Segment exact width in pixels.

pub fn rect(&self) -> Rect<Px, Px>

Bounds of the word or spaces.

pub fn overflow_char_glyph(&self, max_width_px: f32) -> Option<(usize, usize)>

Gets the first char and glyph with advance that overflows max_width.

pub fn inline_info(&self) -> InlineSegmentInfo

Segment info for widget inline segments.

pub fn overline(&self) -> (Point2D<Px, Px>, Px)

Overline spanning the word or spaces, start point + width.

pub fn strikethrough(&self) -> (Point2D<Px, Px>, Px)

Strikethrough spanning the word or spaces, start point + width.

pub fn underline(&self) -> (Point2D<Px, Px>, Px)

Underline spanning the word or spaces, not skipping.

The y is defined by the font metrics.

Returns start point + width.

pub fn underline_skip_glyphs( &self, thickness: Px, ) -> impl Iterator<Item = (Point2D<Px, Px>, Px)> + 'a

Underline spanning the word or spaces, skipping glyph descends that intercept the line.

Returns an iterator of start point + width for underline segments.

pub fn underline_descent(&self) -> (Point2D<Px, Px>, Px)

Underline spanning the word or spaces, not skipping.

The y is the baseline + descent + 1px.

Returns start point + width.

pub fn text_range(&self) -> Range<usize>

Get the text bytes range of this segment in the original text.

pub fn text_start(&self) -> usize

Get the text byte range start of this segment in the original text.

pub fn text_end(&self) -> usize

Get the text byte range end of this segment in the original text.

pub fn text_glyph_range( &self, glyph_range: impl RangeBounds<usize>, ) -> Range<usize>

Get the text bytes range of the glyph_range in this segmentโs text.

pub fn text<'s>(&self, full_text: &'s str) -> &'s str

Select the string represented by this segment.

The full_text must be equal to the original text that was used to generate the parent ShapedText.

pub fn nearest_char_index(&self, x: Px, full_text: &str) -> usize

Gets the insert index in the segment text that is nearest to x.

pub fn index(&self) -> usize

Gets the segment index in the line.