zng::fs_watcher

Struct WriteFile

pub struct WriteFile { /* private fields */ }

Represents an open write file provided by WATCHER.sync.

This struct writes to a temporary file and renames it over the actual file on commit only. The dereferenced fs::File is the temporary file, not the actual one.

Transaction

To minimize the risk of file corruption exclusive locks are used, both the target file and the temp file are locked. An empty lock file is also used to cover the moment when both files are unlocked for the rename operation and the moment the temp file is acquired.

The temp file is the actual file path with file extension replaced with {path/.file-name.ext}.{GUID}-{n}.tmp, the n is a number from 0 to 999, if a temp file exists unlocked it will be reused.

The lock file is {path/.file-name.ext}.{GUID}-lock.tmp. Note that this lock file only helps for apps that use WriteFile, but even without it the risk is minimal as the slow write operations are already flushed when it is time to commit.

Implementations

impl WriteFile

pub fn open(path: PathBuf) -> Result<WriteFile, Error>

Open or create the file.

pub fn write_text(&mut self, txt: &str) -> Result<(), Error>

Write the text string.

pub fn write_json<O>(&mut self, value: &O, pretty: bool) -> Result<(), Error>
where O: Serialize,

Serialize and write.

If pretty is true the JSON is formatted for human reading.

pub fn write_toml<O>(&mut self, value: &O, pretty: bool) -> Result<(), Error>
where O: Serialize,

Serialize and write.

If pretty is true the TOML is formatted for human reading.

pub fn write_ron<O>(&mut self, value: &O, pretty: bool) -> Result<(), Error>
where O: Serialize,

Serialize and write.

If pretty is true the RON if formatted for human reading using the default pretty config.

pub fn write_yaml<O>(&mut self, value: &O) -> Result<(), Error>
where O: Serialize,

Serialize and write.

pub fn commit(self) -> Result<(), Error>

Commit write, flush and replace the actual file with the new one.

pub fn cancel(self)

Cancel write, the file will not be updated.