zng::fs_watcher

Trait IMAGES_Ext

pub trait IMAGES_Ext {
    // Required methods
    fn watch(&self, path: impl Into<PathBuf>) -> ImageVar;
    fn watch_image(
        &self,
        path: impl Into<PathBuf>,
        options: ImageOptions,
        limits: Option<ImageLimits>,
    ) -> ImageVar;
}

File watcher extensions for IMAGES service.

Required Methods

fn watch(&self, path: impl Into<PathBuf>) -> ImageVar

Like IMAGES.read with automatic reload when the file at path is modified.

fn watch_image( &self, path: impl Into<PathBuf>, options: ImageOptions, limits: Option<ImageLimits>, ) -> ImageVar

Like IMAGES.image with automatic cache reload when the file at path is modified.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.