Struct zng::gesture::GESTURES

pub struct GESTURES;

Gesture events config service.

This service is provided by GestureManager.

Shortcuts

This service coordinates shortcut associations with widgets and commands. To define a command’s shortcut use the CommandShortcutExt methods. To define the shortcut that focus and clicks a widget use click_shortcut. To define the shortcut that only focuses on a widget use focus_shortcut. To define a custom handle for a shortcut use on_pre_event or on_event with the SHORTCUT_EVENT.

Event Order

The same shortcut can end-up registered for multiple targets, activation of a shortcut causes these effects in this order:

The gestures manager receives a KEY_INPUT_EVENT in the event track, if a shortcut is completed it gets combined with any primed chords and become the shortcut that will cause the following actions:

The click, command and focus shortcuts are resolved in this order:

First exclusively:

Primary click_shortcut targeting a widget that is enabled and focused.

Command scoped in a widget that is enabled and focused.

Contextual click_shortcut targeting a widget that is enabled and focused.

Primary, command scoped or contextual targeting a widget that is enabled and closest to the focused widget, child first (of the focused).

Primary, command scoped or contextual targeting a widget that is enabled in the focused window.

Primary, command scoped or contextual targeting a widget that is enabled.

focus_shortcut targeting a widget that is in the focused window.

focus_shortcut targeting a widget that is enabled.

The click_focused and context_click_focused.

Same as the above, but for disabled widgets

And then:

a. All enabled commands targeting the focused window.

b. All enabled commands targeting the app.

The app level SHORTCUT_EVENT is notified, with the list of actions that will run, app extensions can handle it before event to stop the resolved actions.

The gestures manager receives the shortcut in event, if propagation is not stopped and it contains any actions they are run, the click and command events are linked by the same propagation. If the shortcut contains no action and it

If the shortcut is a KeyChord::starter for one of the registered shortcuts, and was not claimed by any of the above, the chord starter is primed for the next shortcut press.

The event propagation flag of shortcut, click and command events are linked, so stopping propagation in one signal all others.

Implementations

impl GESTURES

pub fn click_focused(&self) -> ArcVar<Shortcuts>

Shortcuts that generate a primary CLICK_EVENT for the focused widget. The shortcut only works if no widget or command claims it.

Clicks generated by this shortcut count as primary.

Initial shortcuts are Enter and Space.

pub fn context_click_focused(&self) -> ArcVar<Shortcuts>

Shortcuts that generate a context CLICK_EVENT for the focused widget. The shortcut only works if no widget or command claims it.

Clicks generated by this shortcut count as context.

Initial shortcut is ContextMenu.

pub fn shortcut_pressed_duration(&self) -> ArcVar<Duration>

When a shortcut or access primary click happens, targeted widgets can indicate that they are pressed for this duration.

Initial value is 50ms, set to 0 to deactivate this type of indication.

pub fn click_shortcut( &self, shortcuts: impl Into<Shortcuts>, kind: ShortcutClick, target: WidgetId ) -> ShortcutsHandle

Register a widget to receive shortcut clicks when any of the shortcuts are pressed.

pub fn focus_shortcut( &self, shortcuts: impl Into<Shortcuts>, target: WidgetId ) -> ShortcutsHandle

Register a widget to receive keyboard focus when any of the shortcuts are pressed.

If the widget is not focusable the focus moves to the first focusable descendant or the first focusable ancestor.

pub fn shortcut_actions(&self, shortcut: Shortcut) -> ShortcutActions

Gets all the event notifications that are send if the shortcut was pressed at this moment.

See the struct level docs for details of how shortcut targets are resolved.