Trait zng::gesture::CommandShortcutMatchesExt

pub trait CommandShortcutMatchesExt: CommandShortcutExt {
    // Required method
    fn shortcut_matches(self, shortcut: &Shortcut) -> bool;
}

Adds the shortcut_matches method to commands.

Required Methods

fn shortcut_matches(self, shortcut: &Shortcut) -> bool

Returns true if the command has handlers, enabled or disabled, and the shortcut if one of the command shortcuts.

Object Safety

This trait is not object safe.