zng::grid

Module cell

Cell widget and properties.

Structs

Cell

W Grid cell container.

CellInfo

Represents values set by cell properties in a widget.

Constants

AT_AUTO

Value for at that causes the cell to be positioned based on the logical index i, for columns i % columns and for rows i / columns.

Functions

at

P Cell column and row indexes.

column

P Cell column index.

column_span

P Cell column span.

row

P Cell row index.

row_span

P Cell row span.

span

P Cell column and row span.