zng::grid
Module row

Row widget and properties.

Structs

Row

W Grid row definition.

Functions

get_index

P Get the row index.

get_index_len

P Get the row index and number of rows.

get_rev_index

P Get the row index, starting from the last row at 0.

is_even

P If the row index is even.

is_first

P If the row is the first.

is_last

P If the row is the last.

is_odd

P If the row index is odd.