zng::handler

Trait AppWeakHandle

pub trait AppWeakHandle:
    Send
    + Sync
    + 'static {
    // Required methods
    fn clone_boxed(&self) -> Box<dyn AppWeakHandle>;
    fn unsubscribe(&self);
}

Represents a weak handle to a Handler subscription in the app context.

Inside the handler use APP_HANDLER to access this handle.

Required Methods

fn clone_boxed(&self) -> Box<dyn AppWeakHandle>

Dynamic clone.

fn unsubscribe(&self)

Unsubscribes the Handler.

This stops the handler from being called again and causes it to be dropped in a future app update.