zng::image

Function img_scale_density

pub fn img_scale_density(
    child: impl IntoUiNode,
    enabled: impl IntoVar<bool>,
) -> UiNode

P Sets if the image desired size is scaled considering the image and monitor pixel density.

The image desired size is its original size, after img_crop, and it can be in pixels or scaled considering the image pixel density, monitor pixel density and scale factor.

By default this is false, if true the image is scaled in a attempt to recreate the original physical dimensions, this only works if the image and monitor pixel density are set correctly. The monitor pixel density can be set using the MONITORS service.

This value supersedes img_scale_factor, this this enabled the scale factor is ignored.