Module zng::image::mask

Mask image properties.

See zng_wgt_image::mask for the full API.

Enums

ImageMaskMode

Defines how the A8 image mask pixels are to be derived from a source mask image.

Functions

mask_align

P Defines the align of the mask image in relation to the widget bounds in all mask_image inside the widget in descendants.

mask_fit

P Defines how the mask image fits the widget bounds in all mask_image inside the widget in descendants.

mask_image

P Sets an image mask.

mask_image_cache

P Defines if the mask images loaded in all mask_image inside the widget in descendants are cached.

mask_image_downscale

P Custom pixel resize applied during mask image load/decode.

mask_image_limits

P Sets custom mask image load and decode limits.

mask_mode

P Defines how the A8 image mask pixels are to be derived from a source mask image in all mask_image inside the widget in descendants.

mask_offset

P Defines the offset applied to the mask image after all measure and arrange. in all mask_image inside the widget in descendants.