Type Alias zng::image::PathFilter

pub type PathFilter = ImageSourceFilter<PathBuf>;

Represents a ImageSource::Read path request filter.

Only absolute, normalized paths are shared with the Custom filter, there is no relative paths or .. components.

The paths are not canonicalized and existence is not verified, no system requests are made with unfiltered paths.

See ImageLimits::allow_path for more information.

Aliased Type

enum PathFilter {
    BlockAll,
    AllowAll,
    Custom(Arc<dyn Fn(&PathBuf) -> bool + Send + Sync>),
}

Variants

BlockAll

Block all requests of this type.

AllowAll

Allow all requests of this type.

Custom(Arc<dyn Fn(&PathBuf) -> bool + Send + Sync>)

Custom filter, returns true to allow a request, false to block.