Struct zng::keyboard::KeyInputArgs

pub struct KeyInputArgs {
    pub timestamp: DInstant,
    pub window_id: WindowId,
    pub device_id: DeviceId,
    pub key_code: KeyCode,
    pub state: KeyState,
    pub key: Key,
    pub key_modified: Key,
    pub text: Txt,
    pub modifiers: ModifiersState,
    pub repeat_count: u32,
    pub target: InteractionPath,
    /* private fields */
}

Arguments for KEY_INPUT_EVENT.

Fields
timestamp: DInstant

Instant the event happened.

window_id: WindowId

Window that received the event.

device_id: DeviceId

Device that generated the event.

key_code: KeyCode

Physical key.

state: KeyState

If the key was pressed or released.

key: Key

Semantic key.

Pressing Shift+A key will produce Key::Char('a') in QWERTY keyboards, the modifiers are not applied.

key_modified: Key

Semantic key modified by the current active modifiers.

Pressing Shift+A key will produce Key::Char('A') in QWERTY keyboards, the modifiers are applied.

text: Txt

Text typed.

This is only set during KeyState::Pressed of a key that generates text.

This is usually the key_modified char, but is also '\r' for Key::Enter. On Windows when a dead key was pressed earlier but cannot be combined with the character from this key press, the produced text will consist of two characters: the dead-key-character followed by the character resulting from this key press.

modifiers: ModifiersState

What modifier keys where pressed when this event happened.

repeat_count: u32

Number of repeats generated by holding the key pressed.

This is zero for the first key press, increments by one for each event while the key is held pressed.

target: InteractionPath

The focused element at the time of the key input.

Implementations

impl KeyInputArgs

pub fn new( timestamp: impl Into<DInstant>, propagation_handle: EventPropagationHandle, window_id: impl Into<WindowId>, device_id: impl Into<DeviceId>, key_code: impl Into<KeyCode>, state: impl Into<KeyState>, key: impl Into<Key>, key_modified: impl Into<Key>, text: impl Into<Txt>, modifiers: impl Into<ModifiersState>, repeat_count: impl Into<u32>, target: impl Into<InteractionPath> ) -> KeyInputArgs

New args from values that convert into the argument types.

pub fn now( window_id: impl Into<WindowId>, device_id: impl Into<DeviceId>, key_code: impl Into<KeyCode>, state: impl Into<KeyState>, key: impl Into<Key>, key_modified: impl Into<Key>, text: impl Into<Txt>, modifiers: impl Into<ModifiersState>, repeat_count: impl Into<u32>, target: impl Into<InteractionPath> ) -> KeyInputArgs

Arguments for event that happened now (INSTANT.now).

impl KeyInputArgs

pub fn is_enabled(&self, widget_id: WidgetId) -> bool

Returns true if the widget is enabled in target.

pub fn is_disabled(&self, widget_id: WidgetId) -> bool

Returns true if the widget is disabled in target.

impl KeyInputArgs

Text input methods.

The text field contains the raw text associated with the key-press by the operating system, these methods normalize and filter this text.

pub fn is_backspace(&self) -> bool

Returns true if the character is the backspace and CTRL is not pressed.

pub fn is_delete(&self) -> bool

Returns true if the character is delete and CTRL is not pressed.

pub fn is_tab(&self) -> bool

Returns true if the character is a tab space and CTRL is not pressed.

pub fn is_line_break(&self) -> bool

Returns true if the character is a line-break and CTRL is not pressed.

pub fn insert_str(&self) -> &str

Gets the characters to insert in a typed text.

Replaces all is_tab with \t and all is_line_break with \n. Returns "" if there is no text or it contains ASCII control characters or CTRL is pressed.