zng

Module label

Label widget and properties.

The Label! widget is a text presenter that is focusable, when it receives focus it can transfer it to another target widget.

use zng::prelude::*;

Container! {
    child_start = zng::label::Label!("Name", "name-field");
    child_spacing = 5;
    child = TextInput! {
        id = "name-field";
        txt = var_from("");
    };
}

Full API

See zng_wgt_text_input::label for the full widget API.

Structs

DefaultStyle

W Default label style.

Label

W Styleable and focusable read-only text widget.

Functions

style_fn

P Extends or replaces the widget style.