zng::layer

Struct AnchorMode

pub struct AnchorMode {
    pub transform: AnchorTransform,
    pub min_size: AnchorSize,
    pub max_size: AnchorSize,
    pub viewport_bound: bool,
    pub visibility: bool,
    pub interactivity: bool,
    pub corner_radius: bool,
}

Defines what properties the layered widget takes from the anchor widget.

Fields
transform: AnchorTransform

What transforms are copied from the anchor widget and applied as a parent transform of the widget.

min_size: AnchorSize

What size is copied from the anchor widget and used as the available and final min size of the widget.

max_size: AnchorSize

What size is copied from the anchor widget and used as the available and final max size of the widget.

viewport_bound: bool

After the transform and size are resolved the transform is adjusted so that the layered widget is fully visible in the window.

Has no effect if AnchorTransform::None.

visibility: bool

If the widget is only layout if the anchor widget is not Collapsed and is only rendered if the anchor widget is rendered.

interactivity: bool

The widget interactivity is set to the same as the anchor widget.

corner_radius: bool

The widget’s corner radius is set for the layer.

If size is InnerBorder the corner radius are deflated to fit the inner curve of the borders.

Implementations

impl AnchorMode

pub fn window() -> AnchorMode

Mode where widget behaves like an unanchored widget, except that it is still only layout an rendered if the anchor widget exists in the same window.

pub fn foreground() -> AnchorMode

Mode where the widget behaves like a foreground to the target widget.

pub fn popup(place: AnchorOffset) -> AnchorMode

Mode where widget behaves like a flyout menu for the anchor.

pub fn tooltip() -> AnchorMode

Mode where the widget behaves like a tooltip anchored to the cursor.

pub fn tooltip_shortcut() -> AnchorMode

Mode where the widget behaves like a tooltip anchored to the widget.

pub fn context_menu() -> AnchorMode

Mode where the widget behaves like a context-menu anchored to the cursor.

pub fn context_menu_shortcut() -> AnchorMode

Mode where the widget behaves like a context-menu anchored to widget.

pub fn with_transform(self, transform: impl Into<AnchorTransform>) -> AnchorMode

Returns the mode with transform set.

pub fn with_min_size(self, size: impl Into<AnchorSize>) -> AnchorMode

Returns the mode with min_size set.

pub fn with_max_size(self, size: impl Into<AnchorSize>) -> AnchorMode

Returns the mode with max_size set.

pub fn with_size(self, size: impl Into<AnchorSize>) -> AnchorMode

Returns the mode with min_size and max_size set.

pub fn with_visibility(self, visibility: bool) -> AnchorMode

Returns the mode with visibility set.

pub fn with_interactivity(self, interactivity: bool) -> AnchorMode

Returns the mode with interactivity set.

pub fn with_corner_radius(self, corner_radius: bool) -> AnchorMode

Returns the mode with corner_radius set.

pub fn with_viewport_bound(self, viewport_bound: bool) -> AnchorMode

Returns the mode with viewport_bound set.