zng::layout

Function max_width

pub fn max_width(
    child: impl IntoUiNode,
    max_width: impl IntoVar<Length>,
) -> UiNode

P Maximum width of the widget.

The widget width can be smaller then this but not larger. Relative values are computed from the constraints maximum bounded width.

This property does not force the maximum constrained width, the max_width is only used if it is less then the constrained maximum, or the maximum was not constrained.

This property disables inline layout for the widget.

max_size

You can set both max_width and max_height at the same time using the max_size property.