zng::layout

Function min_size

pub fn min_size(child: impl IntoUiNode, min_size: impl IntoVar<Size>) -> UiNode

P Minimum size of the widget.

The widget size can be larger then this but not smaller. Relative values are computed from the constraints maximum bounded size.

This property does not force the minimum constrained size, the min_size is only used in a dimension if it is greater then the constrained minimum.

This property disables inline layout for the widget.

min_width and min_height

You can use the min_width and min_height properties to only set the minimum size of one dimension.