zng::layout

Function min_width

pub fn min_width(
    child: impl IntoUiNode,
    min_width: impl IntoVar<Length>,
) -> UiNode

P Minimum width of the widget.

The widget width can be larger then this but not smaller. Relative values are computed from the constraints maximum bounded width.

This property does not force the minimum constrained width, the min_width is only used if it is greater then the constrained minimum.

This property disables inline layout for the widget.

min_size

You can set both min_width and min_height at the same time using the min_size property.