zng::layout

Struct AngleTurn

pub struct AngleTurn(pub f32);

Angle in turns (complete rotations).

See AngleUnits for more details.

Equality

Equality is determined using about_eq with 0.00001 granularity.

Tuple Fields
0: f32
Implementations

impl AngleTurn

pub fn modulo(self) -> AngleTurn

Turns in [0.0 ..= 1.0].

pub fn lerp(self, to: AngleTurn, factor: Factor) -> AngleTurn

Linear interpolation.

pub fn slerp(self, to: AngleTurn, factor: Factor) -> AngleTurn

Spherical linear interpolation.

Always uses the shortest path from self to to.

The lerp linear interpolation always covers the numeric range between angles, so a transition from 358º to 1º iterates over almost a full counterclockwise turn to reach the final value, slerp simply goes from 358º to 361º modulo normalized.