zng::layout
Struct Factor

#[repr(transparent)]
pub struct Factor(pub f32);

Normalized multiplication factor.

Values of this type are normalized to generally be in between 0.0 and 1.0 to indicate a fraction of a unit. However, values are not clamped to this range, Factor(2.0) is a valid value and so are negative values.

Equality

Equality is determined to within 0.00001 epsilon.

Tuple Fields
0: f32
Implementations

impl Factor

pub fn clamp_range(self) -> Factor

Clamp factor to [0.0..=1.0] range.

pub fn max(self, other: impl Into<Factor>) -> Factor

Returns the maximum of two factors.

pub fn min(self, other: impl Into<Factor>) -> Factor

Returns the minimum of two factors.

pub fn clamp(self, min: impl Into<Factor>, max: impl Into<Factor>) -> Factor

Returns self if min <= self <= max, returns min if self < min or returns max if self > max.

pub fn abs(self) -> Factor

Computes the absolute value of self.

pub fn flip(self) -> Factor

Flip factor, around 0.5,

Returns 1.0 - self.

pub fn pct(self) -> FactorPercent

Factor as percentage.