zng::layout

Struct Rect

pub struct Rect {
    pub origin: Point,
    pub size: Size,
}

2D rect in Length units.

Fields
origin: Point

Top-left origin of the rectangle in length units.

size: Size

Size of the rectangle in length units.

Implementations

impl Rect

pub fn new<O, S>(origin: O, size: S) -> Rect
where O: Into<Point>, S: Into<Size>,

New rectangle defined by an origin point (top-left) and a size, both in any type that converts to Point and Size.

Also see RectFromTuplesBuilder for another way of initializing a rectangle value.

pub fn from_size<S>(size: S) -> Rect
where S: Into<Size>,

New rectangle at Point::zero. The size is in any Length unit.

pub fn zero() -> Rect

New rectangle at Point::zero and Size::zero.

pub fn fill() -> Rect

Rect that fills the available space.

pub fn min(&self) -> Point

Min x and y, this is the origin.

pub fn max(&self) -> Point

Max x and y, this is the sum of origin and size.

pub fn min_x(&self) -> Length

Min x, this is the origin.x.

pub fn min_y(&self) -> Length

Min y, this is the origin.y.

pub fn max_x(&self) -> Length

Max x, this is the origin.x + width.

pub fn max_y(&self) -> Length

Max y, this is the origin.y + height.

pub fn translate(&self, by: impl Into<Vector>) -> Rect

Returns a rectangle of same size that adds the vector to the origin.

pub fn is_default(&self) -> bool

Returns true if all values are Length::Default.

pub fn has_default(&self) -> bool

Returns ´true if any value is [Length::Default`].

pub fn replace_default(&mut self, overwrite: &Rect)

Replaces Length::Default values with overwrite values.