zng::layout
Struct SideOffsets

pub struct SideOffsets {
    pub top: Length,
    pub right: Length,
    pub bottom: Length,
    pub left: Length,
}

2D size offsets in Length units.

This unit defines spacing around all four sides of a box, a widget margin can be defined using a value of this type.

Fields
top: Length

Spacing above, in length units.

right: Length

Spacing to the right, in length units.

bottom: Length

Spacing bellow, in length units.

left: Length

Spacing to the left ,in length units.

Implementations

impl SideOffsets

pub fn new<T, R, B, L>(top: T, right: R, bottom: B, left: L) -> SideOffsets
where T: Into<Length>, R: Into<Length>, B: Into<Length>, L: Into<Length>,

New top, right, bottom left offsets. From any Length type.

pub fn new_vh<TB, LR>(top_bottom: TB, left_right: LR) -> SideOffsets
where TB: Into<Length>, LR: Into<Length>,

Top-bottom and left-right equal. From any Length type.

pub fn new_all<T>(all_sides: T) -> SideOffsets
where T: Into<Length>,

All sides equal. From any Length type.

pub fn zero() -> SideOffsets

All sides zero.

pub fn all_eq(&self) -> bool

If all sides are equal.

pub fn dimensions_eq(&self) -> bool

If top and bottom are equal; and left and right are equal.