zng::layout

Struct Vector

pub struct Vector {
    pub x: Length,
    pub y: Length,
}

2D vector in Length units.

Fields
x: Length

x displacement in length units.

y: Length

y displacement in length units.

Implementations

impl Vector

pub fn new<X, Y>(x: X, y: Y) -> Vector
where X: Into<Length>, Y: Into<Length>,

New x, y from any Length unit.

pub fn splat(xy: impl Into<Length>) -> Vector

New x, y from single value of any Length unit.

pub fn zero() -> Vector

x: Length::zero, y: Length::zero.

pub fn one() -> Vector

(1, 1).

pub fn one_px() -> Vector

(1.px(), 1.px()).

pub fn yx(self) -> Vector

Swap x and y.

pub fn as_tuple(self) -> (Length, Length)

Returns (x, y).

pub fn as_array(self) -> [Length; 2]

Returns [x, y].

pub fn abs(&self) -> Vector

Returns a vector that computes the absolute layout vector of self.

pub fn is_default(&self) -> bool

Returns true if all values are Length::Default.

pub fn replace_default(&mut self, overwrite: &Vector)

Replaces Length::Default values with overwrite values.

pub fn as_point(self) -> Point

Cast to Point.

pub fn as_size(self) -> Size

Cast to Size.

pub fn into_transform(self) -> Transform

Create a translate transform from self.