zng::layout

Type Alias DipSize

pub type DipSize = Size2D<Dip, Dip>;

A size in device pixels.

Aliased Type

#[repr(C)]
pub struct DipSize {
    pub width: Dip,
    pub height: Dip,
}

Fields
width: Dip

The extent of the element in the U units along the x axis (usually horizontal).

height: Dip

The extent of the element in the U units along the y axis (usually vertical).