zng::markdown
Function image_resolver

pub fn image_resolver(
    child: impl UiNode,
    resolver: impl IntoVar<ImageResolver>,
) -> impl UiNode

P Markdown image resolver.

This can be used to override image source resolution, by default the image URL or URI is passed as parsed to the image_fn.

Note that image downloads are blocked by default, you can enable this by using the image::img_limits property.

Sets the IMAGE_RESOLVER_VAR.