Struct zng::markdown::Markdown

pub struct Markdown(/* private fields */);

W Render markdown styled text.

Implementations

impl Markdown

pub fn image_resolver(&self, resolver: impl IntoVar<ImageResolver>)

P Markdown image resolver.

This can be used to override image source resolution, by default the image URL or URI is passed as parsed to the image_fn.

Note that image downloads are blocked by default, you can enable this by using the image::img_limits property.

Sets the IMAGE_RESOLVER_VAR.

impl Markdown

pub fn link_resolver(&self, resolver: impl IntoVar<LinkResolver>)

P Markdown link resolver.

This can be used to expand or replace links.

Sets the LINK_RESOLVER_VAR.

impl Markdown

pub fn link_scroll_mode(&self, mode: impl IntoVar<ScrollToMode>)

P Scroll-to mode used by anchor links.

impl Markdown

pub fn text_fn(&self, wgt_fn: impl IntoVar<WidgetFn<TextFnArgs>>)

P Widget function that converts TextFnArgs to widgets.

Sets the TEXT_FN_VAR.

impl Markdown

pub fn link_fn(&self, wgt_fn: impl IntoVar<WidgetFn<LinkFnArgs>>)

P Widget function that converts LinkFnArgs to widgets.

Sets the LINK_FN_VAR.

impl Markdown

pub fn code_inline_fn(&self, wgt_fn: impl IntoVar<WidgetFn<CodeInlineFnArgs>>)

P Widget function that converts CodeInlineFnArgs to widgets.

Sets the CODE_INLINE_FN_VAR.

impl Markdown

pub fn code_block_fn(&self, wgt_fn: impl IntoVar<WidgetFn<CodeBlockFnArgs>>)

P Widget function that converts CodeBlockFnArgs to widgets.

Sets the CODE_BLOCK_FN_VAR.

impl Markdown

pub fn paragraph_fn(&self, wgt_fn: impl IntoVar<WidgetFn<ParagraphFnArgs>>)

P Widget function that converts ParagraphFnArgs to widgets.

Sets the PARAGRAPH_FN_VAR.

impl Markdown

pub fn heading_fn(&self, wgt_fn: impl IntoVar<WidgetFn<HeadingFnArgs>>)

P Widget function that converts HeadingFnArgs to widgets.

Sets the HEADING_FN_VAR.

impl Markdown

pub fn list_fn(&self, wgt_fn: impl IntoVar<WidgetFn<ListFnArgs>>)

P Widget function that converts ListFnArgs to widgets.

Sets the LIST_FN_VAR.

impl Markdown

pub fn def_list_fn(&self, wgt_fn: impl IntoVar<WidgetFn<DefListArgs>>)

P Widget function that converts DefListArgs to widgets.

Sets the DEF_LIST_FN_VAR.

impl Markdown

pub fn def_list_item_title_fn( &self, wgt_fn: impl IntoVar<WidgetFn<DefListItemTitleArgs>>, )

P Widget function that converts DefListItemTitleArgs to widgets.

Sets the DEF_LIST_ITEM_TITLE_FN_VAR.

impl Markdown

pub fn def_list_item_definition_fn( &self, wgt_fn: impl IntoVar<WidgetFn<DefListItemDefinitionArgs>>, )

P Widget function that converts DefListItemDefinitionArgs to widgets.

Sets the DEF_LIST_ITEM_DEFINITION_FN_VAR.

impl Markdown

pub fn list_item_bullet_fn( &self, wgt_fn: impl IntoVar<WidgetFn<ListItemBulletFnArgs>>, )

P Widget function that converts ListItemBulletFnArgs to widgets.

Sets the LIST_ITEM_BULLET_FN_VAR.

impl Markdown

pub fn list_item_fn(&self, wgt_fn: impl IntoVar<WidgetFn<ListItemFnArgs>>)

P Widget function that converts ListItemFnArgs to widgets.

Sets the LIST_ITEM_FN_VAR.

impl Markdown

pub fn image_fn(&self, wgt_fn: impl IntoVar<WidgetFn<ImageFnArgs>>)

P Widget function that converts ImageFnArgs to widgets.

Sets the IMAGE_FN_VAR.

impl Markdown

pub fn rule_fn(&self, wgt_fn: impl IntoVar<WidgetFn<RuleFnArgs>>)

P Widget function that converts RuleFnArgs to widgets.

Sets the RULE_FN_VAR.

impl Markdown

pub fn block_quote_fn(&self, wgt_fn: impl IntoVar<WidgetFn<BlockQuoteFnArgs>>)

P Widget function that converts BlockQuoteFnArgs to widgets.

Sets the BLOCK_QUOTE_FN_VAR.

impl Markdown

pub fn footnote_ref_fn(&self, wgt_fn: impl IntoVar<WidgetFn<FootnoteRefFnArgs>>)

P Widget function that converts FootnoteRefFnArgs to widgets.

Sets the FOOTNOTE_REF_FN_VAR.

impl Markdown

pub fn footnote_def_fn(&self, wgt_fn: impl IntoVar<WidgetFn<FootnoteDefFnArgs>>)

P Widget function that converts FootnoteDefFnArgs to widgets.

Sets the FOOTNOTE_DEF_FN_VAR.

impl Markdown

pub fn table_fn(&self, wgt_fn: impl IntoVar<WidgetFn<TableFnArgs>>)

P Widget function that converts TableFnArgs to widgets.

Sets the TABLE_FN_VAR.

impl Markdown

pub fn panel_fn(&self, wgt_fn: impl IntoVar<WidgetFn<PanelFnArgs>>)

P Widget function that converts PanelFnArgs to a widget.

This generates the panel that contains all markdown blocks, it is the child of the Markdown! widget.

Sets the PANEL_FN_VAR.

impl Markdown

pub fn widget_new() -> Markdown

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Markdown

pub fn txt(&self, txt: impl IntoVar<Txt>)

P Markdown text.