zng::menu
Function shortcut_spacing

pub fn shortcut_spacing(
    child: impl UiNode,
    spacing: impl IntoVar<Length>,
) -> impl UiNode

P Minimum space between a menu item child and the shortcut_txt child.

The spacing is applied only if the shortcut child is set to a non-zero sized widget and there is no other wider menu item.

This property sets the SHORTCUT_SPACING_VAR.