zng::mouse

Function on_disabled_mouse_hovered

pub fn on_disabled_mouse_hovered(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<MouseHoverArgs>,
) -> UiNode

P Mouse entered or left the widget and descendant widgets area, the widget is disabled and cursor capture allows it.

Preview

You can preview this event using on_pre_disabled_mouse_hovered. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.