zng::mouse

Module raw_device_events

Raw mouse hardware events, received independent of what window is under the pointer.

You must enable device events in the app to receive this events.

Structs

ButtonArgs

Arguments for the BUTTON_EVENT.

MouseMotionArgs

Arguments for MOUSE_MOTION_EVENT.

MouseWheelArgs

Arguments for MOUSE_WHEEL_EVENT.

Statics

BUTTON_EVENT

Button press/release from a device, probably a mouse.

MOUSE_MOTION_EVENT

Mouse device unfiltered move delta.

MOUSE_WHEEL_EVENT

Mouse device unfiltered wheel motion delta.