zng::mouse

Struct MouseClickArgs

pub struct MouseClickArgs {
    pub timestamp: DInstant,
    pub window_id: WindowId,
    pub device_id: InputDeviceId,
    pub button: MouseButton,
    pub position: Point2D<Dip, Dip>,
    pub modifiers: ModifiersState,
    pub click_count: NonZero<u32>,
    pub is_repeat: bool,
    pub hits: HitTestInfo,
    pub target: InteractionPath,
    /* private fields */
}

MOUSE_CLICK_EVENT arguments.

Fields
timestamp: DInstant

Instant the event happened.

window_id: WindowId

Id of window that received the event.

device_id: InputDeviceId

Id of device that generated the event.

button: MouseButton

Which mouse button generated the event.

position: Point2D<Dip, Dip>

Position of the mouse in the coordinates of target.

modifiers: ModifiersState

What modifier keys where pressed when this event happened.

click_count: NonZero<u32>

Count of clicks within the double-click interval. Number 1 is single click, 2 is double click, etc.

Auto repeated clicks also increment this count.

is_repeat: bool

If the event was generated by holding the button pressed over a widget with ClickMode::repeat.

hits: HitTestInfo

Hit-test result for the mouse point in the window, at the moment the click event was generated.

target: InteractionPath

Full path to the widget that got clicked.

A widget is clicked if the mouse down and mouse up happen in sequence in the same widget. Subsequent clicks (double, triple) happen on mouse down.

If a mouse down happen in a child widget and the pointer is dragged to a larger parent widget and then let go (mouse up), the click target is the parent widget.

Multi-clicks (click_count > 1) only happen to the same target.

Implementations

impl MouseClickArgs

pub fn new( timestamp: impl Into<DInstant>, propagation_handle: EventPropagationHandle, window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, button: impl Into<MouseButton>, position: impl Into<Point2D<Dip, Dip>>, modifiers: impl Into<ModifiersState>, click_count: impl Into<NonZero<u32>>, is_repeat: impl Into<bool>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, ) -> MouseClickArgs

New args from values that convert into the argument types.

pub fn now( window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, button: impl Into<MouseButton>, position: impl Into<Point2D<Dip, Dip>>, modifiers: impl Into<ModifiersState>, click_count: impl Into<NonZero<u32>>, is_repeat: impl Into<bool>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, ) -> MouseClickArgs

Arguments for event that happened now (INSTANT.now).

impl MouseClickArgs

pub fn is_primary(&self) -> bool

If the button is the primary.

pub fn is_context(&self) -> bool

If the button is the context (right).

pub fn is_single(&self) -> bool

If the click_count is 1.

pub fn is_double(&self) -> bool

If the click_count is 2.

pub fn is_triple(&self) -> bool

If the click_count is 3.

pub fn position_wgt(&self) -> Option<Point2D<Px, Px>>

Gets position in the widget inner bounds.