zng::render

Struct FrameUpdate

pub struct FrameUpdate { /* private fields */ }

A frame quick update.

A frame update causes a frame render without needing to fully rebuild the display list. It is a more performant but also more limited way of generating a frame.

Any FrameValueKey used in the creation of the frame can be used for updating the frame.

Implementations

impl FrameUpdate

pub fn new( render_update_widgets: Arc<RenderUpdates>, frame_id: FrameId, root_id: WidgetId, root_bounds: WidgetBoundsInfo, clear_color: Rgba, ) -> FrameUpdate

New frame update builder.

render_update_widgets - External update requests.

frame_id - Id of the new frame.

root_id - Id of the window root widget.

root_bounds - Bounds info of the window root widget.

clear_color - The current clear color.

pub fn frame_id(&self) -> FrameId

Id of the new frame.

pub fn is_outer(&self) -> bool

Returns true if the widget inner transform update is still being build.

This is true when inside an update_widget call but false when inside an update_inner call.

pub fn transform(&self) -> &PxTransform

Current transform.

pub fn set_clear_color(&mut self, color: Rgba)

Change the color used to clear the pixel buffer when redrawing the frame.

pub fn auto_hit_test(&self) -> bool

Returns true if all transform updates are also applied to hit-test transforms.

pub fn with_auto_hit_test( &mut self, auto_hit_test: bool, render_update: impl FnOnce(&mut FrameUpdate), )

Runs render_update with auto_hit_test set to a value for the duration of the render call.

pub fn is_visible(&self) -> bool

Returns true if view updates are actually collected, if false only transforms and hit-test are updated.

pub fn hidden(&mut self, update: impl FnOnce(&mut FrameUpdate))

Calls update with is_visible set to false.

Nodes that set the visibility to Hidden must render using the FrameBuilder::hide method and update using this method.

pub fn update_transform( &mut self, new_value: FrameValueUpdate<PxTransform>, hit_test: bool, )

Update a transform value that does not potentially affect widget bounds.

Use with_transform to update transforms that affect widget bounds.

If hit_test is true the hit-test transform is also updated.

pub fn update_transform_opt( &mut self, new_value: Option<FrameValueUpdate<PxTransform>>, hit_test: bool, )

Update a transform value, if there is one.

pub fn with_transform( &mut self, new_value: FrameValueUpdate<PxTransform>, hit_test: bool, render_update: impl FnOnce(&mut FrameUpdate), )

Update a transform that potentially affects widget bounds.

The transform is updated to include this space for the call to the render_update closure. The closure must call render update on child nodes.

If hit_test is true the hit-test transform is also updated.

pub fn with_transform_opt( &mut self, new_value: Option<FrameValueUpdate<PxTransform>>, hit_test: bool, render_update: impl FnOnce(&mut FrameUpdate), )

Update a transform that potentially affects widget bounds, if there is one.

The render_update is always called.

pub fn with_child( &mut self, offset: Vector2D<Px, Px>, render_update: impl FnOnce(&mut FrameUpdate), )

Calls render_update with an offset that affects the first inner child inner bounds.

Nodes that used FrameBuilder::push_child during render must use this method to update the value.

pub fn with_transform_value( &mut self, value: &PxTransform, render_update: impl FnOnce(&mut FrameUpdate), )

Calls render_update while the transform is updated to include the value space.

This is useful for cases where the inner transforms are affected by a value that is only rendered, never updated.

pub fn with_inner_transform( &mut self, transform: &PxTransform, render_update: impl FnOnce(&mut FrameUpdate), )

Update the transform applied after the inner bounds translate.

This is only valid if is_outer.

pub fn can_reuse_widget(&self) -> bool

If widget update can be skipped by setting reuse in update_widget.

pub fn with_no_reuse(&mut self, render_update: impl FnOnce(&mut FrameUpdate))

Calls render_update with can_reuse_widget set to false.

pub fn update_widget(&mut self, render_update: impl FnOnce(&mut FrameUpdate))

Update the widget’s outer transform.

If render-update was not requested for the widget and can_reuse_widget only update outer/inner transforms of descendants. If the widget is reused the render_update is not called.

pub fn reuse_widget(&mut self)

Update the info transforms of the widget and descendants.

Widgets that did not request render-update can use this method to update only the outer and inner transforms of itself and descendants as those values are global and the parent widget may have changed.

pub fn update_inner( &mut self, layout_translation_key: FrameValueKey<PxTransform>, layout_translation_animating: bool, render_update: impl FnOnce(&mut FrameUpdate), )

Update the widget’s inner transform.

The layout_translation_animating affects some webrender caches, see FrameBuilder::push_inner for details.

pub fn update_f32(&mut self, new_value: FrameValueUpdate<f32>)

Update a float value.

pub fn update_f32_opt(&mut self, new_value: Option<FrameValueUpdate<f32>>)

Update a float value, if there is one.

pub fn update_color(&mut self, new_value: FrameValueUpdate<Rgba>)

Update a color value.

See FrameBuilder::push_color for details.

pub fn update_color_opt(&mut self, new_value: Option<FrameValueUpdate<Rgba>>)

Update a color value, if there is one.

pub fn update_extension_raw( &mut self, extension_id: ApiExtensionId, extension_payload: ApiExtensionPayload, )

Update a custom extension value with custom encoding.

pub fn update_extension<T>(&mut self, extension_id: ApiExtensionId, payload: &T)
where T: Serialize,

Update a custom extension value.

pub fn parallel_split(&self) -> ParallelBuilder<FrameUpdate>

Create an update builder that can be send to a parallel task and must be folded back into this builder.

This should be called just before the call to update_widget, an error is traced if called inside a widget outer bounds.

pub fn parallel_fold(&mut self, split: ParallelBuilder<FrameUpdate>)

Collect updates from split into self.

pub fn with_nested_window( &mut self, render_update_widgets: Arc<RenderUpdates>, root_id: WidgetId, root_bounds: WidgetBoundsInfo, update: impl FnOnce(&mut FrameUpdate), )

Calls update to render update a separate nested window on this frame.

pub fn render_update_widgets(&self) -> &Arc<RenderUpdates>

External render update requests for this frame.

pub fn finalize(self, info_tree: &WidgetInfoTree) -> BuiltFrameUpdate

Finalize the update.

Returns the property updates and the new clear color if any was set.