Module zng::rule_line::hr

Horizontal rule line widget and properties.

Structs

Hr

W Draws an horizontal RuleLine!.

Functions

color

P Sets the COLOR_VAR that affects all horizontal rules inside the widget.

line_style

P Sets the LINE_STYLE_VAR that affects all horizontal rules inside the widget.

margin

P Sets the MARGIN_VAR that affects all horizontal rules inside the widget.

stroke_thickness

P Sets the STROKE_THICKNESS_VAR that affects all horizontal rules inside the widget.