zng::scroll
Module cmd

Scroll commands.

Structs

ScrollRequest

Parameters for the scroll and page commands.

ScrollToRequest

Parameters for the SCROLL_TO_CMD.

Enums

ScrollToMode

Defines how much the SCROLL_TO_CMD will scroll to showcase the target widget.

ScrollToTarget

Target for the SCROLL_TO_CMD.

Statics

PAGE_DOWN_CMD

Represents the page down by one v_page_unit action.

PAGE_LEFT_CMD

Represents the page left by one h_page_unit action.

PAGE_RIGHT_CMD

Represents the page right by one h_page_unit action.

PAGE_UP_CMD

Represents the page up by one v_page_unit action.

SCROLL_DOWN_CMD

Represents the scroll down by one v_line_unit action.

SCROLL_LEFT_CMD

Represents the scroll left by one h_line_unit action.

SCROLL_RIGHT_CMD

Represents the scroll right by one h_line_unit action.

SCROLL_TO_BOTTOM_CMD

Represents the scroll to bottom action.

SCROLL_TO_CMD

Represents the action of scrolling until a child widget is fully visible, the command can also adjust the zoom scale.

SCROLL_TO_LEFTMOST_CMD

Represents the scroll to leftmost action.

SCROLL_TO_RIGHTMOST_CMD

Represents the scroll to rightmost action.

SCROLL_TO_TOP_CMD

Represents the scroll to top action.

SCROLL_UP_CMD

Represents the scroll up by one v_line_unit action.

ZOOM_IN_CMD

Represents the zoom in action.

ZOOM_OUT_CMD

Represents the zoom out action.

ZOOM_RESET_CMD

Represents the reset zoom action.

ZOOM_TO_FIT_CMD

Represents the zoom to fit action.

Functions

scroll_to

Scroll all parent is_scroll widgets of target so that it becomes visible.

scroll_to_zoom

Like scroll_to, but also adjusts the zoom scale.