zng::scroll

Function clip_to_viewport

pub fn clip_to_viewport(wgt: &mut WidgetBuilding, clip: impl IntoVar<bool>)

B Clip content to only be visible within the viewport, not under scrollbars.

Disabled by default.

Mixin Property

This property is a mixin, it modifies the widget builder during build, it does not produce a standalone UI node.