zng::scroll

Function scroll_to_focused_mode

pub fn scroll_to_focused_mode(
    child: impl IntoUiNode,
    mode: impl IntoVar<Option<ScrollToMode>>,
) -> UiNode

P Scroll-to mode used by scroll widgets when scrolling to make the focused child visible.

Default is minimal 0dip on all sides, set to None to disable.

Note that SCROLL_TO_CMD requests have priority over scroll-to focused if both requests happen in the same event cycle.

This property sets the SCROLL_TO_FOCUSED_MODE_VAR.