zng::scroll
Function zoom_wheel_origin

pub fn zoom_wheel_origin(
    child: impl UiNode,
    origin: impl IntoVar<Point>,
) -> impl UiNode

P Center point of zoom scaling done using the mouse scroll wheel.

Relative values are resolved in the viewport space. The scroll offsets so that the point in the viewport and content stays as close as possible after the scale change.

The default value (Length::Default) is the cursor position.

This property sets the ZOOM_WHEEL_ORIGIN_VAR