Struct zng::scroll::ScrollInfo

pub struct ScrollInfo(/* private fields */);

Shared reference to the viewport bounds of a scroll.

Implementations

impl ScrollInfo

pub fn viewport(&self) -> Rect<Px, Px>

Gets the viewport bounds in the window space.

pub fn viewport_size(&self) -> Size2D<Px, Px>

Gets the layout size of the viewport.

pub fn viewport_transform(&self) -> PxTransform

Gets the render transform of the viewport.

pub fn joiner_size(&self) -> Size2D<Px, Px>

Gets the layout size of both scroll-bars.

Joiner here is the corner joiner visual, it is sized by the width of the vertical bar and height of the horizontal bar.

pub fn content(&self) -> Rect<Px, Px>

Latest content offset and size.

This is the content bounds, scaled and in the viewport space.