zng::setup

Module page

Common setup wizard pages.

The types in this module are builders for zng::wizard::Page instances that can be used with SetupWizard!.

Full API

See zng_wgt_setup::page for the full API.

Structs

EulaPage

End-user agreement page.

FailedPage

Setup operation failed page.

FailedTaskInfoFnArgs

Arguments for generating a task status UI.

FinishAction

Represents an action that FinishPage runs once the wizard finishes.

FinishPage

Setup operation finish without errors page.

InstallDirPage

Page for reviewing and changing the install directory.

StatusPage

Setup operation status page.

StatusTaskInfoFnArgs

Arguments for generating a task status UI.

TaskInfo

Represents display info about a setup task component of a setup operation.

WelcomePage

Page that shows a markdown message that introduces the purpose of the wizard.

Enums

EulaTxt

Supported formats for EulaPage::license.