zng::setup

Module task

Common setup tasks and custom task API.

Setup tasks can be instantiated with InstallConfig::push, usually after configuration is collected using a SetupWizard! or CLI.

Full API

See zng_ext_setup::task for the full API.

Structs

CopyCurrentExeConfig

Config for CopyCurrentExe.

CreateShortcutConfig

Config for CreateShortcut.

ExtractTarConfig

Config for ExtractTar

Enums

CopyCurrentExe

Setup task that copies std::env::current_exe to a location.

CreateShortcut

Setup task that creates a shortcut to an executable file.

ExtractTar

Setup task that extracts TAR container to a new or existing directory on install and removes these files on uninstall.

SetupTaskError

Represents a SetupTask step error.

Traits

SetupTask

Represents an install and uninstall task implementation.