Function zng::stack::stack_nodes

pub fn stack_nodes(nodes: impl UiNodeList) -> impl UiNode

Basic z-stack node.

Creates a node that updates and layouts the nodes in the logical order they appear in the list and renders them one on top of the other from back(0) to front(len-1). The layout size is the largest item width and height, the parent constraints are used for the layout of each item.

This is the most simple z-stack implementation possible, it is a building block useful for quickly declaring overlaying effects composed of multiple nodes, it does not do any alignment layout or z-sorting render.