zng::state_map

Trait StateValue

pub trait StateValue:
    Any
    + Send
    + Sync { }

Represents a type that can be a StateId value.

Trait Alias

This trait is used like a type alias for traits and is already implemented for all types it applies to.