zng::task::channel

Function ipc_unbounded

pub fn ipc_unbounded<T>() -> Result<(IpcSender<T>, IpcReceiver<T>), Error>
where
    T: IpcValue,

Create an unbounded IPC channel.

Note that the channel endpoints can also be send over IPC, the first channel is setup by Worker. You can also use the NamedIpcReceiver or NamedIpcSender to create the first channel with a custom process.

Note that the channel is only IPC if build with "ipc" crate feature, otherwise it will falls back to channel::unbounded.