zng::task::channel

Function with_ipc_serialization

pub fn with_ipc_serialization<R>(serialize: impl FnOnce() -> R) -> R

Enables special serialization of memory mapped files for the serialize call.

IPC channels like IpcSender serialize messages inside this context to support IpcBytes fast memory map sharing across processes.

You can use the is_ipc_serialization to check if inside context.