zng::task::channel

Struct IpcBytesWriterBlocking

pub struct IpcBytesWriterBlocking { /* private fields */ }

Represents a blocking IpcBytes writer.

Use IpcBytes::new_writer_blocking to start writing.

Implementations

impl IpcBytesWriterBlocking

pub fn finish(self) -> Result<IpcBytes, Error>

Finish writing and move data to a shareable IpcBytes.

pub fn finish_mut(self) -> Result<IpcBytesMut, Error>

Mode data to an exclusive mutable IpcBytes that can be further modified, but not resized.