zng::task

Function poll_respond

pub fn poll_respond<R, F>(
    task: impl IntoFuture<IntoFuture = F>,
) -> ResponseVar<R>
where
    R: VarValue,
    F: Future<Output = R> + Send + 'static,

Polls the task once immediately on the calling thread, if the task is ready returns the response already set, if the task is pending continues execution like respond.