Struct zng::task::http::Body

pub struct Body(/* private fields */);

HTTP request body.

Use TryBody to convert types to body.

Implementations

impl Body

pub fn empty() -> Body

Create a new empty body.

An empty body represents the absence of a body, which is semantically different than the presence of a body of zero length.

pub fn from_bytes_static(bytes: impl AsRef<[u8]> + 'static) -> Body

Create a new body from a potentially static byte buffer.

The body will have a known length equal to the number of bytes given.

This will try to prevent a copy if the type passed in can be re-used, otherwise the buffer will be copied first. This method guarantees to not require a copy for the following types:

pub fn from_reader(read: impl AsyncRead + Send + Sync + 'static) -> Body

Create a streaming body of unknown length.

pub fn from_reader_sized( read: impl AsyncRead + Send + Sync + 'static, size: u64 ) -> Body

Create a streaming body of with known length.

pub fn is_empty(&self) -> bool

Report if this body is empty.

This is not necessarily the same as checking for zero length, since HTTP message bodies are optional, there is a semantic difference between the absence of a body and the presence of a zero-length body. This method will only return true for the former.

pub fn len(&self) -> Option<u64>

Get the size of the body, if known.

pub fn reset(&mut self) -> bool

If this body is repeatable, reset the body stream back to the start of the content.

Returns false if the body cannot be reset.

pub async fn bytes(&mut self) -> Result<Vec<u8>, Error>

Read the body as raw bytes.

pub async fn text_utf8(&mut self) -> Result<Txt, Box<dyn Error>>

Read the body and try to convert to UTF-8.

Consider using Response::text, it uses the header encoding information if available.

pub async fn read(&mut self, buf: &mut [u8]) -> Result<usize, Error>

Read some bytes from the body, returns how many bytes where read.

pub async fn read_exact(&mut self, buf: &mut [u8]) -> Result<(), Error>

Read the from the body to exactly fill the buffer.