zng::task::http

Struct Request

#[non_exhaustive]
pub struct Request {

    pub uri: Uri,
    pub method: Method,
    pub headers: HeaderMap,
    pub timeout: Duration,
    pub connect_timeout: Duration,
    pub low_speed_timeout: (Duration, ByteLength),
    pub redirect_limit: u16,
    pub auto_decompress: bool,
    pub max_upload_speed: ByteLength,
    pub max_download_speed: ByteLength,
    pub require_length: bool,
    pub max_length: ByteLength,
    pub cache: CacheMode,
    pub cookies: bool,
    pub metrics: bool,
    pub body: IpcBytes,

}

HTTP request.

Use send to send a request.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

uri: Uri

The URI.

method: Method

The HTTP method.

headers: HeaderMap

Header values.

Is empty by default.

timeout: Duration

Maximum amount of time that a complete request/response cycle is allowed to take before being aborted. This includes DNS resolution, connecting to the server, writing the request, and reading the response.

Note that this includes the response read operation, so if you get a response but don’t read-it within this timeout you will get a TimedOut IO error.

By default no timeout is used, Duration::MAX.

connect_timeout: Duration

Maximum amount of time to await for establishing connections to a host.

Is 90 seconds by default.

low_speed_timeout: (Duration, ByteLength)

Maximum amount of time allowed when transfer speed is under the given speed in bytes per second.

By default not timeout is used, (Duration::MAX, 0).

redirect_limit: u16

Maximum redirects to follow.

When redirecting the Referer header is updated automatically.

Is 20 by default.

auto_decompress: bool

If should auto decompress received data.

If enabled the “Accept-Encoding” will also be set automatically, if it was not set on the header.

This is enabled by default.

max_upload_speed: ByteLength

Maximum upload speed in bytes per second.

No maximum by default, ByteLength::MAX.

max_download_speed: ByteLength

Maximum download speed in bytes per second.

No maximum by default, ByteLength::MAX.

require_length: bool

If the Content-Length header must be present in the response.

By default this is not required.

max_length: ByteLength

Set the maximum response content length allowed.

If the Content-Length is present on the response and it exceeds this limit an error is returned immediately, otherwise if require_length is not enabled an error will be returned only when the downloaded body length exceeds the limit.

By default no limit is set, ByteLength::MAX.

cache: CacheMode

Response cache mode.

Is CacheMode::Default by default.

cookies: bool

If cookies should be send and stored.

When enabled the http_cache is used to retrieve and store cookies.

Is not enabled by default.

metrics: bool

If transfer metrics should be measured.

When enabled you can get the information using the Response::metrics method.

This is enabled by default.

body: IpcBytes

Request body content.

Is empty by default.

Implementations

impl Request

pub fn new(method: Method, uri: Uri) -> Request

Starts building a request.

Examples

use zng_task::http;

let request = http::Request::new(http::Method::PUT, "https://httpbin.org/put".try_into()?);

pub fn get<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a GET request.

Examples

use zng_task::http;

let get = http::Request::get("https://httpbin.org/get")?;

pub fn put<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a PUT request.

Examples

use zng_task::http;

let put = http::Request::put("https://httpbin.org/put")?.header("accept", "application/json")?;

pub fn post<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a POST request.

Examples

use zng_task::http;

let post = http::Request::post("https://httpbin.org/post")?.header("accept", "application/json")?;

pub fn delete<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a DELETE request.

Examples

use zng_task::http;

let delete = http::Request::delete("https://httpbin.org/delete")?.header("accept", "application/json")?;

pub fn patch<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a PATCH request.

Examples

use zng_task::http;

let patch = http::Request::patch("https://httpbin.org/patch")?.header("accept", "application/json")?;

pub fn head<U>(uri: U) -> Result<Request, <U as TryInto<Uri>>::Error>
where U: TryInto<Uri>,

Starts building a HEAD request.

Examples

use zng_task::http;

let head = http::Request::head("https://httpbin.org")?;

pub fn header<K, V>( self, name: K, value: V, ) -> Result<Request, Box<dyn Error + Send + Sync>>
where K: TryInto<HeaderName>, V: TryInto<HeaderValue>, Box<dyn Error + Send + Sync>: From<<K as TryInto<HeaderName>>::Error> + From<<V as TryInto<HeaderValue>>::Error>,

Appends a header to headers to this request.

pub fn timeout(self, timeout: Duration) -> Request

Set the timeout.

pub fn connect_timeout(self, timeout: Duration) -> Request

Set the connect_timeout.

pub fn low_speed_timeout( self, timeout: Duration, bytes_per_sec: ByteLength, ) -> Request

Set the low_speed_timeout.

pub fn redirect_limit(self, count: u16) -> Request

Set the redirect_limit.

pub fn auto_decompress(self, enabled: bool) -> Request

Set the auto_decompress.

pub fn require_length(self, enabled: bool) -> Request

Set require_length.

pub fn max_length(self, max: ByteLength) -> Request

Set max_length.

pub fn max_upload_speed(self, bytes_per_sec: ByteLength) -> Request

Set the max_upload_speed.

pub fn max_download_speed(self, bytes_per_sec: ByteLength) -> Request

Set the max_download_speed.

pub fn cookies(self, enable: bool) -> Request

Set the cookies.

pub fn metrics(self, enabled: bool) -> Request

Set the metrics.

pub fn body(self, body: IpcBytes) -> Request

Set the body.

pub fn body_text( self, body: &str, ) -> Result<Request, Box<dyn Error + Send + Sync>>

Set the body to a plain text UTF-8 payload. Also sets the Content-Type header if it is not set.

pub fn body_json<T>( self, body: &T, ) -> Result<Request, Box<dyn Error + Send + Sync>>
where T: Serialize,

Set the body to a JSON payload. Also sets the Content-Type header if it is not set.