zng::task

Module io

IO tasks.

Most of the types in this module are re-exported from futures_lite::io.

Re-exports

pub use io::AsyncBufRead;

pub use io::AsyncBufReadExt;

pub use io::AsyncRead;

pub use io::AsyncReadExt;

pub use io::AsyncSeek;

pub use io::AsyncSeekExt;

pub use io::AsyncWrite;

pub use io::AsyncWriteExt;

pub use io::BoxedReader;

pub use io::BoxedWriter;

pub use io::BufReader;

pub use io::BufWriter;

pub use io::Cursor;

pub use io::ReadHalf;

pub use io::WriteHalf;

pub use io::copy;

pub use io::empty;

pub use io::repeat;

pub use io::sink;

pub use io::split;

pub use io::Unblock;

Structs

CloneableError

Represents the cloneable parts of an Error.

McBufReader

Multiple consumer buffered read.

Measure

Measure read/write task.

Metrics

Information about the state of an async IO task.

ReadLimited

Represents a stream reader that generates an error if the source stream exceeds a limit.

Statics

METRICS_ID

Metrics in a Progress::with_meta metadata.

Traits

McBufErrorExt

Extension methods for std::io::Error to be used with errors returned by McBufReader.