Struct zng::task::io::AsyncAsSync

pub struct AsyncAsSync<'r, 'ctx, T> {
    pub context: &'r mut Context<'ctx>,
    pub inner: T,
}

A wrapper around a type that implements AsyncRead or AsyncWrite that converts Pending polls to WouldBlock errors.

This wrapper can be used as a compatibility layer between AsyncRead and Read, for types that take Read as a parameter.

Examples

use std::io::Read;
use std::task::{Poll, Context};

fn poll_for_io(cx: &mut Context<'_>) -> Poll<usize> {
    // Assume we have a library that's built around `Read` and `Write` traits.
    use cooltls::Session;

    // We want to use it with our writer that implements `AsyncWrite`.
    let writer = Stream::new();

    // First, we wrap our `Writer` with `AsyncAsSync` to convert `Pending` polls to `WouldBlock`.
    use futures_lite::io::AsyncAsSync;
    let writer = AsyncAsSync::new(cx, writer);

    // Now, we can use it with `cooltls`.
    let mut session = Session::new(writer);

    // Match on the result of `read()` and translate it to poll.
    match session.read(&mut [0; 1024]) {
        Ok(n) => Poll::Ready(n),
        Err(err) if err.kind() == std::io::ErrorKind::WouldBlock => Poll::Pending,
        Err(err) => panic!("unexpected error: {}", err),
    }
}

// Usually, poll-based functions are best wrapped using `poll_fn`.
use futures_lite::future::poll_fn;
poll_fn(|cx| poll_for_io(cx)).await;

Fields
context: &'r mut Context<'ctx>

The context we are using to poll the future.

inner: T

The actual reader/writer we are wrapping.

Implementations

impl<'r, 'ctx, T> AsyncAsSync<'r, 'ctx, T>

pub fn new(context: &'r mut Context<'ctx>, inner: T) -> AsyncAsSync<'r, 'ctx, T>

Wraps an I/O handle implementing AsyncRead or AsyncWrite traits.

Examples

use futures_lite::io::AsyncAsSync;
use std::task::Context;
use waker_fn::waker_fn;

let reader: &[u8] = b"hello";
let waker = waker_fn(|| {});
let mut context = Context::from_waker(&waker);

let async_reader = AsyncAsSync::new(&mut context, reader);

pub fn close(&mut self) -> Result<(), Error>
where T: AsyncWrite + Unpin,

Attempt to shutdown the I/O handle.

Examples

use futures_lite::io::AsyncAsSync;
use std::task::Context;
use waker_fn::waker_fn;

let reader: Vec<u8> = b"hello".to_vec();
let waker = waker_fn(|| {});
let mut context = Context::from_waker(&waker);

let mut async_reader = AsyncAsSync::new(&mut context, reader);
async_reader.close().unwrap();

pub fn poll_with<R>( &mut self, f: impl FnOnce(Pin<&mut T>, &mut Context<'_>) -> Poll<Result<R, Error>> ) -> Result<R, Error>
where T: Unpin,

Poll this AsyncAsSync for some function.

Examples

use futures_lite::io::{AsyncAsSync, AsyncRead};
use std::task::Context;
use waker_fn::waker_fn;

let reader: &[u8] = b"hello";
let waker = waker_fn(|| {});
let mut context = Context::from_waker(&waker);

let mut async_reader = AsyncAsSync::new(&mut context, reader);
let r = async_reader.poll_with(|io, cx| io.poll_read(cx, &mut [0; 1024]));
assert_eq!(r.unwrap(), 5);