Struct zng::task::io::ReadLimited

pub struct ReadLimited<S, L> { /* private fields */ }

Represents a future that generates an error if an AsyncRead exceeds a limit.

Implementations

impl<S, L> ReadLimited<S, L>
where S: AsyncRead, L: Fn() -> Error,

pub fn new(source: S, limit: ByteLength, on_limit: L) -> ReadLimited<S, L>

Construct a limited reader.

The on_limit closure is called if the limit is reached.