Trait zng::task::io::McBufErrorExt

pub trait McBufErrorExt {
    // Required method
    fn is_only_lazy_left(&self) -> bool;
}

Extension methods for std::io::Error to be used with errors returned by McBufReader.

Required Methods

fn is_only_lazy_left(&self) -> bool

Returns true if this error represents the condition where there are only McBufReader::is_lazy readers left, the buffer is drained and the inner reader is not EOF.

You can recover from this error by turning the reader non-lazy using McBufReader::set_lazy.