zng::task

Module process

Async process API and worker.

This module reexports the async-process for convenience.

Modules

unix

Unix-specific extensions.

worker

Async worker process tasks.

Structs

Child

A spawned child process.

ChildStderr

A handle to a child process’s standard error (stderr).

ChildStdin

A handle to a child process’s standard input (stdin).

ChildStdout

A handle to a child process’s standard output (stdout).

Command

A builder for spawning processes.

ExitStatus

Describes the result of a process after it has terminated.

Output

The output of a finished process.

Stdio

Describes what to do with a standard I/O stream for a child process when passed to the stdin, stdout, and stderr methods of Command.

Functions

driver

Runs the driver for the asynchronous processes.