zng::text::cmd

Struct TextEditOp

pub struct TextEditOp(/* private fields */);

Represents a text edit operation that can be send to an editable text using EDIT_CMD.

Implementations

impl TextEditOp

pub fn new<D>( data: D, op: impl FnMut(&mut D, UndoFullOp<'_>) + Send + 'static, ) -> TextEditOp
where D: Send + Any + 'static,

New text edit operation.

The editable text widget that handles EDIT_CMD will call op during event handling in the node::resolve_text context meaning the TEXT.resolved and TEXT.resolve_caret service is available in op. The text is edited by modifying ResolvedText::txt. The text widget will detect changes to the caret and react s accordingly (updating caret position and animation), the caret index is also snapped to the nearest grapheme start.

The op arguments are a custom data D and what UndoFullOp to run, all text edit operations must be undoable, first UndoOp::Redo is called to โdoโ, then undo and redo again if the user requests undo & redo. The text variable is always read-write when op is called, more than one op can be called before the text variable updates, and ResolvedText::pending_edit is always false.

pub fn insert(insert: impl Into<Txt>) -> TextEditOp

Insert operation.

The insert text is inserted at the current caret index or at 0, or replaces the current selection, after insert the caret is positioned after the inserted text.

pub fn backspace() -> TextEditOp

Remove one backspace range ending at the caret index, or removes the selection.

See SegmentedText::backspace_range for more details about what is removed.

pub fn backspace_word() -> TextEditOp

Remove one backspace word range ending at the caret index, or removes the selection.

See SegmentedText::backspace_word_range for more details about what is removed.

pub fn delete() -> TextEditOp

Remove one delete range starting at the caret index, or removes the selection.

See SegmentedText::delete_range for more details about what is removed.

pub fn delete_word() -> TextEditOp

Remove one delete word range starting at the caret index, or removes the selection.

See SegmentedText::delete_word_range for more details about what is removed.

pub fn clear() -> TextEditOp

Remove all the text.

pub fn replace( select_before: Range<usize>, insert: impl Into<Txt>, select_after: Range<usize>, ) -> TextEditOp

Replace operation.

The select_before is removed, and insert inserted at the select_before.start, after insertion the select_after is applied, you can use an empty insert to just remove.

All indexes are snapped to the nearest grapheme, you can use empty ranges to just position the caret.

pub fn apply_transforms() -> TextEditOp

Applies TEXT_TRANSFORM_VAR and WHITE_SPACE_VAR to the text.