Function zng::text::font_family

pub fn font_family(
    child: impl UiNode,
    names: impl IntoVar<FontNames>,
) -> impl UiNode

P Font family name or list of names for texts in this widget or descendants.

All fonts in the list are resolved according to the font_style, font_weight and font_stretch config. During text shaping the first font on the list is preferred, but if the font does not cover a character or word, that character or word to the second font in the list and so on.

Sets the FONT_FAMILY_VAR.