zng::text
Function justify

pub fn justify(
    child: impl UiNode,
    mode: impl IntoVar<Option<Justify>>,
) -> impl UiNode

P Config the automatic spacing inserted between words and letters when text is aligned to fill.

Text alignment can be set to Align::FILL, if this config is set to Some(mode) when that happens the text layout will automatically insert spaces to try and fill the text block. When justify is not enabled, that is set to None, fill alignment is the same as Align::START.

Sets the JUSTIFY_VAR.