zng::text

Function letter_spacing

pub fn letter_spacing(
    child: impl UiNode,
    extra: impl IntoVar<Length>,
) -> impl UiNode

P Extra spacing added in between text letters. If not set inherits the letter_spacing from the parent widget.

Letter spacing is computed using the font data, this unit represents extra space added to the computed spacing.

A “letter” is a character glyph cluster, e.g.: a, â, 1, -, 漢.

The Default value signals that letter spacing can be tweaked when text justification is enabled, all other values disable automatic adjustments for justification inside words.

Relative values are computed from the length of the space ' ' character.

This property sets the LETTER_SPACING_VAR context var that affects all inner texts.