zng::text

Function on_change_stop

pub fn on_change_stop(
    child: impl UiNode,
    handler: impl WidgetHandler<ChangeStopArgs>,
) -> impl UiNode

P Called after the text changed and interaction has stopped.

The handler will be called after change and change_stop_delay elapses, or the widget loses focus, or the Key::Enter is pressed and accepts_enter is false.