zng::text

Function txt_align

pub fn txt_align(child: impl UiNode, mode: impl IntoVar<Align>) -> impl UiNode

P Alignment of text inside available space.

Horizontal alignment is applied for each line independently, vertical alignment is applied for the entire text block together.

Note that the Text! widget only implements this for text inside each instance in isolation, multiple text instances in an inline row will not all align together by the Text! layout implementation alone.

Sets the TEXT_ALIGN_VAR.

See also txt_overflow_align, used when the text overflows.