zng::text

Function txt_overflow_align

pub fn txt_overflow_align(
    child: impl IntoUiNode,
    mode: impl IntoVar<Align>,
) -> UiNode

P Alignment of text inside available space when the text overflows.

Note that the Text! widget only implements this for text inside each instance in isolation, multiple text instances in an inline row will not all align together. Also note that txt_overflow truncation only applies to the end of the text after it is aligned, so unless this is Align::TOP_START (default) the start of the text maybe still be clipped after truncation.

Sets the TEXT_OVERFLOW_ALIGN_VAR.