zng::text
Struct CaretStatus

pub struct CaretStatus { /* private fields */ }

Display info of edit caret position.

Implementations

impl CaretStatus

pub fn none() -> CaretStatus

Status for text without caret.

pub fn new(index: usize, text: &SegmentedText) -> CaretStatus

New position from char index and text.

Panics

Panics if index is greater then text length.

pub fn index(&self) -> Option<usize>

Char index on the text string, starts a 0, can be the length of the text.

pub fn line(&self) -> Option<NonZero<u32>>

Display line, starts at 1.

Note that this does not count soft line breaks (wrapped lines), this is the actual text line.

pub fn column(&self) -> Option<NonZero<u32>>

Display column, starts at 1.

This is the char count from the start of the text line to the index.