zng::text_input

Function field_style_fn

pub fn field_style_fn(
    child: impl IntoUiNode,
    style_fn: impl IntoVar<StyleFn>,
) -> UiNode

P Extends or replaces the FieldStyle!.

Properties and when conditions set here are applied to widgets using the style.

Note that style instances extend the contextual style by default, you can set replace on a style to true to fully replace.