Struct zng::timer::TimerArgs

pub struct TimerArgs {
    pub timestamp: DInstant,
    pub deadline: Deadline,
    /* private fields */
}

Arguments for an on_interval handler.

Note the timer can be stopped using the handlers unsubscribe, and once handlers stop the timer automatically.

The field values are about the specific call to handler that received the args, the methods on the other hand are connected with the timer by a weak reference and always show the up-to-date state of the timer. For synchronous handlers this does not matter, but for async handlers this means that the values can be different after each .await. This can be useful to for example, disable the timer until the async task finishes but it can also be surprising.

Fields
timestamp: DInstant

When the handler was called.

deadline: Deadline

Expected deadline, is less-or-equal to the timestamp.

Implementations

impl TimerArgs

pub fn interval(&self) -> Duration

The timer interval. Enabled handlers are called every time this interval elapses.

pub fn set_interval(&self, new_interval: Duration)

Set the interval.

Note that this method does not awake the app, so if this is called from outside the app thread it will only apply on the next app update.

pub fn is_paused(&self) -> bool

If the timer is not ticking, but can be started again.

pub fn is_playing(&self) -> bool

If the timer is ticking.

pub fn pause(&self)

Disable the timer, this causes the timer to stop ticking until play is called.

pub fn play(&self, reset: bool)

Enable the timer, this causes it to start ticking again.

If reset is true the last timestamp is set to now.

pub fn count(&self) -> usize

Count incremented by one every time the timer elapses.

pub fn set_count(&self, count: usize)

Resets the count.

pub fn last_timestamp(&self) -> DInstant

The timestamp of the last update. This can be different from timestamp after the first .await in async handlers of if called from a different thread.

pub fn next_deadline(&self) -> Deadline

The next timer deadline.

This is last_timestamp plus interval.

pub fn is_stopped(&self) -> bool

If the timer was stopped while the handler was running after it started handling.

Note the timer can be stopped from the inside of the handler using the handlers unsubscribe, and once handlers stop the timer automatically.

Outside of the handler the TimerHandle can be used to stop the timer at any time, even from another thread.