zng::toggle
Function selector

pub fn selector(
    child: impl UiNode,
    selector: impl IntoValue<Selector>,
) -> impl UiNode

P Sets the contextual selector that all inner widgets will target from the value property.

All value properties declared in widgets inside child will use the Selector to manipulate the selection.

Selection in a context can be blocked by setting the selector to Selector::nil(), this is also the default selector so the value property only works if a contextual selector is present.

This property sets the SELECTOR context and handles cmd::SelectOp requests. It also sets the widget access role to AccessRole::RadioGroup.