Struct zng::toggle::ComboStyle

pub struct ComboStyle(/* private fields */);

W Popup toggle style.

Style a Toggle! widget to look like the popup toggle button in a combo-box.

Implementations

impl ComboStyle

pub fn widget_new() -> ComboStyle

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl ComboStyle

pub fn combo_spacing(&self, spacing: impl IntoVar<Length>)

P Spacing between the arrow symbol and the content.