Trait zng::toggle::SelectorImpl

pub trait SelectorImpl: Send + 'static {
    // Required methods
    fn subscribe(&self);
    fn select(
        &mut self,
        value: Box<dyn AnyVarValue>
    ) -> Result<(), SelectorError>;
    fn deselect(
        &mut self,
        value: &(dyn AnyVarValue + 'static)
    ) -> Result<(), SelectorError>;
    fn is_selected(&self, value: &(dyn AnyVarValue + 'static)) -> bool;
}

Represents a Selector implementation.

Required Methods

fn subscribe(&self)

Add the selector subscriptions in the WIDGET.

fn select(&mut self, value: Box<dyn AnyVarValue>) -> Result<(), SelectorError>

Insert the value in the selection, returns Ok(()) if the value was inserted or was already selected.

fn deselect( &mut self, value: &(dyn AnyVarValue + 'static) ) -> Result<(), SelectorError>

Remove the value from the selection, returns Ok(()) if the value was removed or was not selected.

fn is_selected(&self, value: &(dyn AnyVarValue + 'static)) -> bool

Returns true if the value is selected.