zng::touch

Function is_touch_active

pub fn is_touch_active(
    child: impl IntoUiNode,
    state: impl IntoVar<bool>,
) -> UiNode

P If an unhandled touch tap has happened on the widget within a time duration defined by contextual touch_active_config.

This property is the touch equivalent to is_mouse_active.