zng::touch

Function on_touch_end

pub fn on_touch_end(
    child: impl UiNode,
    handler: impl WidgetHandler<TouchInputArgs>,
) -> impl UiNode

P Touch contact ended over the widget, it is enabled and cursor capture allows it.

Preview

You can preview this event using on_pre_touch_end. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.