zng::touch

Struct TouchTransformArgs

pub struct TouchTransformArgs {

    pub timestamp: DInstant,
    pub window_id: WindowId,
    pub device_id: InputDeviceId,
    pub first_info: TouchTransformInfo,
    pub latest_info: TouchTransformInfo,
    pub velocity: [Vector2D<Px, Px>; 2],
    pub scale_factor: Factor,
    pub phase: TouchPhase,
    pub hits: HitTestInfo,
    pub target: InteractionPath,
    pub capture: Option<CaptureInfo>,
    pub modifiers: ModifiersState,
    pub propagation: EventPropagationHandle,

}

Arguments for TOUCH_TRANSFORM_EVENT.

Fields
timestamp: DInstant

Instant the event happened.

window_id: WindowId

Id of window that received the touch events.

device_id: InputDeviceId

Id of the device that generated the touch events.

first_info: TouchTransformInfo

Info collected when the second touch point started.

latest_info: TouchTransformInfo

Latest update of the two points.

velocity: [Vector2D<Px, Px>; 2]

Velocity of the latest_info touch points.

scale_factor: Factor

Scale factor used in the computed pixel values.

This is the window’s scale factor when the first touch started.

phase: TouchPhase

Gesture phase.

hits: HitTestInfo

Hit-test result for the center point between the first position of the two touches in the window when the gesture started.

target: InteractionPath

Full path to the top-most hit in hits.

capture: Option<CaptureInfo>

Current pointer capture.

modifiers: ModifiersState

What modifier keys where pressed when this event happened.

propagation: EventPropagationHandle

Propagation handle associated with this event instance.

Cloned arguments share the same handle, some arguments may also share the handle of another event if they share the same cause.

Implementations

impl TouchTransformArgs

pub fn new( timestamp: impl Into<DInstant>, propagation: EventPropagationHandle, window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, first_info: impl Into<TouchTransformInfo>, latest_info: impl Into<TouchTransformInfo>, velocity: impl Into<[Vector2D<Px, Px>; 2]>, scale_factor: impl Into<Factor>, phase: impl Into<TouchPhase>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, capture: impl Into<Option<CaptureInfo>>, modifiers: impl Into<ModifiersState>, ) -> TouchTransformArgs

New args from values that convert into the argument types.

pub fn now( window_id: impl Into<WindowId>, device_id: impl Into<InputDeviceId>, first_info: impl Into<TouchTransformInfo>, latest_info: impl Into<TouchTransformInfo>, velocity: impl Into<[Vector2D<Px, Px>; 2]>, scale_factor: impl Into<Factor>, phase: impl Into<TouchPhase>, hits: impl Into<HitTestInfo>, target: impl Into<InteractionPath>, capture: impl Into<Option<CaptureInfo>>, modifiers: impl Into<ModifiersState>, ) -> TouchTransformArgs

Arguments for event that happened now (INSTANT.now).

impl TouchTransformArgs

pub fn capture_allows(&self, wgt: (WindowId, WidgetId)) -> bool

If capture is None or allows the wgt to receive this event.

pub fn local_info(&self, wgt: (WindowId, WidgetId)) -> [TouchTransformInfo; 2]

Gets the first_info and latest_info in the wgt inner bounds space.

pub fn translation(&self) -> Vector2D<f32, Px>

Computes the translation to transform from first_info to latest_info.

pub fn translation_x(&self) -> f32

Computes the translation-x to transform from first_info to latest_info.

pub fn translation_y(&self) -> f32

Computes the translation-y to transform from first_info to latest_info.

pub fn rotation(&self) -> AngleRadian

Computes the rotation to transform from first_info to latest_info.

pub fn scale(&self) -> Factor

Computes the scale to transform from first_info to latest_info.

pub fn scale_x(&self) -> Factor

Computes the scale-y to transform from first_info to latest_info.

pub fn scale_y(&self) -> Factor

Computes the scale-y to transform from first_info to latest_info.

pub fn transform(&self, mode: TouchTransformMode) -> PxTransform

Computes the transform from first_info to latest_info.

pub fn local_transform( &self, mode: TouchTransformMode, wgt: (WindowId, WidgetId), ) -> PxTransform

Computes the transform between the local_info values, rotates and scales around the latest center.

pub fn translation_velocity(&self) -> Vector2D<Px, Px>

Average velocity.

pub fn translation_inertia_x(&self, deceleration: Dip) -> (Px, Duration)

Compute the final offset and duration for a fling animation that simulates inertia movement from the translation_velocity().x and deceleration. Returns 0 if velocity less than min_fling_velocity.

Deceleration is in dip/s, a good value is 1000. The recommended animation easing function is |t| easing::ease_out(easing::quad, t).

pub fn translation_inertia_y(&self, deceleration: Dip) -> (Px, Duration)

Compute the final offset and duration for a fling animation that simulates inertia movement from the translation_velocity().y and deceleration. Returns 0 if velocity less than min_fling_velocity.

Deceleration is in dip/s, a good value is 1000. The recommended animation easing function is |t| easing::ease_out(easing::quad, t).

pub fn is_start(&self) -> bool

If the phase is start.

Note that the latest_info may already be different from first_info if the gesture detector awaited to disambiguate before starting the gesture.

pub fn is_end(&self) -> bool

If the phase is end.

Any transform already applied must be committed.

pub fn is_cancel(&self) -> bool

If the phase is cancel.

Any transform already applied must be undone.