zng::undo

Trait UndoInfo

pub trait UndoInfo:
    Send
    + Sync
    + Any {
    // Required method
    fn description(&self) -> Txt;

    // Provided methods
    fn meta(&self) -> StateMapRef<'_, UNDO> { ... }
    fn into_dyn(self) -> Arc<dyn UndoInfo>
       where Self: Sized { ... }
}

Metadata info about an action registered for undo action.

Required Methods

fn description(&self) -> Txt

Short display description of the action that will be undone/redone.

Provided Methods

fn meta(&self) -> StateMapRef<'_, UNDO>

Any extra metadata associated with the item. This can be a thumbnail of an image edit action for example, or an icon.

Is empty by default.

fn into_dyn(self) -> Arc<dyn UndoInfo>
where Self: Sized,

Into Arc<dyn UndoInfo> without double wrapping.