Struct zng::update::ContextUpdates

pub struct ContextUpdates {
    pub events: Vec<EventUpdate>,
    pub update: bool,
    pub info: bool,
    pub layout: bool,
    pub render: bool,
    pub update_widgets: WidgetUpdates,
    pub info_widgets: InfoUpdates,
    pub layout_widgets: LayoutUpdates,
    pub render_widgets: RenderUpdates,
    pub render_update_widgets: RenderUpdates,
}

Updates that must be reacted by an app owner.

This type is public only for testing, it is the return type for test methods of WINDOW.

Fields
events: Vec<EventUpdate>

Events to notify.

When this is not empty update is true.

update: bool

Update requested.

When this is true, update_widgets may contain widgets, if not then only app extensions must update.

info: bool

Info rebuild requested.

When this is true, info_widgets may contain widgets, if not then only app extensions must update.

layout: bool

Layout requested.

When this is true, layout_widgets may contain widgets, if not then only app extensions must update.

render: bool

Render requested.

When this is true, render_widgets or render_update_widgets may contain widgets, if not then only app extensions must update.

update_widgets: WidgetUpdates

Update targets.

When this is not empty update is true.

info_widgets: InfoUpdates

Info rebuild targets.

When this is not empty info is true.

layout_widgets: LayoutUpdates

Layout targets.

When this is not empty layout is true.

render_widgets: RenderUpdates

Full render targets.

When this is not empty render is true.

render_update_widgets: RenderUpdates

Render update targets.

When this is not empty render is true.

Implementations

impl ContextUpdates

pub fn has_updates(&self) -> bool

If has events, update, layout or render was requested.