zng::var::animation::easing

Function step_floor

pub fn step_floor(steps: u32, time: EasingTime) -> Factor

Jumps to the final value by a number of steps.

Waits until first step to output the first step value.