Trait zng::var::animation::Transitionable

pub trait Transitionable: VarValue {
    // Required method
    fn lerp(self, to: &Self, step: Factor) -> Self;
}

Represents a type that can be animated between two values.

This trait is auto-implemented for all Copy types that can add, subtract and multiply by Factor, Clone only types must implement this trait manually.

Required Methods

fn lerp(self, to: &Self, step: Factor) -> Self

Sample the linear interpolation from self -> to by step.

Object Safety

This trait is not object safe.