zng::var
Function getter_var

pub fn getter_var<T>() -> ArcVar<T>
where
    T: VarValue + Default,

Variable for getter properties (get_*, actual_*).

Getter variables are inited with a default value that is overridden by the property on node init and updated by the property when the internal state they track changes. They are created automatically by the property default when used in when expressions, but can be created manually.