zng::var

Trait IntoValue

pub trait IntoValue<T>: Into<T>
where
    T: VarValue,
{ }

A property value that is not a variable but can be inspected.

Implementing

The trait is only auto-implemented for T: Into<T> + VarValue, unfortunately actual type conversions must be manually implemented, note that the impl_from_and_into_var! macro auto-implements this conversion.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.