zng::var

Trait VarValue

pub trait VarValue:
    Debug
    + Clone
    + PartialEq
    + Any
    + Send
    + Sync { }

Represents a type that can be a Var<T> value.

Trait Alias

This trait is used like a type alias for traits and is already implemented for all types it applies to.

Implementing

Types need to be Debug + Clone + PartialEq + Send + Sync + Any to auto-implement this trait, if you want to place an external type in a variable and it does not implement all the traits you may need to declare a newtype wrapper, if the external type is Debug + Send + Sync + Any at least you can use the ArcEq<T> wrapper to quickly implement Clone + PartialEq, this is particularly useful for error types in ResponseVar<Result<_, E>>.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.