zng::view_process::prebuilt

Function run_same_process

pub fn run_same_process(run_app: impl FnOnce() + Send + 'static) -> !

Call the prebuilt run_same_process.

This function exits the process after run_app returns.

Panics

Panics if it fails to install the prebuilt binary.

Aborts

Kills the process with code 101 if there is a panic generated by the pre-built code or by threads started by the pre-build code.