zng::widget::builder
Struct NestGroup

pub struct NestGroup(/* private fields */);

Property nest position group.

Each group has u16::MAX / 9 in between, custom groups can be created using the +/- operations, SIZE+1 is still outside BORDER, but slightly inside SIZE.

See NestPosition for more details.

Implementations

impl NestGroup

pub const WIDGET: NestGroup = _

Minimal nest position, property is outside even context properties and is only inside the widget node.

This is rarely used, prefer using CONTEXT-n if you must have a property outside the widget context.

pub const CONTEXT: NestGroup = _

Property defines a contextual value or variable.

Usually these properties don’t define behavior, they just configure the widget. A common pattern is defining all widget config as context vars, that are all used by a widget intrinsic node.

These properties are not expected to affect layout or render, if they do some errors may be logged by the default widget base.

pub const EVENT: NestGroup = _

Property defines an event handler, or state monitor, they are placed inside all context properties, so can be configured by context, but are still outside of the layout and render nodes.

Event handlers can be notified before or after the inner child delegation, if handled before the event is said to be preview. Implementers can use this intrinsic feature of the UI tree to interrupt notification for child properties and widgets.

These properties are not expected to affect layout or render, if they do some errors may be logged by the default widget base.

pub const LAYOUT: NestGroup = _

Property defines the position and size of the widget inside the space made available by the parent widget.

These properties must accumulatively affect the measure and layout, they must avoid rendering. The computed layout is usually rendered by the widget as a single transform, the layout properties don’t need to render transforms.

pub const SIZE: NestGroup = _

Property strongly enforces a widget size.

Usually the widget final size is a side-effect of all the layout properties, but some properties may enforce a size, they can use this group to ensure that they are inside the other layout properties.

pub const WIDGET_INNER: NestGroup = _

Minimal widget visual position, any property or node can render, but usually only properties inside this position render. For example, borders will only render correctly inside this nest position.

This is rarely used, prefer using BORDER-n to declare properties that are visually outside the bounds, only use this node for intrinsics that define some inner context or service for the visual properties.

pub const BORDER: NestGroup = _

Property renders a border visual.

Borders are strictly coordinated, see the border module for more details. All nodes of this group may render at will, the renderer is already configured to apply the final layout and size.

pub const FILL: NestGroup = _

Property defines a visual of the widget.

This is the main render group, it usually defines things like a background fill, but it can render over child nodes simply by choosing to render after the render is delegated to the inner child.

pub const CHILD_CONTEXT: NestGroup = _

Property defines contextual value or variable for the inner child or children widgets. Config set here does not affect the widget where it is set, it only affects the descendants.

pub const CHILD_LAYOUT: NestGroup = _

Property defines the layout and size of the child or children widgets. These properties don’t affect the layout of the widget where they are set. Some properties are functionally the same, only changing their effect depending on their group, the margin and padding properties are like this, margin is LAYOUT and padding is CHILD_LAYOUT.

pub const CHILD: NestGroup = _

Maximum nest position, property is inside all others and only wraps the widget child node.

Properties that insert child nodes may use this group, properties that only affect the child layout and want to be inside other child layout should use CHILD_LAYOUT+n instead.

pub const ITEMS: [NestGroup; 11] = _

All groups, from outermost(WIDGET) to innermost(CHILD).

pub fn name(self) -> Txt

Group name.