Struct zng::widget::builder::WidgetBuilder

pub struct WidgetBuilder { /* private fields */ }

Widget instance builder.

Implementations

impl WidgetBuilder

pub fn new(widget: WidgetType) -> WidgetBuilder

New empty default.

pub fn widget_type(&self) -> WidgetType

The widget that started this builder.

pub fn push_property( &mut self, importance: Importance, args: Box<dyn PropertyArgs>, )

Insert/override a property.

You can use the property_args! macro to collect args for a property.

pub fn push_property_positioned( &mut self, importance: Importance, position: NestPosition, args: Box<dyn PropertyArgs>, )

Insert property with custom nest position.

pub fn push_when(&mut self, importance: Importance, when: WhenInfo)

Insert a when block.

pub fn push_unset(&mut self, importance: Importance, property_id: PropertyId)

Insert a name = unset!; property.

pub fn push_property_build_action( &mut self, property_id: PropertyId, action_name: &'static str, importance: Importance, input_actions: Vec<Box<dyn AnyPropertyBuildAction>>, )

Add or override custom builder actions that are called to finalize the inputs for a property.

The importance overrides previous build action of the same name and property. The input_actions vec must contain one action for each property input.

pub fn push_unset_property_build_action( &mut self, property_id: PropertyId, action_name: &'static str, importance: Importance, )

Insert a property build action filter.

pub fn clear_property_build_actions(&mut self)

Remove all registered property build actions.

pub fn push_build_action( &mut self, action: impl FnMut(&mut WidgetBuilding) + Send + 'static, )

Add an action closure that is called every time this builder or a clone of it builds a widget instance.

pub fn clear_build_actions(&mut self)

Remove all registered build actions.

pub fn is_custom_build(&self) -> bool

Returns true if a custom build handler is registered.

pub fn set_custom_build<R>( &mut self, build: impl FnMut(WidgetBuilder) -> R + Send + 'static, )
where R: UiNode,

Set a build closure to run instead of default_build when build is called.

Overrides the previous custom build, if any was set.

pub fn clear_custom_build(&mut self)

Remove the custom build handler, if any was set.

pub fn extend(&mut self, other: WidgetBuilder)

Apply other over self.

All properties, unsets, whens, build actions and custom build of other are inserted in self, override importance rules apply, other items only replace self items if they have the same or greater importance.

Note that properties of the same position index from other are pushed after properties of the same position in self, this means that fill properties of other will render over fill properties of self.

pub fn has_properties(&self) -> bool

If any property is present in the builder.

pub fn has_unsets(&self) -> bool

If any unset filter is present in the builder.

pub fn has_whens(&self) -> bool

If any when block is present in the builder.

pub fn split_off( &mut self, properties: impl IntoIterator<Item = PropertyId>, out: &mut WidgetBuilder, )

Move all properties to a new builder.

The properties are removed from self, any when assign is also moved, properties used in WhenInput that affect the properties are cloned or moved into the new builder.

Note that properties can depend on others in the widget contextually, this is not preserved on split-off.

pub fn build(self) -> Box<dyn UiNodeBoxed>

Instantiate the widget.

If a custom build is set it is run, unless it is already running, otherwise the default_build is called.

pub fn default_build(self) -> Box<dyn UiNodeBoxed>

Instantiate the widget.

Runs all build actions, but ignores custom build.