zng::widget
Function on_interactivity_changed

pub fn on_interactivity_changed(
    child: impl UiNode,
    handler: impl WidgetHandler<InteractivityChangedArgs>,
) -> impl UiNode

P Widget interactivity changed.

Note that there are multiple specific events for interactivity changes, on_enable, on_disable, on_block and on_unblock are some of then.

Note that an event is received when the widget first initializes in the widget info tree, this is because the interactivity changed from None, this initial event can be detected using the is_new method in the args.

Preview

You can preview this event using on_pre_interactivity_changed. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.