zng::widget

Function on_vis_enabled_changed

pub fn on_vis_enabled_changed(
    child: impl IntoUiNode,
    handler: impl WidgetHandler<InteractivityChangedArgs>,
) -> UiNode

P Widget changed to enabled or disabled visuals.

Note that this event tracks the visual enabled status of the widget, not the actual status, the widget may still be blocked, see Interactivity for more details.

Note that an event is received when the widget first initializes in the widget info tree, this is because the interactivity changed from None, this initial event can be detected using the is_new method in the args.

See on_interactivity_changed for a more general interactivity event.

Preview

You can preview this event using on_pre_vis_enabled_changed. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.