zng::widget::info

Struct WidgetInfo

pub struct WidgetInfo { /* private fields */ }

Reference to a widget info in a WidgetInfoTree.

Implementations

impl WidgetInfo

pub fn access(&self) -> Option<WidgetAccessInfo>

Accessibility info, if the widget is accessible.

The widget is accessible only if access_enabled and some accessibility metadata was set on the widget.

pub fn access_children(&self) -> impl Iterator<Item = WidgetAccessInfo>

Descendant branches that have accessibility info.

The iterator enters descendants only until it finds a node that has access info, these nodes are yielded.

pub fn access_parent(&self) -> Option<WidgetAccessInfo>

First ancestor that is accessible.

impl WidgetInfo

pub fn id(&self) -> WidgetId

Widget id.

pub fn path(&self) -> WidgetPath

Full path to this widget.

pub fn trace_path(&self) -> Txt

Path details to help finding the widget during debug.

If the inspector metadata is present the widget type is included.

pub fn trace_id(&self) -> Txt

Detailed id text.

If the inspector metadata is present the widget type is included.

pub fn interaction_path(&self) -> InteractionPath

Full path to this widget with interactivity values.

pub fn new_path(&self, old_path: &WidgetPath) -> Option<WidgetPath>

Gets the path if it is different from old_path.

Only allocates a new path if needed.

Panics

If old_path does not point to the same widget id as self.

pub fn new_interaction_path( &self, old_path: &InteractionPath, ) -> Option<InteractionPath>

Gets the interaction_path if it is different from old_path.

Only allocates a new path if needed.

Panics

If old_path does not point to the same widget id as self.

pub fn z_index(&self) -> Option<(ZIndex, ZIndex)>

Get the z-index of the widget in the latest frame if it was rendered.

Note that widgets can render in the back and front of each descendant, these indexes are the back-most index, the moment the widget starts rendering and the front-most index at the moment the widget and all contents finishes rendering.

This value is updated every render without causing a tree rebuild.

pub fn visibility(&self) -> Visibility

Gets the visibility of the widget or the widget’s descendants in the last rendered frame.

A widget is Visible if it rendered at least one display item, Hidden if it rendered only space and hit-test items, Collapsed if it did not render. All widgets are Visible if no frame was ever rendered.

pub fn interactivity(&self) -> Interactivity

Get or compute the interactivity of the widget.

The interactivity of a widget is the combined result of all interactivity filters applied to it and its ancestors. If a parent is blocked this is blocked, same for disabled.

pub fn bounds_info(&self) -> WidgetBoundsInfo

All the transforms introduced by this widget, starting from the outer info.

This information is up-to-date, it is updated every layout and render without causing a tree rebuild.

pub fn border_info(&self) -> WidgetBorderInfo

Clone a reference to the widget border and corner radius information.

This information is up-to-date, it is updated every layout without causing a tree rebuild.

pub fn perspective(&self) -> Option<(f32, Point2D<Px, Px>)>

Gets the 3D perspective for this widget.

The f32 is a distance from the Z-plane to the viewer, the point is the vanishing center in the parent widget inner bounds.

pub fn transform_style(&self) -> TransformStyle

Gets the transform style for this widget.

Is Flat unless it or the parent widget sets Preserve3D.

pub fn outer_size(&self) -> Size2D<Px, Px>

Size of the widget outer area, not transformed.

Returns an up-to-date size, the size is updated every layout without causing a tree rebuild.

pub fn inner_size(&self) -> Size2D<Px, Px>

Size of the widget inner area, not transformed.

Returns an up-to-date size, the size is updated every layout without causing a tree rebuild.

pub fn inner_border_size(&self) -> Size2D<Px, Px>

Size of the widget child area, not transformed.

Returns an up-to-date size, the size is updated every layout without causing a tree rebuild.

pub fn baseline(&self) -> Px

Gets the baseline offset up from the inner bounds bottom line.

pub fn outer_transform(&self) -> PxTransform

Widget outer transform in window space.

Returns an up-to-date transform, the transform is updated every render or render update without causing a tree rebuild.

pub fn inner_transform(&self) -> PxTransform

Widget inner transform in the window space.

Returns an up-to-date transform, the transform is updated every render or render update without causing a tree rebuild.

pub fn outer_bounds(&self) -> Rect<Px, Px>

Widget outer rectangle in the window space.

Returns an up-to-date rect, the bounds are updated every render or render update without causing a tree rebuild.

pub fn inner_bounds(&self) -> Rect<Px, Px>

Widget inner rectangle in the window space.

Returns an up-to-date rect, the bounds are updated every render or render update without causing a tree rebuild.

pub fn spatial_bounds(&self) -> Box2D<Px, Px>

Compute the bounding box that envelops self and descendants inner bounds.

pub fn center(&self) -> Point2D<Px, Px>

Widget inner bounds center in the window space.

pub fn meta(&self) -> StateMapRef<'_, WidgetInfoMeta>

Custom metadata associated with the widget during info build.

pub fn tree(&self) -> &WidgetInfoTree

Reference the WidgetInfoTree that owns self.

pub fn is_reused(&self) -> bool

If the widget info and all descendants did not change in the last rebuild.

pub fn root(&self) -> WidgetInfo

Reference to the root widget.

pub fn parent(&self) -> Option<WidgetInfo>

Reference to the widget that contains this widget.

Is None only for root.

pub fn prev_sibling(&self) -> Option<WidgetInfo>

Reference to the previous widget within the same parent.

pub fn next_sibling(&self) -> Option<WidgetInfo>

Reference to the next widget within the same parent.

pub fn first_child(&self) -> Option<WidgetInfo>

Reference to the first widget within this widget.

pub fn last_child(&self) -> Option<WidgetInfo>

Reference to the last widget within this widget.

pub fn has_siblings(&self) -> bool

If the parent widget has multiple children.

pub fn has_children(&self) -> bool

If the widget has at least one child.

pub fn siblings(&self) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

All parent children except this widget.

pub fn children(&self) -> Children

Iterator over the direct descendants of the widget.

pub fn children_count(&self) -> usize

Count of children.

pub fn self_and_children(&self) -> Children

Iterator over the widget and the direct descendants of the widget.

pub fn descendants(&self) -> TreeIter

Iterator over all widgets contained by this widget.

pub fn descendants_len(&self) -> usize

Total number of descendants.

pub fn self_and_descendants(&self) -> TreeIter

Iterator over the widget and all widgets contained by it.

pub fn ancestors(&self) -> Ancestors

Iterator over parent -> grandparent -> .. -> root.

pub fn descendants_range(&self) -> WidgetDescendantsRange

Gets a value that can check if widgets are descendant of self in O(1) time.

pub fn cmp_sibling_in( &self, sibling: &WidgetInfo, ancestor: &WidgetInfo, ) -> Option<Ordering>

Compare the position of self and sibling in the ancestor descendants.

pub fn is_ancestor(&self, maybe_descendant: &WidgetInfo) -> bool

If self is an ancestor of maybe_descendant.

pub fn is_descendant(&self, maybe_ancestor: &WidgetInfo) -> bool

If self is inside maybe_ancestor.

pub fn self_and_ancestors(&self) -> Ancestors

Iterator over self -> parent -> grandparent -> .. -> root.

pub fn prev_siblings(&self) -> PrevSiblings

Iterator over all previous widgets within the same parent.

pub fn self_and_prev_siblings(&self) -> PrevSiblings

Iterator over self and all previous widgets within the same parent.

pub fn next_siblings(&self) -> NextSiblings

Iterator over all next widgets within the same parent.

pub fn self_and_next_siblings(&self) -> NextSiblings

Iterator over self and all next widgets within the same parent.

pub fn prev_siblings_in(&self, ancestor: &WidgetInfo) -> RevTreeIter

Iterator over all previous widgets within the same ancestor, including descendants of siblings.

If ancestor is not actually an ancestor iterates to the root.

pub fn self_and_prev_siblings_in(&self, ancestor: &WidgetInfo) -> RevTreeIter

Iterator over self, descendants and all previous widgets within the same ancestor.

If ancestor is not actually an ancestor iterates to the root.

pub fn next_siblings_in(&self, ancestor: &WidgetInfo) -> TreeIter

Iterator over all next widgets within the same ancestor, including descendants of siblings.

If ancestor is not actually an ancestor iterates to the root.

pub fn self_and_next_siblings_in(&self, ancestor: &WidgetInfo) -> TreeIter

Iterator over self, descendants and all next widgets within the same ancestor.

If ancestor is not actually an ancestor iterates to the root.

pub fn orientation_from(&self, origin: Point2D<Px, Px>) -> Option<Orientation2D>

The center orientation in relation to an origin.

Returns None if the origin is the center.

pub fn distance_key(&self, origin: Point2D<Px, Px>) -> DistanceKey

Value that indicates the distance between this widget center and origin.

pub fn rect_distance_key(&self, origin: Point2D<Px, Px>) -> DistanceKey

Value that indicates the distance between the nearest point inside this widgets rectangles and origin.

The widgets rectangles is the inner bounds for block widgets or the row rectangles for inline widgets.

pub fn rect_distance_key_filtered( &self, origin: Point2D<Px, Px>, filter: impl FnMut(Rect<Px, Px>, usize, usize) -> bool, ) -> DistanceKey

Like rect_distance_key, but only consider rectangles approved by filter.

The filter parameters are the rectangle, the inline row index and the total inline rows length. If the widget is not inlined both the index and len are zero.

pub fn depth(&self) -> usize

Count of ancestors.

pub fn shared_ancestor(&self, other: &WidgetInfo) -> Option<WidgetInfo>

First ancestor of self and other.

Returns None if other is not from the same tree.

pub fn is_in_bounds(&self) -> bool

Returns true if this widget’s inner bounds are fully contained by the parent inner bounds.

pub fn out_of_bounds( &self, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Iterator over all descendants with inner bounds not fully contained by their parent inner bounds.

pub fn spatial_iter<F>( &self, filter: F, ) -> impl Iterator<Item = WidgetInfo> + use<F>
where F: Fn(&WidgetInfo) -> bool + Clone,

Iterator over self and descendants, first self, then all in-bounds descendants, then all out-of-bounds descendants.

If the filter returns false the widget and all it’s in-bounds descendants are skipped, otherwise they are yielded. After all in-bounds descendants reachable from self and filtered the iterator changes to each out-of-bounds descendants and their in-bounds descendants that are also filtered.

pub fn inner_contains( &self, point: Point2D<Px, Px>, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Iterator over self and all descendants with inner bounds that contain the point.

pub fn inner_intersects( &self, rect: Rect<Px, Px>, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over self and descendants with inner bounds that intersects the rect.

pub fn inner_contains_rect( &self, rect: Rect<Px, Px>, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over self and descendants with inner bounds that fully envelops the rect.

pub fn inner_contained( &self, rect: Rect<Px, Px>, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over self and descendants with inner bounds that are fully inside the rect.

pub fn center_contained( &self, area: Rect<Px, Px>, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over self and descendants with center point inside the area.

pub fn center_in_distance( &self, origin: Point2D<Px, Px>, max_radius: Px, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over self and descendants with center point within the max_radius of the origin.

pub fn hit_test(&self, point: Point2D<Px, Px>) -> HitTestInfo

Gets all widgets of self and descendants hit by a point, sorted by z-index of the hit, front to back.

pub fn nearest( &self, origin: Point2D<Px, Px>, max_radius: Px, ) -> Option<WidgetInfo>

Find the descendant with center point nearest of origin within the max_radius.

This method is faster than sorting the result of center_in_distance, but is slower if any point in distance is acceptable.

pub fn nearest_filtered( &self, origin: Point2D<Px, Px>, max_radius: Px, filter: impl FnMut(&WidgetInfo) -> bool, ) -> Option<WidgetInfo>

Find the widget, self or descendant, with center point nearest of origin within the max_radius and approved by the filter closure.

pub fn nearest_bounded_filtered( &self, origin: Point2D<Px, Px>, max_radius: Px, bounds: Rect<Px, Px>, filter: impl FnMut(&WidgetInfo) -> bool, ) -> Option<WidgetInfo>

Find the widget, self or descendant, with center point nearest of origin within the max_radius and inside bounds; and approved by the filter closure.

pub fn nearest_rect( &self, origin: Point2D<Px, Px>, max_radius: Px, ) -> Option<WidgetInfo>

Find the descendant that has inner bounds or inline rows nearest to origin and are within max_radius.

The distance is from any given point inside the bounds or inline rows rectangle to the origin. If origin is inside the rectangle the distance is zero. If multiple widgets have the same distance the nearest center point widget is used.

pub fn nearest_rect_filtered( &self, origin: Point2D<Px, Px>, max_radius: Px, filter: impl FnMut(&WidgetInfo, Rect<Px, Px>, usize, usize) -> bool, ) -> Option<WidgetInfo>

Find the descendant that has inner bounds or inline rows nearest to origin and are within max_radius and are approved by the filter closure.

The filter parameters are the widget, the rect, the rect row index and the widget inline rows length. If the widget is not inlined both index and len are zero.

pub fn nearest_rect_bounded_filtered( &self, origin: Point2D<Px, Px>, max_radius: Px, bounds: Rect<Px, Px>, filter: impl FnMut(&WidgetInfo, Rect<Px, Px>, usize, usize) -> bool, ) -> Option<WidgetInfo>

Find the widget, self or descendant, with inner bounds or inline rows nearest of origin within the max_radius and inside bounds; and approved by the filter closure.

The filter parameters are the widget, the rect, the rect row index and the widget inline rows length. If the widget is not inlined both index and len are zero.

pub fn oriented( &self, origin: Point2D<Px, Px>, max_distance: Px, orientation: Orientation2D, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over all widgets, self and descendants, with center in the direction defined by orientation and within max_distance of the origin, widgets are only visited once and the distance is clipped by the spatial_bounds.

Use Px::MAX on the distance to visit all widgets in the direction.

The direction is defined by a 45º frustum cast from the origin, see Orientation2D::point_is for more details.

pub fn oriented_box( &self, origin: Box2D<Px, Px>, max_distance: Px, orientation: Orientation2D, ) -> impl Iterator<Item = WidgetInfo> + 'static + use<>

Spatial iterator over all widgets, self and descendants, with inner_bounds in the direction defined by orientation in relation to origin and with center within max_distance of the origin center. Widgets are only visited once and the distance is clipped by the spatial_bounds.

Use Px::MAX on the distance to visit all widgets in the direction.

The direction is a collision check between inner-bounds and origin, see Orientation2D::box_is for more details.

pub fn nearest_oriented( &self, origin: Point2D<Px, Px>, max_distance: Px, orientation: Orientation2D, ) -> Option<WidgetInfo>

Find the widget with center point nearest of origin within the max_distance and with orientation to origin.

This method is faster than searching the result of oriented.

pub fn nearest_oriented_filtered( &self, origin: Point2D<Px, Px>, max_distance: Px, orientation: Orientation2D, filter: impl FnMut(&WidgetInfo) -> bool, ) -> Option<WidgetInfo>

Find the widget with center point nearest of origin within the max_distance and with orientation to origin, and approved by the filter closure.

This method is faster than searching the result of oriented.

pub fn nearest_box_oriented( &self, origin: Box2D<Px, Px>, max_distance: Px, orientation: Orientation2D, ) -> Option<WidgetInfo>

Find the widget with center point nearest to origin center within the max_distance and with box orientation to origin.

This method is faster than searching the result of oriented_box.

pub fn nearest_box_oriented_filtered( &self, origin: Box2D<Px, Px>, max_distance: Px, orientation: Orientation2D, filter: impl FnMut(&WidgetInfo) -> bool, ) -> Option<WidgetInfo>

Find the widget with center point nearest to origin center within the max_distance and with box orientation to origin, and approved by the filter closure.

This method is faster than searching the result of oriented_box.