Struct zng::widget::info::WidgetInfoBuilder

pub struct WidgetInfoBuilder { /* private fields */ }

Widget info tree builder.

See WidgetInfoTree for more details.

Implementations

impl WidgetInfoBuilder

pub fn access(&mut self) -> Option<WidgetAccessInfoBuilder<'_>>

Accessibility metadata builder.

Only available if accessibility info is required for the window.

impl WidgetInfoBuilder

pub fn new( info_widgets: Arc<InfoUpdates>, window_id: WindowId, access_enabled: AccessEnabled, root_id: WidgetId, root_bounds_info: WidgetBoundsInfo, root_border_info: WidgetBorderInfo, scale_factor: Factor ) -> WidgetInfoBuilder

Starts building a info tree with the root information.

pub fn widget_id(&self) -> WidgetId

Current widget id.

pub fn with_build_meta<R>( &mut self, visitor: impl FnOnce(StateMapMut<'_, WidgetInfoMeta>) -> R ) -> R

Widget info tree build metadata.

This metadata can be modified only by pushed widgets, not by the reused widgets.

pub fn set_build_meta<T>( &mut self, id: impl Into<StateId<T>>, value: impl Into<T> )
where T: StateValue,

Set the info tree build metadata id to value.

pub fn flag_build_meta(&mut self, id: impl Into<StateId<()>>)

Sets the info tree build metadata id without value.

pub fn with_meta<R>( &mut self, visitor: impl FnOnce(StateMapMut<'_, WidgetInfoMeta>) -> R ) -> R

Current widget info metadata.

pub fn set_meta<T>(&mut self, id: impl Into<StateId<T>>, value: impl Into<T>)
where T: StateValue,

Set the widget info metadata id to value.

Returns the previous set value.

pub fn flag_meta(&mut self, id: impl Into<StateId<()>>)

Sets the widget info metadata id without value.

pub fn push_widget(&mut self, f: impl FnOnce(&mut WidgetInfoBuilder))

Calls f to build the context widget info.

Note that f is only called if the widget info cannot be reused.

pub fn push_interactivity(&mut self, interactivity: Interactivity)

Add the interactivity bits to the current widget’s interactivity, it will affect the widget and all descendants.

Also see push_interactivity_filter to affect the interactivity of widgets outside the current one.

pub fn push_interactivity_filter( &mut self, filter: impl Fn(&InteractivityFilterArgs) -> Interactivity + Send + Sync + 'static )

Register a closure that returns the Interactivity allowed for each widget.

Widgets interactivity is computed from all interactivity filters and parents. Interactivity filters are global to the widget tree, and are re-registered for the tree if the current widget is reused.

Note that the filter can make the assumption that parent widgets affect all descendants and if the filter is intended to affect only the current widget and descendants you can use push_interactivity instead.

pub fn with_children_range( &mut self, info: impl FnOnce(&mut WidgetInfoBuilder) ) -> Range<usize>

Calls the info closure and returns the range of children inserted by it.

pub fn parallel_split(&self) -> ParallelBuilder<WidgetInfoBuilder>

Create a new info builder that can be built in parallel and merged back onto this one using parallel_fold.

pub fn parallel_fold(&mut self, split: ParallelBuilder<WidgetInfoBuilder>)

Collect info from split into self.

pub fn finalize( self, previous_tree: Option<WidgetInfoTree>, notify: bool ) -> WidgetInfoTree

Build the info tree.

Also notifies WIDGET_INFO_CHANGED_EVENT and INTERACTIVITY_CHANGED_EVENT if notify is true.