zng::widget::node

Function presenter

pub fn presenter<D>(
    data: impl IntoVar<D>,
    wgt_fn: impl IntoVar<WidgetFn<D>>,
) -> impl UiNode
where
    D: VarValue,

Node that presents data using wgt_fn.

The node’s child is always the result of wgt_fn called for the data value, it is reinited every time either variable changes.

See also presenter_opt for a presenter that is nil with the data is None.

See also the present method that can be called on the data`` variable and [present_data](VarPresentData::present_data) that can be called on the wgt_fn` variable.