zng::widget

Struct BorderSide

#[repr(C)]
pub struct BorderSide {
    pub color: Rgba,
    pub style: BorderStyle,
}

The line style and color for the sides of a widget’s border.

Fields
color: Rgba

Line color.

style: BorderStyle

Line style.

Implementations

impl BorderSide

pub fn new<C, S>(color: C, style: S) -> BorderSide
where C: Into<Rgba>, S: Into<BorderStyle>,

New border side from color and style value.

pub fn solid<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Solid style.

pub fn double<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Double style.

pub fn dotted<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Dotted style.

pub fn dashed<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Dashed style.

pub fn groove<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Groove style.

pub fn ridge<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Ridge style.

pub fn inset<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Inset style.

pub fn outset<C>(color: C) -> BorderSide
where C: Into<Rgba>,

New border side with Outset style.

pub fn hidden() -> BorderSide

New border side with Hidden style and transparent color.