Struct zng::widget::WIDGET

pub struct WIDGET;

Current context widget.

Panics

Most of the methods on this service panic if not called inside a widget context.

Implementations

impl WIDGET

pub fn is_in_widget(&self) -> bool

Returns true if called inside a widget.

pub fn try_id(&self) -> Option<WidgetId>

Get the widget ID, if called inside a widget.

pub fn trace_path(&self) -> Txt

Gets a text with detailed path to the current widget.

This can be used to quickly identify the current widget during debug, the path printout will contain the widget types if the inspector metadata is found for the widget.

This method does not panic if called outside of a widget.

pub fn trace_id(&self) -> Txt

Gets a text with a detailed widget id.

This can be used to quickly identify the current widget during debug, the printout will contain the widget type if the inspector metadata is found for the widget.

This method does not panic if called outside of a widget.

pub fn id(&self) -> WidgetId

Get the widget ID.

pub fn info(&self) -> WidgetInfo

Gets the widget info.

pub fn bounds(&self) -> WidgetBoundsInfo

Widget bounds, updated every layout.

pub fn border(&self) -> WidgetBorderInfo

Widget border, updated every layout.

pub fn parent_id(&self) -> Option<WidgetId>

Gets the parent widget or None if is root.

Panics if not called inside a widget.

pub fn update_op(&self, op: UpdateOp) -> &WIDGET

Schedule an UpdateOp for the current widget.

pub fn update(&self) -> &WIDGET

Schedule an update for the current widget.

After the current update the app-extensions, parent window and widgets will update again.

pub fn update_info(&self) -> &WIDGET

Schedule an info rebuild for the current widget.

After all requested updates apply the parent window and widgets will re-build the info tree.

pub fn layout(&self) -> &WIDGET

Schedule a re-layout for the current widget.

After all requested updates apply the parent window and widgets will re-layout.

pub fn render(&self) -> &WIDGET

Schedule a re-render for the current widget.

After all requested updates and layouts apply the parent window and widgets will re-render.

This also overrides any pending render_update request.

pub fn render_update(&self) -> &WIDGET

Schedule a frame update for the current widget.

After all requested updates and layouts apply the parent window and widgets will update the frame.

This request is supplanted by any render request.

pub fn reinit(&self)

Flags the widget to re-init after the current update returns.

The widget responds to this request differently depending on the node method that calls it:

UiNode::init and UiNode::deinit: Request is ignored, removed.

UiNode::event: If the widget is pending a reinit, it is reinited first, then the event is propagated to child nodes. If a reinit is requested during event handling the widget is reinited immediately after the event handler.

UiNode::update: If the widget is pending a reinit, it is reinited and the update ignored. If a reinit is requested during update the widget is reinited immediately after the update.

Other methods: Reinit request is flagged and an UiNode::update is requested for the widget.

pub fn with_state<R>(&self, f: impl FnOnce(StateMapRef<'_, WIDGET>) -> R) -> R

Calls f with a read lock on the current widget state map.

pub fn with_state_mut<R>( &self, f: impl FnOnce(StateMapMut<'_, WIDGET>) -> R ) -> R

Calls f with a write lock on the current widget state map.

pub fn get_state<T>(&self, id: impl Into<StateId<T>>) -> Option<T>
where T: StateValue + Clone,

Get the widget state id, if it is set.

pub fn req_state<T>(&self, id: impl Into<StateId<T>>) -> T
where T: StateValue + Clone,

Require the widget state id.

Panics if the id is not set.

pub fn set_state<T>( &self, id: impl Into<StateId<T>>, value: impl Into<T> ) -> Option<T>
where T: StateValue,

Set the widget state id to value.

Returns the previous set value.

pub fn flag_state(&self, id: impl Into<StateId<()>>) -> bool

Sets the widget state id without value.

Returns if the state id was already flagged.

pub fn init_state<T>(&self, id: impl Into<StateId<T>>, init: impl FnOnce() -> T)
where T: StateValue,

Calls init and sets id if it is not already set in the widget.

pub fn init_state_default<T>(&self, id: impl Into<StateId<T>>)
where T: StateValue + Default,

Sets the id to the default value if it is not already set.

pub fn contains_state<T>(&self, id: impl Into<StateId<T>>) -> bool
where T: StateValue,

Returns true if the id is set or flagged in the widget.

pub fn sub_var_op(&self, op: UpdateOp, var: &impl AnyVar) -> &WIDGET

Subscribe to receive UpdateOp when the var changes.

pub fn sub_var_op_when<T>( &self, op: UpdateOp, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive UpdateOp when the var changes and predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_var(&self, var: &impl AnyVar) -> &WIDGET

Subscribe to receive updates when the var changes.

pub fn sub_var_when<T>( &self, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive updates when the var changes and the predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_var_info(&self, var: &impl AnyVar) -> &WIDGET

Subscribe to receive info rebuild requests when the var changes.

pub fn sub_var_info_when<T>( &self, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive info rebuild requests when the var changes and the predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_var_layout(&self, var: &impl AnyVar) -> &WIDGET

Subscribe to receive layout requests when the var changes.

pub fn sub_var_layout_when<T>( &self, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive layout requests when the var changes and the predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_var_render(&self, var: &impl AnyVar) -> &WIDGET

Subscribe to receive render requests when the var changes.

pub fn sub_var_render_when<T>( &self, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive render requests when the var changes and the predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_var_render_update(&self, var: &impl AnyVar) -> &WIDGET

Subscribe to receive render update requests when the var changes.

pub fn sub_var_render_update_when<T>( &self, var: &impl Var<T>, predicate: impl Fn(&T) -> bool + Send + Sync + 'static ) -> &WIDGET
where T: VarValue,

Subscribe to receive render update requests when the var changes and the predicate approves the new value.

Note that the predicate does not run in the widget context, it runs on the app context.

pub fn sub_event<A>(&self, event: &Event<A>) -> &WIDGET
where A: EventArgs,

Subscribe to receive events from event when the event targets this widget.

pub fn push_event_handle(&self, handle: EventHandle)

Hold the event handle until the widget is deinited.

pub fn push_event_handles(&self, handles: EventHandles)

Hold the event handles until the widget is deinited.

pub fn push_var_handle(&self, handle: VarHandle)

Hold the var handle until the widget is deinited.

pub fn push_var_handles(&self, handles: VarHandles)

Hold the var handles until the widget is deinited.

pub fn win_point_to_wgt( &self, point: Point2D<Dip, Dip> ) -> Option<Point2D<Px, Px>>

Transform point in the window space to the widget inner bounds.

pub fn win_to_wgt(&self) -> Option<PxTransform>

Gets the transform from the window space to the widget inner bounds.

pub fn with_handles<R>( &self, handles: &mut WidgetHandlesCtx, f: impl FnOnce() -> R ) -> R

Calls f with an override target for var and event subscription handles.

By default when vars and events are subscribed using the methods of this service the subscriptions live until the widget is deinited. This method intersects these subscriptions, registering then in handles instead.

pub fn with_context<R>( &self, ctx: &mut WidgetCtx, update_mode: WidgetUpdateMode, f: impl FnOnce() -> R ) -> R

Calls f while the widget is set to ctx.

If update_mode is WidgetUpdateMode::Bubble the update flags requested for the ctx after f will be copied to the caller widget context, otherwise they are ignored.

This method can be used to manually define a widget context, note that widgets already define their own context.

pub fn with_no_context<R>(&self, f: impl FnOnce() -> R) -> R

Calls f while no widget is available in the context.