zng::widget
Struct ZIndex

pub struct ZIndex(pub u32);

Position of a widget inside an UiNodeList render operation.

When two widgets have the same index their logical position defines the render order.

Examples

Create a Z-index that causes the widget to render in front of all siblings that don’t set Z-index.

let highlight_z = ZIndex::DEFAULT + 1;

Tuple Fields
0: u32
Implementations

impl ZIndex

pub const BACK: ZIndex = _

Widget is rendered first causing all overlapping siblings to render on top of it.

The value is 0.

pub const DEFAULT: ZIndex = _

Z-index of widgets that don’t set the index.

The value is u32::MAX / 2.

pub const FRONT: ZIndex = _

Widget is rendered after all siblings causing it to render on top.

pub fn saturating_add(self, other: impl Into<ZIndex>) -> ZIndex

Computes other above self, caps at FRONT.

This is the default ZIndex addition, equivalent to self + other.

pub fn saturating_sub(self, other: impl Into<ZIndex>) -> ZIndex

Computes other below self, stops at BACK.

This is the default ZIndex subtraction, equivalent to self - other.