zng::window

Enum WindowIcon

pub enum WindowIcon {
    Default,
    Image(ImageSource),
}

Window icon.

Variants

Default

The operating system’s default icon.

Image(ImageSource)

Image is requested from IMAGES.

Implementations

impl WindowIcon

pub fn render<I, F>(new_icon: F) -> WindowIcon
where I: UiNode, F: Fn() -> I + Send + Sync + 'static,

New window icon from a closure that generates a new icon UiNode for the window.

The closure is called once on init and every time the window icon property changes, the closure runs in a headless window context, it must return a node to be rendered as an icon.

The icon node is deinited and dropped after the first render, you can enable image::render_retain on it to cause the icon to continue rendering on updates.

Examples

WindowIcon::render(
    || Container! {
        // image::render_retain = true;
        size = (36, 36);
        background_gradient = Line::to_bottom_right(), stops![colors::MIDNIGHT_BLUE, 70.pct(), colors::CRIMSON];
        corner_radius = 6;
        font_size = 28;
        font_weight = FontWeight::BOLD;
        child = Text!("A");
    }
)