zng::window

Function on_window_moved

pub fn on_window_moved(
    child: impl IntoUiNode,
    handler: Box<dyn FnMut(&WindowChangedArgs) -> HandlerResult + Send>,
) -> UiNode

P On window position changed.

Preview

You can preview this event using on_pre_window_moved. Otherwise the handler is only called after the widget content has a chance to stop propagation.

Async

You can use async event handlers with this property.