zng::window::inspector

Function show_hit_test

pub fn show_hit_test(
    child: impl IntoUiNode,
    enabled: impl IntoVar<bool>,
) -> UiNode

P Draws the inner bounds that where tested for the mouse point.

Window Only

This property only works if set in a window, if set in another widget it will log an error and not render anything.