Struct zng::window::WINDOW

pub struct WINDOW;

Current context window.

This represents the minimum features required for a window context, see WINDOW_Ext for more features provided by the default window implementation.

Panics

Most of the methods on this service panic if not called inside a window context.

Implementations

impl WINDOW

pub fn is_in_window(&self) -> bool

Returns true if called inside a window.

pub fn try_id(&self) -> Option<WindowId>

Gets the window ID, if called inside a window.

pub fn id(&self) -> WindowId

Gets the window ID.

pub fn mode(&self) -> WindowMode

Gets the window mode.

pub fn info(&self) -> WidgetInfoTree

Gets the window info tree.

Panics if called before the window future yields the window.

pub fn with_state<R>(&self, f: impl FnOnce(StateMapRef<'_, WINDOW>) -> R) -> R

Calls f with a read lock on the current window state map.

pub fn with_state_mut<R>( &self, f: impl FnOnce(StateMapMut<'_, WINDOW>) -> R, ) -> R

Calls f with a write lock on the current window state map.

pub fn get_state<T>(&self, id: impl Into<StateId<T>>) -> Option<T>
where T: StateValue + Clone,

Get the window state id, if it is set.

pub fn req_state<T>(&self, id: impl Into<StateId<T>>) -> T
where T: StateValue + Clone,

Require the window state id.

Panics if the id is not set.

pub fn set_state<T>( &self, id: impl Into<StateId<T>>, value: impl Into<T>, ) -> Option<T>
where T: StateValue,

Set the window state id to value.

Returns the previous set value.

pub fn flag_state(&self, id: impl Into<StateId<()>>) -> bool

Sets the window state id without value.

Returns if the state id was already flagged.

pub fn init_state<T>(&self, id: impl Into<StateId<T>>, init: impl FnOnce() -> T)
where T: StateValue,

Calls init and sets id if the id is not already set in the widget.

pub fn init_state_default<T>(&self, id: impl Into<StateId<T>>)
where T: StateValue + Default,

Sets the id to the default value if it is not already set.

pub fn contains_state<T>(&self, id: impl Into<StateId<T>>) -> bool
where T: StateValue,

Returns true if the id is set or flagged in the window.

pub fn with_context<R>(&self, ctx: &mut WindowCtx, f: impl FnOnce() -> R) -> R

Calls f while the window is set to ctx.

pub fn with_no_context<R>(&self, f: impl FnOnce() -> R) -> R

Calls f while no window is available in the context.

impl WINDOW

Window test helpers.

Panics

Most of the test methods panic if not called inside with_test_context.

pub fn with_test_context<R>( &self, update_mode: WidgetUpdateMode, f: impl FnOnce() -> R, ) -> R

Calls f inside a new headless window and root widget.

pub fn test_window_size(&self) -> Size2D<Px, Px>

Get the test window size.

This size is used by the test_* methods that need a window size.

pub fn set_test_window_size(&self, size: Size2D<Px, Px>)

Set test window size.

pub fn test_init(&self, content: &mut impl UiNode) -> ContextUpdates

Call inside with_test_context to init the content as a child of the test window root.

pub fn test_deinit(&self, content: &mut impl UiNode) -> ContextUpdates

Call inside with_test_context to deinit the content as a child of the test window root.

pub fn test_info(&self, content: &mut impl UiNode) -> ContextUpdates

Call inside with_test_context to rebuild info the content as a child of the test window root.

pub fn test_event( &self, content: &mut impl UiNode, update: &mut EventUpdate, ) -> ContextUpdates

Call inside with_test_context to delivery an event to the content as a child of the test window root.

pub fn test_update( &self, content: &mut impl UiNode, updates: Option<&mut WidgetUpdates>, ) -> ContextUpdates

Call inside with_test_context to update the content as a child of the test window root.

The updates can be set to a custom delivery list, otherwise window root and content widget are flagged for update.

pub fn test_layout( &self, content: &mut impl UiNode, constraints: Option<PxConstraints2d>, ) -> (Size2D<Px, Px>, ContextUpdates)

Call inside with_test_context to layout the content as a child of the test window root.

pub fn test_layout_inline( &self, content: &mut impl UiNode, measure_constraints: (PxConstraints2d, InlineConstraintsMeasure), layout_constraints: (PxConstraints2d, InlineConstraintsLayout), ) -> ((Size2D<Px, Px>, Size2D<Px, Px>), ContextUpdates)

Call inside with_test_context to layout the content as a child of the test window root.

Returns the measure and layout size, and the requested updates.

pub fn test_render( &self, content: &mut impl UiNode, ) -> (BuiltFrame, ContextUpdates)

Call inside with_test_context to render the content as a child of the test window root.

pub fn test_render_update( &self, content: &mut impl UiNode, ) -> (BuiltFrameUpdate, ContextUpdates)

Call inside with_test_context to render_update the content as a child of the test window root.