zng::window
Trait WidgetInfoImeArea

pub trait WidgetInfoImeArea {
    // Required method
    fn ime_area(&self) -> Rect<Px, Px>;
}

IME extension methods for WidgetInfo.

Required Methods

fn ime_area(&self) -> Rect<Px, Px>

IME exclusion area in the window space.

Widgets are IME targets when they are focused and subscribe to IME_EVENT. This value is an area the IME window should avoid covering, by default it is the widget inner-bounds, but the widget can override it using set_ime_area.

This value can change after every render update.