zng::color

Module filter

Color filter types and properties.

Structs

ColorMatrix

Represents a custom color filter.

Filter

A color filter or combination of filters.

Functions

backdrop_blur

P Blur the everything behind the widget.

backdrop_brightness

P Adjust color brightness of everything behind the widget.

backdrop_color_matrix

P Custom backdrop filter.

backdrop_contrast

P Adjust the color contrast of everything behind the widget.

backdrop_filter

P Backdrop filter, or combination of filters.

backdrop_grayscale

P Grayscale tone everything behind the widget.

backdrop_hue_rotate

P Hue shift the colors behind the widget.

backdrop_invert

P Inverts the colors of everything behind the widget.

backdrop_saturate

P Adjust color saturation of everything behind the widget.

backdrop_sepia

P Sepia tone everything behind the widget.

blur

P Blur the widget.

brightness

P Adjust the widget colors brightness.

child_filter

P Color filter, or combination of filters targeting the widget’s descendants and not the widget itself.

child_mix_blend

P Sets how the widget’s child content blends with the widget.

child_opacity

P Opacity/transparency of the widget’s child.

color_matrix

P Custom color filter.

contrast

P Adjust the widget colors contrast.

drop_shadow

P Drop-shadow effect for the widget.

filter

P Color filter, or combination of filters.

grayscale

P Grayscale tone the widget.

hue_rotate

P Hue shift the widget colors.

invert_color

P Inverts the colors of the widget.

mix_blend

P Sets how the widget blends with the parent widget.

opacity

P Opacity/transparency of the widget.

saturate

P Adjust the widget colors saturation.

sepia

P Sepia tone the widget.

Type Aliases

RenderFilter

A computed Filter, ready for Webrender.