zng::color

Module gradient

Color gradient types and nodes.

Macros

stops

Creates a GradientStops containing the arguments.

Structs

ColorStop

A color stop in a gradient.

ConicGradient

Conic gradient.

GradientBuilder

Gradient builder start.

GradientRadius

The radial gradient radius length in both dimensions.

GradientStops

Stops in a gradient.

LinearGradient

Linear gradient.

RadialGradient

Radial gradient.

TiledConicGradient

Tiled conic gradient.

TiledLinearGradient

Tiled linear gradient.

TiledRadialGradient

Tiled radial gradient.

Enums

ExtendMode

Specifies how to draw the gradient outside the first and last stop.

GradientRadiusBase

The radial gradient radius base length.

GradientStop

A stop in a gradient.

LinearGradientAxis

The angle or line that defines a linear gradient.

Functions

conic_gradient

Starts building a conic gradient with the angle and color stops.

gradient

Starts building a gradient with the color stops.

linear_gradient

Starts building a linear gradient with the axis and color stops.

radial_gradient

Starts building a radial gradient with the radius and color stops.

Type Aliases

RenderExtendMode

Gradient extend mode supported by the render.

RenderGradientStop

Computed GradientStop.