Function zng::color::gradient::conic_gradient

pub fn conic_gradient<C, A, S>(
    center: C,
    angle: A,
    stops: S
) -> ConicGradient<<S as IntoVar<GradientStops>>::Var, <C as IntoVar<Point>>::Var, <A as IntoVar<AngleRadian>>::Var, LocalVar<ExtendMode>>
where
    C: IntoVar<Point>,
    A: IntoVar<AngleRadian>,
    S: IntoVar<GradientStops>,

Starts building a conic gradient with the angle and color stops.

Returns a node that is also a builder that can be used to refine the gradient definition.