zng::color::gradient
Function radial_gradient

pub fn radial_gradient<C, R, S>(
    center: C,
    radius: R,
    stops: S,
) -> RadialGradient<<S as IntoVar<GradientStops>>::Var, <C as IntoVar<Point>>::Var, <R as IntoVar<GradientRadius>>::Var, LocalVar<ExtendMode>>
where
    C: IntoVar<Point>,
    R: IntoVar<GradientRadius>,
    S: IntoVar<GradientStops>,

Starts building a radial gradient with the radius and color stops.

Returns a node that is also a builder that can be used to refine the gradient definition.