zng::color::gradient

Type Alias RenderExtendMode

pub type RenderExtendMode = ExtendMode;

Gradient extend mode supported by the render.

Note that ExtendMode::Reflect is not supported directly, you must duplicate and mirror the stops and use the Repeat render mode.

Aliased Type

enum RenderExtendMode {
    Clamp = 0,
    Repeat = 1,
}

Variants

Clamp = 0

Repeat = 1