zng::config

Module settings

Settings are the config the user can directly edit, this module implements a basic settings data model.

Full API

See zng_ext_config::settings for the full settings API.

Modules

editor

Settings editor widget.

Structs

CategoriesBuilder

Setting categories builder.

Category

Settings category.

CategoryBuilder

Category entry builder.

CategoryId

Unique ID of a Category.

SETTINGS

Settings metadata service.

Setting

Setting entry.

SettingBuilder

Setting entry builder.

SettingsBuilder

Settings builder.

Statics

SETTINGS_CMD

Represents the open settings action.

Functions

on_pre_settings

P Preview on_settings command.

on_settings

P On settings command.