zng::config::settings

Module editor

Settings editor widget.

Full API

See zng_wgt_settings for the full settings editor API.

Structs

CategoriesListArgs

Arguments for a widget function that makes a list of category items that can be selected.

CategoryHeaderArgs

Arguments for a widget function that makes a category header in a settings list.

CategoryItemArgs

Arguments for a widget function that makes a category item for a categories list.

SettingArgs

Arguments for a widget function that makes a setting container.

SettingsArgs

Arguments for a widget function that makes a settings for a category list.

SettingsEditor

W Settings editor widget.

Traits

SettingBuilderEditorExt

Extends SettingBuilder to set custom editor metadata.

SettingsCtxExt

Extends SETTINGS to provide contextual information in an editor.

Functions

categories_list_fn

P Widget function that converts CategoriesListArgs to a category list.

category_header_fn

P Widget function that converts CategoryHeaderArgs to a category settings header.

category_item_fn

P Widget function that converts CategoryItemArgs to a category item on a category list.

reset_button

Reset button used by default_setting_fn.

setting_fn

P Widget function that converts SettingArgs to a setting editor entry on a settings list.

settings_fn

P Widget function that converts SettingsArgs to a settings list.