zng::container

Function child_insert

pub fn child_insert(
    child: impl IntoUiNode,
    placement: impl IntoVar<ChildInsert>,
    node: impl IntoUiNode,
) -> UiNode

P Insert node in the placement relative to the widget’s child.

The node is inserted inside the CHILD_LAYOUT scope, meaning inside padding, just like the child. See also child_out_insert for inserting a node outside the child layout.

Spacing between the widget’s child and node can be configured using child_spacing.

A property for each direction is also provided, see child_start, child_end, child_left, child_right, child_top, child_bottom, child_over and child_under.

This property disables inline layout for the widget.