zng::event
Trait CommandInfoExt

pub trait CommandInfoExt {
    // Required methods
    fn info(self) -> Box<dyn VarBoxed<Txt>>;
    fn init_info(self, info: impl Into<Txt>) -> Self;
}

Adds the info command metadata.

Required Methods

fn info(self) -> Box<dyn VarBoxed<Txt>>

Gets a read-write variable that is a short informational string about the command.

fn init_info(self, info: impl Into<Txt>) -> Self

Sets the initial info if it is not set.

Object Safety

This trait is not object safe.