zng::event

Trait EventArgs

pub trait EventArgs: AnyEventArgs + Clone {
    // Provided method
    fn handle<F, R>(&self, handler: F) -> Option<R>
       where F: FnOnce(&Self) -> R { ... }
}

Event<A> arguments.

See AnyEventArgs for the object safe part of event arguments.

Provided Methods

fn handle<F, R>(&self, handler: F) -> Option<R>
where F: FnOnce(&Self) -> R,

Calls handler and stops propagation if propagation is still allowed.

Returns the handler result if it was called.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.