Trait zng::gesture::HeadlessAppGestureExt

pub trait HeadlessAppGestureExt {
    // Required method
    fn press_shortcut(
        &mut self,
        window_id: WindowId,
        shortcut: impl Into<Shortcut>,
    );
}

Extension trait that adds gesture simulation methods to HeadlessApp.

Required Methods

fn press_shortcut(&mut self, window_id: WindowId, shortcut: impl Into<Shortcut>)

Generates key press events to mimic the shortcut and updates.

Object Safety

This trait is not object safe.