zng::gesture

Trait HeadlessAppGestureExt

pub trait HeadlessAppGestureExt {
    // Required method
    fn press_shortcut(
        &mut self,
        window_id: WindowId,
        shortcut: impl Into<Shortcut>,
    );
}

Extension trait that adds gesture simulation methods to HeadlessApp.

Required Methods

fn press_shortcut(&mut self, window_id: WindowId, shortcut: impl Into<Shortcut>)

Generates key press events to mimic the shortcut and updates.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.