zng::grid::cell

Function column

pub fn column(child: impl IntoUiNode, col: impl IntoVar<usize>) -> UiNode

P Cell column index.

If set to usize::MAX the cell is positioned based on the logical index.

Is 0 by default.

This property sets the INFO_ID.

See also the at property to bind both indexes at the same time.