Attribute Macro zng::hot_reload::hot_node

#[hot_node]

Expands an UI node function into a hot reloading one.

See the module level documentation for more details about hot reloading.

Attribute

This attribute has one optional argument, a string literal that uniquely identifies the function among all other hot node functions. The default name is only the function name, so you can use this argument to resolve name conflicts.

Limitations

This attribute only accepts inputs with a single name, no destructuring, and of type that is Clone + Any + Send or the impl generics supported by property. Unlike property this function does not support named generic parameters.

The function output type must be impl UiNode, the attribute will change the internal node type.

Expands an impl into a UiNode impl.

Full Documentation

Read the documentation in the zng::hot_reload::hot_node page.