Module zng::icon

Icon widget and other types.

Icons for this widget are defined in a text font. The Material Design Icons can be embedded using the crate feature "material_icons".

zng = { version = "0.8.2", features = ["material_icons"] }

use zng::{prelude::*, icon};

icon::Icon! {
    ico = icon::material_rounded::ACCESSIBILITY;
    ico_size = 80;
}

You can implement your own icon sets by providing GlyphIcon instances or a type that converts to GlyphIcon, the MaterialIcon type is an example of this. Glyph icons define a font name and a GlyphSource that can be a char or a ligature text.

use zng::{prelude::*, icon, font};

let font = font::CustomFont::from_file(
    "Font Awesome 6 Free-Regular",
    r#"Font Awesome 6 Free-Regular-400.otf"#,
    0,
);
font::FONTS.register(font).wait_into_rsp().await.unwrap();

icon::Icon! {
    ico = icon::GlyphIcon::new("Font Awesome 6 Free-Regular", "address-book").with_ligatures();
    ico_size = 80;
}

The example above loads an icon font and display one of the icons selected using a ligature that matches "address-book".

Full API

See zng_wgt_text::icon for the full widget API.

Modules

material_filled

Filled icons.

material_outlined

Outline icons.

material_rounded

Rounded icons.

material_sharp

Sharp icons.

Structs

GlyphIcon

Represents an icon glyph and font.

Icon

W Render icons defined as glyphs in an icon font.

MaterialFonts

Material fonts.

MaterialIcon

Represents a material font icon.

Enums

GlyphSource

Identifies an icon glyph in the font set.

Traits

CommandIconExt

Adds the icon command metadata.

Functions

ico_color

P Sets the ICON_COLOR_VAR that affects all icons inside the widget.

ico_size

P Sets the ICON_SIZE_VAR that affects all icons inside the widget.