zng::icon

Module material

Material Icons

The Material Design Icons can be embedded using the "material_icons*" crate features.

zng = { version = "0.20.5", features = ["material_icons"] }

Handlers are registered for ICONS that provides the icons, the raw codepoints and glyph icon metadata is available in each font module.

If multiple material icons are enabled they are resolved in this order:

outlined

filled

rounded

sharp

You can disambiguate icons by using a the "material/{set}/{name}" where {set} is one of the items from the list above, and {name} is the icon name.

Full API

See zng_wgt_material_icons for the full API.

Modules

filled

Filled icons.

outlined

Outline icons.

rounded

Rounded icons.

sharp

Sharp icons.