zng::layout

Type Alias PxSize

pub type PxSize = Size2D<Px, Px>;

A size in device pixels.

Aliased Type

#[repr(C)]
pub struct PxSize {
    pub width: Px,
    pub height: Px,
}

Fields
width: Px

The extent of the element in the U units along the x axis (usually horizontal).

height: Px

The extent of the element in the U units along the y axis (usually vertical).