Module zng::markdown

Markdown widget, properties and other types.

This widget displays CommonMark text, without support for HTML code yet .

zng::markdown::Markdown! {
    txt = "# Title\n\n- List item 1.\n- List item 2.";
}

Full API

See zng_wgt_markdown for the full widget API.

Structs

BlockQuoteFnArgs

Arguments for a markdown block quote view.

CodeBlockFnArgs

Arguments for a markdown code block view.

CodeInlineFnArgs

Arguments for a markdown inlined code text view.

FootnoteDefFnArgs

Arguments for a markdown footnote definition view.

FootnoteRefFnArgs

Arguments for a markdown footnote reference view.

HeadingFnArgs

Arguments for a markdown heading view.

ImageFnArgs

Arguments for a markdown image view.

LinkArgs

Arguments for the LINK_EVENT.

LinkFnArgs

Arguments for a markdown inlined link view.

ListFnArgs

Arguments for a markdown list view.

ListItemBulletFnArgs

Arguments for a markdown list item bullet, check mark or number.

ListItemFnArgs

Arguments for a markdown list item view.

Markdown

W Render markdown styled text.

MarkdownStyle

Markdown text run style.

PanelFnArgs

Arguments for a markdown panel.

ParagraphFnArgs

Arguments for a markdown paragraph view.

RuleFnArgs

Arguments for a markdown rule view.

TableCellFnArgs

Arguments for a markdown table cell view.

TableFnArgs

Arguments for a markdown table view.

TextFnArgs

Arguments for a markdown text view.

Enums

HeadingLevel

ImageResolver

Markdown image resolver.

LinkResolver

Markdown link resolver.

Statics

LINK_EVENT

Event raised by markdown links when clicked.

Traits

WidgetInfoExt

Markdown extension methods for widget info.

Functions

anchor

P Set a label that identifies the widget in the context of the parent markdown.

block_quote_fn

P Widget function that converts BlockQuoteFnArgs to widgets.

code_block_fn

P Widget function that converts CodeBlockFnArgs to widgets.

code_inline_fn

P Widget function that converts CodeInlineFnArgs to widgets.

footnote_def_fn

P Widget function that converts FootnoteDefFnArgs to widgets.

footnote_ref_fn

P Widget function that converts FootnoteRefFnArgs to widgets.

heading_anchor

Generate an anchor label for a header.

heading_fn

P Widget function that converts HeadingFnArgs to widgets.

image_fn

P Widget function that converts ImageFnArgs to widgets.

image_resolver

P Markdown image resolver.

link_fn

P Widget function that converts LinkFnArgs to widgets.

link_resolver

P Markdown link resolver.

link_scroll_mode

P Scroll-to mode used by anchor links.

list_fn

P Widget function that converts ListFnArgs to widgets.

list_item_bullet_fn

P Widget function that converts ListItemBulletFnArgs to widgets.

list_item_fn

P Widget function that converts ListItemFnArgs to widgets.

on_link

P Markdown link click.

on_pre_link

P Preview on_link event.

panel_fn

P Widget function that converts PanelFnArgs to a widget.

paragraph_fn

P Widget function that converts ParagraphFnArgs to widgets.

rule_fn

P Widget function that converts RuleFnArgs to widgets.

table_fn

P Widget function that converts TableFnArgs to widgets.

text_fn

P Widget function that converts TextFnArgs to widgets.