zng::menu
Module context

Context menu widget and properties.

See zng_wgt_menu::context for the full widget API.

Structs

ContextMenu

W Context menu popup.

ContextMenuArgs

Arguments for context menu widget functions.

DefaultStyle

W Context menu popup default style.

TouchStyle

W Touch context menu popup default style.

Functions

context_menu

P Defines the context menu shown when the widget is enabled and receives a context click.

context_menu_anchor

P Set the position of the context-menu widgets opened for the widget or its descendants.

context_menu_fn

P Defines the context menu function shown when the widget is enabled and receives a context click.

disabled_context_menu

P Defines the context menu shown when the widget is disabled and receives a context click.

disabled_context_menu_fn

P Defines the context menu function shown when the widget is disabled and receives a context click.

panel_fn

P Widget function that generates the context menu layout.

style_fn

P Extends or replaces the widget style.