Module zng::menu::sub

Submenu widget and properties.

See zng_wgt_menu::sub for the full widget API.

Structs

DefaultStyle

W Style applied to SubMenu! not inside any other sub-menus.

SubMenu

W Submenu header and items.

SubMenuAncestors

Iterator over sub-menu parents.

SubMenuStyle

W Style applied to all SubMenu! widgets inside other sub-menus.

Traits

SubMenuWidgetInfoExt

Extension methods for WidgetInfo.

Functions

column_width_padding

P If the start and end column width is applied as padding.

end_column

P Sets the content to the Align::END side of the button menu item.

end_column_fn

P Sets the content to the Align::END side of the button menu item generated using a WidgetFn<()>.

end_column_width

P Width of the sub-menu expand symbol column.

hover_open_delay

P Delay a sub-menu must be hovered to open the popup.

is_open

P If the sub-menu popup is open or opening.

start_column

P Sets the content to the Align::START side of the button menu item.

start_column_fn

P Sets the content to the Align::START side of the button menu item generated using a WidgetFn<()>.

start_column_width

P Width of the icon/checkmark column.