zng::pointer_capture

Function capture_pointer_on_init

pub fn capture_pointer_on_init(
    child: impl UiNode,
    mode: impl IntoVar<CaptureMode>,
) -> impl UiNode

P Capture mouse and touch for the widget on init.

The capture happens only if any mouse button or touch is pressed on the window and the mode is Widget or Subtree.

Pointer captures are released when all mouse buttons stop being pressed on the window. The capture is also released back to window if the mode changes to Window when the mouse is captured for the widget.