Module zng::rule_line::vr

Vertical rule line widget and properties.

Structs

Vr

W Draws a vertical RuleLine!.

Functions

color

P Sets the COLOR_VAR that affects all vertical rules inside the widget.

line_style

P Sets the LINE_STYLE_VAR that affects all vertical rules inside the widget.

margin

P Sets the MARGIN_VAR that affects all vertical rules inside the widget.

stroke_thickness

P Sets the STROKE_THICKNESS_VAR that affects all vertical rules inside the widget.