zng::scroll

Function clip_to_viewport

pub fn clip_to_viewport(clip: impl IntoVar<bool>)

c Clip content to only be visible within the viewport, not under scrollbars.

Disabled by default.

Capture-Only

This property is capture-only, it only defines a property signature, it does not implement any behavior by itself. Widgets can capture and implement this property as part of their intrinsics, otherwise it will have no effect if set on a widget that does not implement it.