zng::task
Module io

IO tasks.

Most of the types in this module are re-exported from futures_lite::io.

Structs

BufReader

Adds buffering to a reader.

BufWriter

Adds buffering to a writer.

CloneableError

Represents the cloneable parts of an Error.

Cursor

Gives an in-memory buffer a cursor for reading and writing.

McBufReader

Multiple consumer buffered read.

Measure

Measure read/write of an async task.

Metrics

Information about the state of an async IO task.

ReadHalf

The read half returned by split().

ReadLimited

Represents a future that generates an error if an AsyncRead exceeds a limit.

WriteHalf

The write half returned by split().

Statics

METRICS_ID

Metrics in a Progress::with_meta metadata.

Traits

AsyncBufRead

Read bytes asynchronously.

AsyncBufReadExt

Extension trait for AsyncBufRead.

AsyncRead

Read bytes asynchronously.

AsyncReadExt

Extension trait for AsyncRead.

AsyncSeek

Seek bytes asynchronously.

AsyncSeekExt

Extension trait for AsyncSeek.

AsyncWrite

Write bytes asynchronously.

AsyncWriteExt

Extension trait for AsyncWrite.

McBufErrorExt

Extension methods for std::io::Error to be used with errors returned by McBufReader.

Functions

copy

Copies the entire contents of a reader into a writer.

empty

Creates an empty reader.

repeat

Creates an infinite reader that reads the same byte repeatedly.

sink

Creates a writer that consumes and drops all data.

split

Splits a stream into AsyncRead and AsyncWrite halves.

Type Aliases

BoxedReader

Type alias for Pin<Box<dyn AsyncRead + Send + 'static>>.

BoxedWriter

Type alias for Pin<Box<dyn AsyncWrite + Send + 'static>>.