zng::text

Module cmd

Commands that controls the editable/selectable text.

Most of the normal text editing is controlled by keyboard events, these commands.

Full API

See zng_wgt_text::cmd for the full API.

Structs

TextEditOp

Represents a text edit operation that can be send to an editable text using EDIT_CMD.

TextSelectOp

Represents a text caret/selection operation that can be send to an editable text using SELECT_CMD.

Statics

EDIT_CMD

Applies the TextEditOp into the text if it is editable.

PARSE_CMD

Parse text and update value if txt_parse is pending.

SELECT_ALL_CMD

Select all text.

SELECT_CMD

Applies the TextSelectOp into the text if it is editable.