Module zng::text::cmd

Commands that control the editable text.

Most of the normal text editing is controlled by keyboard events, the EDIT_CMD command allows for arbitrary text editing without needing to simulate keyboard events.

The node::resolve_text node implements EDIT_CMD when the text is editable.

Structs

TextEditOp

Represents a text edit operation that can be send to an editable text using EDIT_CMD.

TextSelectOp

Represents a text selection operation that can be send to an editable text using SELECT_CMD.

Statics

EDIT_CMD

Applies the TextEditOp into the text if it is editable.

PARSE_CMD

Parse text and update value if txt_parse is pending.

SELECT_ALL_CMD

Select all text.

SELECT_CMD

Applies the TextSelectOp into the text if it is editable.