zng::text

Struct ParagraphMix

pub struct ParagraphMix<P>(/* private fields */);

m Text paragraph properties.

Note that the Text! widget does not include this mix-in, as raw text does not encode paragraph breaks, other rich text widgets can include it to configure paragraphs.

Implementations

impl ParagraphMix<()>

pub fn context_vars_set(set: &mut ContextValueSet)

Insert context variables used by properties in this mix-in.

impl<P> ParagraphMix<P>
where P: WidgetImpl,

pub fn paragraph_spacing(&self, extra: impl IntoVar<Length>)

P Extra spacing in-between paragraphs.

The default value is 1.em(). Note that the Text! widget does not implement this property, as raw text does not encode paragraph breaks, this property and context var exists to configure rich-text widgets, like the Markdown! widget.

Sets the PARAGRAPH_SPACING_VAR.