zng::tip

Function tooltip_anchor

pub fn tooltip_anchor(
    child: impl UiNode,
    mode: impl IntoVar<AnchorMode>,
) -> impl UiNode

P Set the position of the tip widgets opened for the widget or its descendants.

Tips are inserted as POPUP when shown, this property defines how the tip layer is aligned with the anchor widget, or the cursor.

By default tips are aligned below the cursor position at the time they are opened.

This position is used when the tip opens with cursor interaction, see access_tooltip_anchor for position without the cursor.

This property sets the TOOLTIP_ANCHOR_VAR.