zng::tip

Function tooltip_context_capture

pub fn tooltip_context_capture(
    child: impl IntoUiNode,
    capture: impl IntoVar<ContextCapture>,
) -> UiNode

P Defines if the tooltip captures the build/instantiate context and sets it in the node context.

This is disabled by default, it can be enabled to have the tooltip be affected by context properties in the anchor widget.

Note that updates to this property do not affect tooltips already open, just subsequent tooltips.

This property sets the TOOLTIP_CONTEXT_CAPTURE_VAR.