zng::undo

Module history

Undo history widget.

Full API

See zng_wgt_undo_history for the full undo API.

Structs

UndoEntryArgs

Represents an action in the undo or redo stack.

UndoHistory

W Undo/redo stack view.

UndoPanelArgs

Args to present the child of UndoHistory!.

UndoRedoButtonStyle

W Menu style button for an entry in a undo/redo stack.

UndoStackArgs

Represents an undo or redo stack.

Functions

group_by_undo_interval

P If undo entries are grouped by the UNDO.undo_interval.

is_cap_hovered_timestamp

P State is true when the widget is an undo_entry and it is hovered, has captured the mouse or a sibling with higher timestamp is hovered/has cap.

undo_redo_button_style_fn

P Extends or replaces the UndoRedoButtonStyle!.