Trait zng::undo::UndoSelector

pub trait UndoSelector: Sealed {
    type Select: UndoSelect;

    // Required method
    fn select(self, op: UndoOp) -> Self::Select;
}

Represents a type that can select actions for undo or redo once.

This API is sealed, only core crate can implement it.

See UNDO::undo_select for more details.

Required Associated Types

type Select: UndoSelect

Selection collector.

Required Methods

fn select(self, op: UndoOp) -> Self::Select

Start selecting action for the op.