Trait zng::update::UpdateSubscribers

pub trait UpdateSubscribers: Send + Sync + 'static {
    // Required methods
    fn contains(&self, widget_id: WidgetId) -> bool;
    fn to_set(&self) -> IdSet<WidgetId>;
}

Represents a set of widgets that subscribe to an event source.

Required Methods

fn contains(&self, widget_id: WidgetId) -> bool

Returns true if the widget is one of the subscribers.

fn to_set(&self) -> IdSet<WidgetId>

Gets all subscribers as a set.