zng::var::animation
Module easing

Common easing functions.

Structs

Bezier

A unit cubic Bézier curve, used for timing functions in CSS transitions and animations.

EasingTime

Easing function input.

Enums

EasingFn

Easing functions as a value.

Functions

back

Cubic transition that goes slightly negative to start and ends very fast.

bounce

Oscillating transition that grows in magnitude, does not go negative, when the curve is about to go negative it sharply transitions to a new arc of larger magnitude.

circ

Cubic transition with slightly slowed start then cubic.

cubic

Cubic transition (t³).

cubic_bezier

X coordinate is time, Y coordinate is function advancement. The nominal range for both is 0 to 1.

ease_in

Applies the ease_fn.

ease_in_out

Applies ease_in for the first half then ease_out scaled to fit a single duration (1.0).

ease_out

Applies the ease_fn in reverse and flipped.

ease_out_in

Applies ease_out for the first half then ease_in scaled to fit a single duration (1.0).

elastic

Oscillating transition that grows in magnitude, goes negative twice.

expo

Exponential transition. Very slow start, very fast end.

linear

Simple linear transition, no easing, no acceleration.

none

Always 1.fct(), that is, the completed transition.

quad

Quadratic transition (t²).

quart

Fourth power transition (t⁴).

quint

Fifth power transition (t⁵).

reverse

Applies the ease_fn in reverse.

reverse_out

Applies the ease_fn flipped.

sine

Sine transition. Slow start, fast end.

step_ceil

Jumps to the final value by a number of steps.

step_floor

Jumps to the final value by a number of steps.

Type Aliases

EasingStep

Easing function output.