zng::var::animation

Trait AnimationController

pub trait AnimationController:
    Send
    + Sync
    + Any {
    // Provided methods
    fn on_start(&self, animation: &Animation) { ... }
    fn on_stop(&self, animation: &Animation) { ... }
}

Animations controller.

See VARS.with_animation_controller for more details.

Provided Methods

fn on_start(&self, animation: &Animation)

Called for each animation that starts in the controller context.

Note that this handler itself is not called inside the controller context.

fn on_stop(&self, animation: &Animation)

Called for each animation that ends in the controller context.

Note that this handler itself is not called inside the controller context.