Trait zng::var::AnyWeakVar

pub trait AnyWeakVar: Any + Send + Sync + Sealed {
    // Required methods
    fn clone_any(&self) -> Box<dyn AnyWeakVar>;
    fn as_any(&self) -> &(dyn Any + 'static);
    fn strong_count(&self) -> usize;
    fn weak_count(&self) -> usize;
    fn upgrade_any(&self) -> Option<Box<dyn AnyVar>>;
}

Represents a weak reference to an AnyVar.

Required Methods

fn clone_any(&self) -> Box<dyn AnyWeakVar>

Clone the weak reference.

fn as_any(&self) -> &(dyn Any + 'static)

Access to dyn Any methods.

fn strong_count(&self) -> usize

Gets the number of strong references to the variable.

This is the same as AnyVar::strong_count.

fn weak_count(&self) -> usize

Gets the number of weak references to the variable.

This is the same as AnyVar::weak_count.

fn upgrade_any(&self) -> Option<Box<dyn AnyVar>>

Upgrade to a strong AnyVar clone.

Returns None if the strong_count is zero.