zng::view_process

Module default

Default view-process implementation.

Full API

See zng_view for the full API including view API extensions such as enabling ANGLE backend on Windows.

Functions

run_same_process

Runs the view-process server in the current process and calls run_app to also run the app in the current process. Note that run_app will be called in a different thread.