Trait zng::widget::builder::PropertyArgs

pub trait PropertyArgs: Send + Sync {
    // Required methods
    fn clone_boxed(&self) -> Box<dyn PropertyArgs>;
    fn property(&self) -> PropertyInfo;
    fn instantiate(&self, child: Box<dyn UiNodeBoxed>) -> Box<dyn UiNodeBoxed>;

    // Provided methods
    fn var(&self, i: usize) -> &(dyn AnyVar + 'static) { ... }
    fn value(&self, i: usize) -> &(dyn AnyVarValue + 'static) { ... }
    fn ui_node(&self, i: usize) -> &ArcNode<Box<dyn UiNodeBoxed>> { ... }
    fn ui_node_list(&self, i: usize) -> &ArcNodeList<Box<dyn UiNodeListBoxed>> { ... }
    fn widget_handler(&self, i: usize) -> &(dyn AnyArcWidgetHandler + 'static) { ... }
}

Represents a property instantiation request.

Required Methods

fn clone_boxed(&self) -> Box<dyn PropertyArgs>

Clones the arguments.

fn property(&self) -> PropertyInfo

Property info.

fn instantiate(&self, child: Box<dyn UiNodeBoxed>) -> Box<dyn UiNodeBoxed>

Create a property instance with args clone or taken.

If the property is PropertyInfo::capture the child is returned.

Provided Methods

fn var(&self, i: usize) -> &(dyn AnyVar + 'static)

Gets a InputKind::Var.

Is a BoxedVar<T>.

fn value(&self, i: usize) -> &(dyn AnyVarValue + 'static)

Gets a InputKind::Value.

fn ui_node(&self, i: usize) -> &ArcNode<Box<dyn UiNodeBoxed>>

Gets a InputKind::UiNode.

fn ui_node_list(&self, i: usize) -> &ArcNodeList<Box<dyn UiNodeListBoxed>>

Gets a InputKind::UiNodeList.

fn widget_handler(&self, i: usize) -> &(dyn AnyArcWidgetHandler + 'static)

Gets a InputKind::WidgetHandler.

Is a ArcWidgetHandler<A>.

Implementations

impl dyn PropertyArgs + '_

pub fn id(&self) -> PropertyId

Unique ID.

pub fn downcast_value<T>(&self, i: usize) -> &T
where T: VarValue,

Gets a strongly typed value.

Panics if the type does not match.

pub fn downcast_var<T>(&self, i: usize) -> &Box<dyn VarBoxed<T>>
where T: VarValue,

Gets a strongly typed var.

Panics if the variable value type does not match.

pub fn downcast_handler<A>(&self, i: usize) -> &ArcWidgetHandler<A>
where A: 'static + Clone,

Gets a strongly typed widget_handler.

Panics if the args type does not match.

pub fn live_debug(&self, i: usize) -> Box<dyn VarBoxed<Txt>>

Gets the property input as a debug variable.

If the input is a variable the returned variable will update with it, if not it is a static print.

Note that you must call this in the widget context to get the correct value.

pub fn debug(&self, i: usize) -> Txt

Gets the property input current value as a debug text.

Note that you must call this in the widget context to get the correct value.

pub fn new_build( &self, build_actions: Vec<Vec<Box<dyn AnyPropertyBuildAction>>>, build_actions_when_data: Vec<Vec<Option<Arc<dyn Any + Sync + Send>>>>, ) -> Box<dyn PropertyArgs>

Call new with the same instance info and args, but with the build_actions and build_actions_when_data.