Struct zng::widget::info::access::WidgetAccessInfo

pub struct WidgetAccessInfo { /* private fields */ }

Accessibility info for a widget.

Implementations

impl WidgetAccessInfo

pub fn info(&self) -> &WidgetInfo

Full widget info.

pub fn role(&self) -> Option<AccessRole>

Accessibility role of the widget.

pub fn commands(&self) -> &[AccessCmdName]

Accessibility commands supported by the widget.

pub fn auto_complete(&self) -> Option<AutoComplete>

How input text triggers display of one or more predictions of the user’s intended value.

pub fn checked(&self) -> Option<Option<bool>>

If the widget is checked (Some(true)), unchecked (Some(false)), or if the checked status is indeterminate (None).

Note that the value is wrapped in another Option<_> that indicates if it was set or not.

pub fn current(&self) -> Option<CurrentKind>

Kind of current item the widget represents.

pub fn error_message(&self) -> Option<WidgetInfo>

Gets the invalid widget that this widget is an error message for.

pub fn active_descendant(&self) -> Option<WidgetInfo>

Identifies the currently active widget when focus is on a composite widget.

pub fn expanded(&self) -> Option<bool>

Gets visibility of related widgets.

pub fn has_popup(&self) -> Option<Popup>

Indicates the availability and type of interactive popup widget.

pub fn invalid(&self) -> Invalid

If the widget data has errors.

pub fn label(&self) -> Option<Txt>

Gets the accessibility name explicitly set on this widget.

pub fn labelled_by_child(&self) -> bool

If the widget children must be used like labelled_by.

pub fn lang(&self) -> Option<LanguageIdentifier>

Gets the language of texts inside this widget and descendants.

If not set it is the parents language.

pub fn scroll_horizontal(&self) -> Option<Box<dyn VarBoxed<Factor>>>

Normalized (0..1) horizontal scroll, 0 is showing the content leftmost edge, 1 is showing the content the rightmost edge.

Also signals that the content is horizontally scrollable.

pub fn scroll_vertical(&self) -> Option<Box<dyn VarBoxed<Factor>>>

Normalized (0..1) vertical scroll, 0 is showing the content topmost edge, 1 is showing the content the bottommost edge.

Also signals that the content is vertically scrollable.

pub fn is_multi_selectable(&self) -> bool

Indicates that the user may select more than one item from the current selectable descendants.

pub fn orientation(&self) -> Option<Orientation>

Indicates whether the widget’s orientation is horizontal, vertical, or unknown/ambiguous.

pub fn placeholder(&self) -> Option<Txt>

Short hint (a word or short phrase) intended to help the user with data entry when a form control has no value.

pub fn is_read_only(&self) -> bool

Indicates that the widget is not editable, but is otherwise operable.

pub fn is_required(&self) -> bool

Indicates that user input is required on the widget before a form may be submitted.

pub fn level(&self) -> Option<NonZero<u32>>

Defines the hierarchical level of a widget within a structure.

pub fn is_selected(&self) -> bool

Indicates that the widget is selected.

pub fn sort(&self) -> Option<SortDirection>

Indicates if items in a table or grid are sorted in ascending or descending order.

pub fn value_max(&self) -> Option<f64>

Maximum value (inclusive).

pub fn value_min(&self) -> Option<f64>

Minimum value (inclusive).

pub fn value(&self) -> Option<f64>

Current value.

pub fn value_text(&self) -> Option<Txt>

Current value in a readable format.

Note that this returns Some(_) only when a value text was set, value may or may not be set also.

pub fn live(&self) -> Option<(LiveIndicator, bool, bool)>

Gets the live indicator, atomic and busy.

See AccessState::Live for more details.

pub fn col_count(&self) -> Option<usize>

Defines the total number of columns in a Table, Grid, or TreeGrid when not all columns are present in tree.

The value 0 indicates that not all columns are in the widget and the application cannot determinate the exact number.

pub fn col_index(&self) -> Option<usize>

Defines a widget’s column index in the parent table or grid.

pub fn col_span(&self) -> Option<usize>

Defines the number of columns spanned by the widget in the parent table or grid.

pub fn row_count(&self) -> Option<usize>

Defines the total number of rows in a Table, Grid, or TreeGrid when not all rows are present in tree.

The value 0 indicates that not all rows are in the widget and the application cannot determinate the exact number.

pub fn row_index(&self) -> Option<usize>

Defines a widget’s column index in the parent table or grid.

pub fn row_span(&self) -> Option<usize>

Defines the number of columns spanned by the widget in the parent table or grid.

pub fn item_count(&self) -> Option<usize>

Defines the number of items in the current set of list items or tree items when not all items in the set are present in the tree.

pub fn item_index(&self) -> Option<usize>

Defines the widget’s number or position in the current set of list items or tree items when not all items are present in the tree.

pub fn modal(&self) -> bool

Indicates whether the widget is modal when displayed.

pub fn controls(&self) -> impl Iterator<Item = WidgetInfo>

Widget(s) whose contents or presence are controlled by this widget.

pub fn described_by(&self) -> impl Iterator<Item = WidgetInfo>

Identifies the widget(s) that describes this widget.

pub fn describes(&self) -> Option<WidgetInfo>

Identifies the widget that is described by this widget.

Note that this is not a query for all widgets that have this one in their described_by list, it is only set if it was set explicitly during info build.

pub fn details(&self) -> impl Iterator<Item = WidgetInfo>

Identifies the widget(s) that provide additional information related to this widget.

pub fn labelled_by(&self) -> impl Iterator<Item = WidgetInfo>

Identifies the widget(s) that labels the widget it is applied to.

pub fn labels(&self) -> Option<WidgetInfo>

Identifies the widget that is labelled by this widget.

Note that this is not a query for all widgets that have this one in their labelled_by list, it is only set if it was set explicitly during info build.

pub fn owns(&self) -> impl Iterator<Item = WidgetInfo>

Extra widgets that are child to this widget, but are not descendants on the info tree.

pub fn flows_to(&self) -> impl Iterator<Item = WidgetInfo>

Options for next widget to read.

pub fn is_accessible(&self) -> bool

If the widget and descendants is visible in the view-process and screen readers.

Note that the accessibility info for the widget and descendants is still available in the app-process.