Module zng::widget::node

Widget node types, UiNode, UiNodeList and others.

Macros

ui_vec

Creates an UiNodeVec containing the arguments.

Structs

AdoptiveChildNode

Placeholder for the dynamic child of an adoptive node.

AdoptiveNode

Represents a node setup to dynamically swap child.

ArcNode

A reference counted UiNode.

ArcNodeList

A reference counted UiNodeList.

DefaultPanelListData

Default PanelList associated data.

EditableUiNodeList

Represents an UiNodeList that can be modified using a connected sender.

EditableUiNodeListRef

Represents a sender to an EditableUiNodeList.

FillUiNode

A UI node that fills the available layout space.

MatchNodeChild

Child node of match_node.

MatchNodeChildren

Children node of match_node_list.

MatchWidgetChild

Child node of match_widget.

NilUiNode

A UI node that does nothing and has collapsed layout (zero size).

OffsetUiListObserver

Represents an UiNodeListObserver that applies an offset to all indexes.

PanelList

Represents the final UiNodeList in a panel layout node.

PanelListRange

First and last child widget in a PanelList.

SORTING_LIST

Represents the contextual parent SortingList during an update.

SortingList

Represents a sorted view into an UiNodeList.

UiNodeListChainImpl

Implementation of UiNodeListChain::chain.

UiNodeVec

Vec of boxed UI nodes.

WeakNode

Weak reference to a ArcNode<U>.

WeakNodeList

Weak reference to a ArcNodeList<U>.

WhenUiNodeBuilder

Builds a node that can be one of multiple options, selected by the first condition that is true, or a fallback default.

WhenUiNodeListBuilder

Builds a node list that can be one of multiple options, selected by the first condition that is true, or a fallback default.

Z_INDEX

Access to widget z-index in a parent PanelList.

Enums

UiNodeOp

Represents a node operation in a match_node.

UiNodeOpMethod

Identifies an UiNodeOp method without the associated data.

Traits

PanelListData

Represents an item’s associated data in a PanelList.

UiNode

Represents an UI tree node.

UiNodeList

Represents a list of UI nodes.

UiNodeListChain

Adds the chain method for all UiNodeList implementors.

UiNodeListObserver

Represents an UiNodeList::update_all observer that can be used to monitor widget insertion, removal and re-order.

Functions

bind_state

Helper for declaring state properties that are controlled by a variable.

border_node

Creates a border node that delegates rendering to a border_visual and manages the border_offsets coordinating with the other borders of the widget.

event_state

Helper for declaring state properties that depend on a single event.

event_state2

Helper for declaring state properties that depend on two other event states.

event_state3

Helper for declaring state properties that depend on three other event states.

event_state4

Helper for declaring state properties that depend on four other event states.

extend_widget

Create a widget node that wraps the widget with any number of other non-widget nodes and still delegates with_context to the widget.

fill_node

Transforms and clips the content node according with the default widget border align behavior.

interactive_node

Create a node that controls interaction for all widgets inside node.

list_presenter

Node that presents list using item_fn for each new list item.

match_node

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to another child node.

match_node_leaf

Creates a node that is implemented as a closure that matches over UiNodeOp and does not delegate to any child node.

match_node_list

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to multiple children nodes in a list.

match_node_typed

Like match_node, but does not change the child type when build with dyn_node.

match_widget

Creates a widget that is implemented as a closure that matches over UiNodeOp and delegates to another child widget.

presenter

Node that presents data using wgt_fn.

presenter_opt

Node that presents data using wgt_fn if data is available, otherwise presents nil.

widget_state_get_state

Helper for declaring state getter properties that are controlled by values in the widget state map.

widget_state_is_state

Helper for declaring state properties that are controlled by values in the widget state map.

with_context_blend

Helper for declaring widgets that are recontextualized to take in some of the context of an original parent.

with_context_local

Helper for declaring nodes that sets a context local value.

with_context_local_init

Helper for declaring nodes that sets a context local value generated on init.

with_context_var

Helper for declaring properties that sets a context var.

with_context_var_init

Helper for declaring properties that sets a context var to a value generated on init.

with_index_len_node

Helper for a property that gets the index of the widget in the parent panel and the number of children.

with_index_node

Helper for a property that gets the index of the widget in the parent panel.

with_rev_index_node

Helper for a property that gets the reverse index of the widget in the parent panel.

with_widget_state

Helper for declaring properties that set the widget state.

with_widget_state_modify

Helper for declaring properties that set the widget state with a custom closure.

Type Aliases

BoxedUiNode

An UiNode in a box.

BoxedUiNodeList

An UiNodeList in a box.