zng::widget

Module node

Widget node types, UiNode, UiVec and others.

Macros

ui_vec

Creates an UiVec containing the arguments.

Structs

AdoptiveChildNode

Placeholder for the dynamic child of an adoptive node.

AdoptiveNode

Represents a node setup to dynamically swap child.

ArcNode

A reference counted UiNode.

ChainList

UI node list implementation that flattens child lists.

DefaultPanelListData

Default PanelList associated data.

EditableUiVec

Represents an UiVec that can be modified using a connected sender.

EditableUiVecRef

Represents a sender to an EditableUiVec.

FillUiNode

A UI node that fills the available layout space.

MatchNodeChild

Child node of match_node.

MatchWidgetChild

Child node of match_widget.

OffsetUiListObserver

Represents an UiNodeListObserver that applies an offset to all indexes.

PanelList

Represents the final UI list in a panel layout node.

PanelListRange

First and last child widget in a PanelList.

SORTING_LIST

Represents the contextual parent SortingList during an update.

SortingList

Represents a sorted view into an UiNode::is_list.

UiNode

Represents an UI tree node instance.

UiVec

Vec of boxed UI nodes.

WeakNode

Weak reference to a ArcNode.

WhenUiNodeBuilder

Builds a node that can be one of multiple options, selected by the first condition that is true, or a fallback default.

WidgetUiNode

Extra UiNode methods for nodes that define a widget instance context.

Z_INDEX

Access to widget z-index in a parent PanelList.

Enums

UiNodeMethod

Identifies the UiNode UI methods.

UiNodeOp

Represents a node operation in a match_node.

Traits

IntoUiNode

Represents a value that can become a UiNode instance.

PanelListData

Represents an item’s associated data in a PanelList.

UiNodeImpl

Represents an UiNode implementation.

UiNodeListObserver

Represents an UiNode::update_list observer that can be used to monitor widget insertion, removal and re-order.

WidgetUiNodeImpl

Represents an UiNodeImpl that defines a widget instance scope.

Functions

bind_state

Helper for declaring state properties that are controlled by a variable.

bind_state_init

Helper for declaring state properties that are controlled by a variable.

border_node

Creates a border node that delegates rendering to a border_visual and manages the border_offsets coordinating with the other borders of the widget.

event_state

Helper for declaring state properties that depend on a single event.

event_state2

Helper for declaring state properties that depend on two other event states.

event_state3

Helper for declaring state properties that depend on three other event states.

event_state4

Helper for declaring state properties that depend on four other event states.

extend_widget

Create a widget node that wraps the widget with any number of other non-widget nodes and still delegates with_context to the widget.

fill_node

Transforms and clips the content node according with the default widget border align behavior.

interactive_node

Create a node that controls interaction for all widgets inside node.

list_presenter

Node list that presents list using item_fn for each new list item.

list_presenter_from_iter

Node list that presents list using item_fn for each list item.

match_node

Creates a node that is implemented as a closure that matches over UiNodeOp and delegates to another child node.

match_node_leaf

Creates a node that is implemented as a closure that matches over UiNodeOp and does not delegate to any child node.

match_widget

Creates a widget that is implemented as a closure that matches over UiNodeOp and delegates to another child widget.

presenter

Node that presents data using wgt_fn.

presenter_opt

Node that presents data using wgt_fn if data is available, otherwise presents nil.

widget_state_get_state

Helper for declaring state getter properties that are controlled by values in the widget state map.

widget_state_is_state

Helper for declaring state properties that are controlled by values in the widget state map.

with_context_blend

Helper for declaring widgets that are recontextualized to take in some of the context of an original parent.

with_context_local

Helper for declaring nodes that sets a context local value.

with_context_local_init

Helper for declaring nodes that sets a context local value generated on init.

with_context_var

Helper for declaring properties that sets a context var.

with_context_var_init

Helper for declaring properties that sets a context var to a value generated on init.

with_index_len_node

Helper for a property that gets the index of the widget in the parent panel and the number of children.

with_index_node

Helper for a property that gets the index of the widget in the parent panel.

with_rev_index_node

Helper for a property that gets the reverse index of the widget in the parent panel.

with_widget_state

Helper for declaring properties that set the widget state.

with_widget_state_modify

Helper for declaring properties that set the widget state with a custom closure.