zng::window

Module cmd

Window commands.

Enums

ResizeDirection

Defines the orientation that a window resize will be performed.

Statics

CANCEL_IME_CMD

Represents the close IME action.

CLOSE_CMD

Represents the window close action.

DRAG_MOVE_RESIZE_CMD

Represents the window drag-move and drag-resize actions.

EXCLUSIVE_FULLSCREEN_CMD

Represents the window toggle fullscreen action.

FULLSCREEN_CMD

Represents the window toggle fullscreen action.

INSPECT_CMD

Represent the window inspect action.

MAXIMIZE_CMD

Represents the window maximize action.

MINIMIZE_CMD

Represents the window minimize action.

OPEN_TITLE_BAR_CONTEXT_MENU_CMD

Represents the window open title bar context menu action.

RESTORE_CMD

Represents the window restore action.