Struct zng::window::Window

pub struct Window(/* private fields */);

W A window container.

The instance type is WindowRoot, it can be given to the WINDOWS service to open a system window that is kept in sync with the window properties set in the widget.

See run_window for more details.

Implementations

impl Window

pub fn position(&self, position: impl IntoVar<Point>)

P Binds the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn monitor(&self, monitor: impl IntoVar<MonitorQuery>)

P Binds the monitor window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn state(&self, state: impl IntoVar<WindowState>)

P Binds the state window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn size(&self, size: impl IntoVar<Size>)

P Binds the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn min_size(&self, min_size: impl IntoVar<Size>)

P Binds the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn max_size(&self, max_size: impl IntoVar<Size>)

P Binds the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn font_size(&self, font_size: impl IntoVar<Length>)

P Binds the font_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn chrome(&self, chrome: impl IntoVar<bool>)

P Binds the chrome window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn icon(&self, icon: impl IntoVar<WindowIcon>)

P Binds the icon window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn title(&self, title: impl IntoVar<Txt>)

P Binds the title window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn auto_size(&self, auto_size: impl IntoVar<AutoSize>)

P Binds the auto_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn auto_size_origin(&self, auto_size_origin: impl IntoVar<Point>)

P Binds the auto_size_origin window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn resizable(&self, resizable: impl IntoVar<bool>)

P Binds the resizable window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn movable(&self, movable: impl IntoVar<bool>)

P Binds the movable window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn always_on_top(&self, always_on_top: impl IntoVar<bool>)

P Binds the always_on_top window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn visible(&self, visible: impl IntoVar<bool>)

P Binds the visible window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn taskbar_visible(&self, taskbar_visible: impl IntoVar<bool>)

P Binds the taskbar_visible window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn parent(&self, parent: impl IntoVar<Option<WindowId>>)

P Binds the parent window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn modal(&self, modal: impl IntoVar<bool>)

P Binds the modal window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn color_scheme(&self, color_scheme: impl IntoVar<Option<ColorScheme>>)

P Binds the color_scheme window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn frame_capture_mode( &self, frame_capture_mode: impl IntoVar<FrameCaptureMode> )

P Binds the frame_capture_mode window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn enabled_buttons(&self, enabled_buttons: impl IntoVar<WindowButton>)

P Binds the enabled_buttons window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn x(&self, x: impl IntoVar<Length>)

P Binds the x of the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn y(&self, y: impl IntoVar<Length>)

P Binds the y of the position window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn width(&self, width: impl IntoVar<Length>)

P Binds the width of the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn height(&self, height: impl IntoVar<Length>)

P Binds the height of the size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn min_width(&self, min_width: impl IntoVar<Length>)

P Binds the width of the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn min_height(&self, min_height: impl IntoVar<Length>)

P Binds the height of the min_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn max_width(&self, max_width: impl IntoVar<Length>)

P Binds the width of the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn max_height(&self, max_height: impl IntoVar<Length>)

P Binds the height of the max_size window var with the property value.

The binding is bidirectional and the window variable is assigned on init.

impl Window

pub fn clear_color(&self, color: impl IntoVar<Rgba>)

P Window clear color.

Color used to clear the previous frame pixels before rendering a new frame. It is visible if window content does not completely fill the content area, this can happen if you do not set a background or the background is semi-transparent, also can happen during very fast resizes.

impl Window

pub fn save_state(&self, enabled: impl IntoValue<SaveState>)

P Save and restore the window state.

If enabled a config entry is created for the window state in CONFIG, and if a config backend is set the window state is persisted on change and restored when the app reopens.

It is recommended to open the window with an unique name if the app can open more than one window, otherwise the state will be associated with the sequential ID of the window.

This property is enabled by default in the Window! widget.

impl Window

pub fn widget_new() -> Window

Start building a new instance.

pub fn widget_type() -> WidgetType

Gets the widget type info.

impl Window

pub fn widget_build(&mut self) -> WindowRoot

Build a WindowRoot.

impl Window

pub fn start_position(&self, position: impl IntoValue<StartPosition>)

c Defines how the window is positioned when it first opens.

impl Window

pub fn start_focused(&self, enabled: impl IntoValue<bool>)

c If the window is steals keyboard focus on open.

By default the operating system decides if the window will receive focus after opening, usually it is focused only if the process that started the window already has focus. Enabling this ensures that focus is moved to the new window, potentially stealing the focus from other apps and disrupting the user.

impl Window

pub fn kiosk(&self, kiosk: impl IntoValue<bool>)

c Lock-in kiosk mode.

In kiosk mode the only window states allowed are fullscreen or fullscreen exclusive, and all subsequent windows opened are child of the kiosk window.

Note that this does not configure the operating system, you still need to setup a kiosk environment. This just stops the app itself from accidentally exiting fullscreen.

impl Window

pub fn allow_transparency(&self, allow: impl IntoValue<bool>)

c If semi-transparent content is see-through, mixing with the operating system pixels behind the window.

Note that to actually see behind the window you must set the clear_color and background_color to a transparent color. The composition is a simple alpha blend, effects like blur do not apply to the pixels behind the window.

impl Window

pub fn render_mode(&self, mode: impl IntoValue<Option<RenderMode>>)

c Render performance mode overwrite for this window, if set to None the WINDOWS.default_render_mode is used.

The view-process will try to match the mode, if it is not available a fallback mode is selected, see RenderMode for more details about each mode and fallbacks.

impl Window

pub fn on_open(&self, handler: impl WidgetHandler<WindowOpenArgs>)

P Event just after the window opens.

This event notifies once per window, after the window content is inited.

This property is the same as on_pre_window_open.

impl Window

pub fn on_load(&self, handler: impl WidgetHandler<WindowOpenArgs>)

P Event just after the window loads.

This event notifies once per window, after the first layout and all WindowLoadingHandle have expired or dropped.

This property is the same as on_pre_window_load.

impl Window

pub fn on_close_requested( &self, handler: impl WidgetHandler<WindowCloseRequestedArgs> )

P On window close requested.

This event notifies every time an attempt to close the window is made. Close can be cancelled by stopping propagation on the event args, the window only closes after all handlers receive this event and propagation is not stopped.

This property is the same as on_window_close_requested.

impl Window

pub fn on_close(&self, handler: impl WidgetHandler<WindowCloseArgs>)

P On window close.

The window will deinit after this event.

This property is the same as on_pre_window_close.

impl Window

pub fn on_moved(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window position changed.

This event notifies every time the window position changes. You can also track the window position using the actual_position variable.

This property is the same as on_pre_window_moved.

impl Window

pub fn on_resized(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window size changed.

This event notifies every time the window content area size changes. You can also track the window size using the actual_size variable.

This property is the same as on_pre_window_resized.

impl Window

pub fn on_state_changed(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window state changed.

This event notifies every time the window state changes.

Note that you can also track the window state by setting state to a read-write variable.

This property is the same as on_pre_window_state_changed.

impl Window

pub fn on_maximized(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window maximized.

This event notifies every time the window state changes to maximized.

This property is the same as on_pre_window_maximized.

impl Window

pub fn on_unmaximized(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window exited the maximized state.

This event notifies every time the window state changes to a different state from maximized.

This property is the same as on_pre_window_unmaximized.

impl Window

pub fn on_minimized(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window minimized.

This event notifies every time the window state changes to minimized.

This property is the same as on_pre_window_maximized.

impl Window

pub fn on_unminimized(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window exited the minimized state.

This event notifies every time the window state changes to a different state from minimized.

This property is the same as on_pre_window_unminimized.

impl Window

pub fn on_restored(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window state changed to Normal.

This event notifies every time the window state changes to Normal.

This property is the same as on_pre_window_restored.

impl Window

pub fn on_fullscreen(&self, handler: impl WidgetHandler<WindowChangedArgs>)

P On window enter one of the fullscreen states.

This event notifies every time the window state changes to Fullscreen or Exclusive.

This property is the same as on_pre_window_fullscreen.

impl Window

pub fn on_exited_fullscreen( &self, handler: impl WidgetHandler<WindowChangedArgs> )

P On window is no longer fullscreen.

This event notifies every time the window state changes to one that is not fullscreen.

This property is the same as on_pre_window_exited_fullscreen.

impl Window

pub fn on_frame_image_ready( &self, handler: impl WidgetHandler<FrameImageReadyArgs> )

P On window frame rendered.

If frame_capture_mode is set the image will be available in the event args.

impl Window

pub fn headless_monitor(&self, monitor: impl IntoValue<HeadlessMonitor>)

c Imaginary monitor used by the window when it runs in headless mode.